refactor!: various changes related to icons Removed `<Icon />`, `iconsMap`& `AllIcons` Changed script to generate only the enum Move lint of generated file closer
