+++ .sonarcloud.properties
@@ -1,3 +1,1 @@
-sonar.cpd.exclusions=src/foundation/Icon/IconsMap.tsx
-sonar.exclusions=src/foundation/Icon/IconsMap.tsx,src/components/Button/ButtonClasses.tsx
-sonar.test.exclusions=src/foundation/Icon/IconsMap.tsx
+sonar.exclusions=src/components/Button/ButtonClasses.tsx
