+++ package.json
@@ -35,9 +35,9 @@
         "dev:library": "vite build --emptyOutDir false --watch",
         "clean": "rimraf dist",
         "component:create": "esno scripts/createNewComponent.ts",
-        "generate:icons": "pnpm generate:reactIcons && pnpm generate:iconsMap",
-        "generate:iconsMap": "esno ./scripts/generateIconMap.ts && eslint --fix ./src/foundation/Icon/IconEnum.ts ./src/foundation/Icon/IconsMap.tsx ./src/foundation/Icon/Generated/index.ts",
-        "generate:reactIcons": "esno ./scripts/generateReactIcons.ts",
+        "generate:icons": "pnpm generate:reactIcons && pnpm generate:iconsEnum",
+        "generate:iconsEnum": "esno ./scripts/generateIconEnum.ts && eslint --fix ./src/foundation/Icon/IconEnum.ts",
+        "generate:reactIcons": "esno ./scripts/generateReactIcons.ts && eslint --fix ./src/foundation/Icon/Generated/index.ts",
         "lint": "eslint .",
         "lint:fix": "eslint --fix .",
         "lint:prettier": "prettier --check .",
