+++ src/foundation/Icon/Icon.stories.tsx
@@ -5,8 +5,8 @@ import React from 'react';
 import { Meta, StoryFn } from '@storybook/react';
 import { IconSize } from './IconSize';
 import { IconProps } from './IconProps';
-import { Icon as IconComponent } from './Icon';
 import { IconEnum } from './IconEnum';
+import * as AllIcons from './Generated';
 
 export default {
     title: 'Foundation/Icon',
@@ -29,11 +29,17 @@ export const Icon: StoryFn<IconProps> = (args: IconProps) => (
                     iconName.includes((args?.size || '16').replace('Size', '')) ||
                     !['12', '16', '20', '24', '32'].some((item) => iconName.includes(item)),
             )
-            .map((iconName) => (
-                <li key={iconName} className="tw-flex tw-flex-col tw-items-center tw-text-violet-80 dark:tw-text-white">
-                    <IconComponent {...args} icon={iconName} />
-                    {iconName}
-                </li>
-            ))}
+            .map((iconName) => {
+                const Icon = AllIcons[`Icon${iconName}`];
+                return (
+                    <li
+                        key={iconName}
+                        className="tw-flex tw-flex-col tw-items-center tw-text-violet-80 dark:tw-text-white"
+                    >
+                        <Icon {...args} />
+                        {iconName}
+                    </li>
+                );
+            })}
     </ul>
 );
