--- src/foundation/Icon/IconsMap.tsx
@@ -1,3643 +1,0 @@
-/* (c) Copyright Frontify Ltd., all rights reserved. */
-
-import React, { ReactElement } from 'react';
-import { IconEnum } from './IconEnum';
-import IconAdobeCreativeCloud from '@foundation/Icon/Generated/IconAdobeCreativeCloud';
-import IconAdobeCreativeCloud12 from '@foundation/Icon/Generated/IconAdobeCreativeCloud12';
-import IconAdobeCreativeCloud16 from '@foundation/Icon/Generated/IconAdobeCreativeCloud16';
-import IconAdobeCreativeCloud20 from '@foundation/Icon/Generated/IconAdobeCreativeCloud20';
-import IconAdobeCreativeCloud24 from '@foundation/Icon/Generated/IconAdobeCreativeCloud24';
-import IconAdobeCreativeCloud32 from '@foundation/Icon/Generated/IconAdobeCreativeCloud32';
-import IconAdobeInDesign from '@foundation/Icon/Generated/IconAdobeInDesign';
-import IconAdobeInDesign12 from '@foundation/Icon/Generated/IconAdobeInDesign12';
-import IconAdobeInDesign16 from '@foundation/Icon/Generated/IconAdobeInDesign16';
-import IconAdobeInDesign20 from '@foundation/Icon/Generated/IconAdobeInDesign20';
-import IconAdobeInDesign24 from '@foundation/Icon/Generated/IconAdobeInDesign24';
-import IconAdobeInDesign32 from '@foundation/Icon/Generated/IconAdobeInDesign32';
-import IconAnchor from '@foundation/Icon/Generated/IconAnchor';
-import IconAnchor12 from '@foundation/Icon/Generated/IconAnchor12';
-import IconAnchor16 from '@foundation/Icon/Generated/IconAnchor16';
-import IconAnchor20 from '@foundation/Icon/Generated/IconAnchor20';
-import IconAnchor24 from '@foundation/Icon/Generated/IconAnchor24';
-import IconAnchor32 from '@foundation/Icon/Generated/IconAnchor32';
-import IconAnnotationLines from '@foundation/Icon/Generated/IconAnnotationLines';
-import IconAnnotationLines12 from '@foundation/Icon/Generated/IconAnnotationLines12';
-import IconAnnotationLines16 from '@foundation/Icon/Generated/IconAnnotationLines16';
-import IconAnnotationLines20 from '@foundation/Icon/Generated/IconAnnotationLines20';
-import IconAnnotationLines24 from '@foundation/Icon/Generated/IconAnnotationLines24';
-import IconAnnotationLines32 from '@foundation/Icon/Generated/IconAnnotationLines32';
-import IconAnnotationNumbers from '@foundation/Icon/Generated/IconAnnotationNumbers';
-import IconAnnotationNumbers12 from '@foundation/Icon/Generated/IconAnnotationNumbers12';
-import IconAnnotationNumbers16 from '@foundation/Icon/Generated/IconAnnotationNumbers16';
-import IconAnnotationNumbers20 from '@foundation/Icon/Generated/IconAnnotationNumbers20';
-import IconAnnotationNumbers24 from '@foundation/Icon/Generated/IconAnnotationNumbers24';
-import IconAnnotationNumbers32 from '@foundation/Icon/Generated/IconAnnotationNumbers32';
-import IconArchive from '@foundation/Icon/Generated/IconArchive';
-import IconArchive12 from '@foundation/Icon/Generated/IconArchive12';
-import IconArchive16 from '@foundation/Icon/Generated/IconArchive16';
-import IconArchive20 from '@foundation/Icon/Generated/IconArchive20';
-import IconArchive24 from '@foundation/Icon/Generated/IconArchive24';
-import IconArchive32 from '@foundation/Icon/Generated/IconArchive32';
-import IconArrowAlignDown from '@foundation/Icon/Generated/IconArrowAlignDown';
-import IconArrowAlignDown12 from '@foundation/Icon/Generated/IconArrowAlignDown12';
-import IconArrowAlignDown16 from '@foundation/Icon/Generated/IconArrowAlignDown16';
-import IconArrowAlignDown20 from '@foundation/Icon/Generated/IconArrowAlignDown20';
-import IconArrowAlignDown24 from '@foundation/Icon/Generated/IconArrowAlignDown24';
-import IconArrowAlignDown32 from '@foundation/Icon/Generated/IconArrowAlignDown32';
-import IconArrowAlignHorizontalCentre from '@foundation/Icon/Generated/IconArrowAlignHorizontalCentre';
-import IconArrowAlignHorizontalCentre12 from '@foundation/Icon/Generated/IconArrowAlignHorizontalCentre12';
-import IconArrowAlignHorizontalCentre16 from '@foundation/Icon/Generated/IconArrowAlignHorizontalCentre16';
-import IconArrowAlignHorizontalCentre20 from '@foundation/Icon/Generated/IconArrowAlignHorizontalCentre20';
-import IconArrowAlignHorizontalCentre24 from '@foundation/Icon/Generated/IconArrowAlignHorizontalCentre24';
-import IconArrowAlignHorizontalCentre32 from '@foundation/Icon/Generated/IconArrowAlignHorizontalCentre32';
-import IconArrowAlignLeft from '@foundation/Icon/Generated/IconArrowAlignLeft';
-import IconArrowAlignLeft12 from '@foundation/Icon/Generated/IconArrowAlignLeft12';
-import IconArrowAlignLeft16 from '@foundation/Icon/Generated/IconArrowAlignLeft16';
-import IconArrowAlignLeft20 from '@foundation/Icon/Generated/IconArrowAlignLeft20';
-import IconArrowAlignLeft24 from '@foundation/Icon/Generated/IconArrowAlignLeft24';
-import IconArrowAlignLeft32 from '@foundation/Icon/Generated/IconArrowAlignLeft32';
-import IconArrowAlignRight from '@foundation/Icon/Generated/IconArrowAlignRight';
-import IconArrowAlignRight12 from '@foundation/Icon/Generated/IconArrowAlignRight12';
-import IconArrowAlignRight16 from '@foundation/Icon/Generated/IconArrowAlignRight16';
-import IconArrowAlignRight20 from '@foundation/Icon/Generated/IconArrowAlignRight20';
-import IconArrowAlignRight24 from '@foundation/Icon/Generated/IconArrowAlignRight24';
-import IconArrowAlignRight32 from '@foundation/Icon/Generated/IconArrowAlignRight32';
-import IconArrowAlignUp from '@foundation/Icon/Generated/IconArrowAlignUp';
-import IconArrowAlignUp12 from '@foundation/Icon/Generated/IconArrowAlignUp12';
-import IconArrowAlignUp16 from '@foundation/Icon/Generated/IconArrowAlignUp16';
-import IconArrowAlignUp20 from '@foundation/Icon/Generated/IconArrowAlignUp20';
-import IconArrowAlignUp24 from '@foundation/Icon/Generated/IconArrowAlignUp24';
-import IconArrowAlignUp32 from '@foundation/Icon/Generated/IconArrowAlignUp32';
-import IconArrowAlignVerticalCentre from '@foundation/Icon/Generated/IconArrowAlignVerticalCentre';
-import IconArrowAlignVerticalCentre12 from '@foundation/Icon/Generated/IconArrowAlignVerticalCentre12';
-import IconArrowAlignVerticalCentre16 from '@foundation/Icon/Generated/IconArrowAlignVerticalCentre16';
-import IconArrowAlignVerticalCentre20 from '@foundation/Icon/Generated/IconArrowAlignVerticalCentre20';
-import IconArrowAlignVerticalCentre24 from '@foundation/Icon/Generated/IconArrowAlignVerticalCentre24';
-import IconArrowAlignVerticalCentre32 from '@foundation/Icon/Generated/IconArrowAlignVerticalCentre32';
-import IconArrowBidirectional from '@foundation/Icon/Generated/IconArrowBidirectional';
-import IconArrowBidirectional12 from '@foundation/Icon/Generated/IconArrowBidirectional12';
-import IconArrowBidirectional16 from '@foundation/Icon/Generated/IconArrowBidirectional16';
-import IconArrowBidirectional20 from '@foundation/Icon/Generated/IconArrowBidirectional20';
-import IconArrowBidirectional24 from '@foundation/Icon/Generated/IconArrowBidirectional24';
-import IconArrowBidirectional32 from '@foundation/Icon/Generated/IconArrowBidirectional32';
-import IconArrowBidirectionalHorizontal from '@foundation/Icon/Generated/IconArrowBidirectionalHorizontal';
-import IconArrowBidirectionalHorizontal12 from '@foundation/Icon/Generated/IconArrowBidirectionalHorizontal12';
-import IconArrowBidirectionalHorizontal16 from '@foundation/Icon/Generated/IconArrowBidirectionalHorizontal16';
-import IconArrowBidirectionalHorizontal20 from '@foundation/Icon/Generated/IconArrowBidirectionalHorizontal20';
-import IconArrowBidirectionalHorizontal24 from '@foundation/Icon/Generated/IconArrowBidirectionalHorizontal24';
-import IconArrowBidirectionalHorizontal32 from '@foundation/Icon/Generated/IconArrowBidirectionalHorizontal32';
-import IconArrowBottomToTop from '@foundation/Icon/Generated/IconArrowBottomToTop';
-import IconArrowBottomToTop12 from '@foundation/Icon/Generated/IconArrowBottomToTop12';
-import IconArrowBottomToTop16 from '@foundation/Icon/Generated/IconArrowBottomToTop16';
-import IconArrowBottomToTop20 from '@foundation/Icon/Generated/IconArrowBottomToTop20';
-import IconArrowBottomToTop24 from '@foundation/Icon/Generated/IconArrowBottomToTop24';
-import IconArrowBottomToTop32 from '@foundation/Icon/Generated/IconArrowBottomToTop32';
-import IconArrowCircleDown from '@foundation/Icon/Generated/IconArrowCircleDown';
-import IconArrowCircleDown12 from '@foundation/Icon/Generated/IconArrowCircleDown12';
-import IconArrowCircleDown16 from '@foundation/Icon/Generated/IconArrowCircleDown16';
-import IconArrowCircleDown20 from '@foundation/Icon/Generated/IconArrowCircleDown20';
-import IconArrowCircleDown24 from '@foundation/Icon/Generated/IconArrowCircleDown24';
-import IconArrowCircleDown32 from '@foundation/Icon/Generated/IconArrowCircleDown32';
-import IconArrowCircleUp from '@foundation/Icon/Generated/IconArrowCircleUp';
-import IconArrowCircleUp12 from '@foundation/Icon/Generated/IconArrowCircleUp12';
-import IconArrowCircleUp16 from '@foundation/Icon/Generated/IconArrowCircleUp16';
-import IconArrowCircleUp20 from '@foundation/Icon/Generated/IconArrowCircleUp20';
-import IconArrowCircleUp24 from '@foundation/Icon/Generated/IconArrowCircleUp24';
-import IconArrowCircleUp32 from '@foundation/Icon/Generated/IconArrowCircleUp32';
-import IconArrowDown from '@foundation/Icon/Generated/IconArrowDown';
-import IconArrowDown12 from '@foundation/Icon/Generated/IconArrowDown12';
-import IconArrowDown16 from '@foundation/Icon/Generated/IconArrowDown16';
-import IconArrowDown20 from '@foundation/Icon/Generated/IconArrowDown20';
-import IconArrowDown24 from '@foundation/Icon/Generated/IconArrowDown24';
-import IconArrowDown32 from '@foundation/Icon/Generated/IconArrowDown32';
-import IconArrowExpand from '@foundation/Icon/Generated/IconArrowExpand';
-import IconArrowExpand12 from '@foundation/Icon/Generated/IconArrowExpand12';
-import IconArrowExpand16 from '@foundation/Icon/Generated/IconArrowExpand16';
-import IconArrowExpand20 from '@foundation/Icon/Generated/IconArrowExpand20';
-import IconArrowExpand24 from '@foundation/Icon/Generated/IconArrowExpand24';
-import IconArrowExpand32 from '@foundation/Icon/Generated/IconArrowExpand32';
-import IconArrowFitBottomLeft from '@foundation/Icon/Generated/IconArrowFitBottomLeft';
-import IconArrowFitBottomLeft12 from '@foundation/Icon/Generated/IconArrowFitBottomLeft12';
-import IconArrowFitBottomLeft16 from '@foundation/Icon/Generated/IconArrowFitBottomLeft16';
-import IconArrowFitBottomLeft20 from '@foundation/Icon/Generated/IconArrowFitBottomLeft20';
-import IconArrowFitBottomLeft24 from '@foundation/Icon/Generated/IconArrowFitBottomLeft24';
-import IconArrowFitBottomLeft32 from '@foundation/Icon/Generated/IconArrowFitBottomLeft32';
-import IconArrowFitBottomRight from '@foundation/Icon/Generated/IconArrowFitBottomRight';
-import IconArrowFitBottomRight12 from '@foundation/Icon/Generated/IconArrowFitBottomRight12';
-import IconArrowFitBottomRight16 from '@foundation/Icon/Generated/IconArrowFitBottomRight16';
-import IconArrowFitBottomRight20 from '@foundation/Icon/Generated/IconArrowFitBottomRight20';
-import IconArrowFitBottomRight24 from '@foundation/Icon/Generated/IconArrowFitBottomRight24';
-import IconArrowFitBottomRight32 from '@foundation/Icon/Generated/IconArrowFitBottomRight32';
-import IconArrowFitTopLeft from '@foundation/Icon/Generated/IconArrowFitTopLeft';
-import IconArrowFitTopLeft12 from '@foundation/Icon/Generated/IconArrowFitTopLeft12';
-import IconArrowFitTopLeft16 from '@foundation/Icon/Generated/IconArrowFitTopLeft16';
-import IconArrowFitTopLeft20 from '@foundation/Icon/Generated/IconArrowFitTopLeft20';
-import IconArrowFitTopLeft24 from '@foundation/Icon/Generated/IconArrowFitTopLeft24';
-import IconArrowFitTopLeft32 from '@foundation/Icon/Generated/IconArrowFitTopLeft32';
-import IconArrowFitTopRight from '@foundation/Icon/Generated/IconArrowFitTopRight';
-import IconArrowFitTopRight12 from '@foundation/Icon/Generated/IconArrowFitTopRight12';
-import IconArrowFitTopRight16 from '@foundation/Icon/Generated/IconArrowFitTopRight16';
-import IconArrowFitTopRight20 from '@foundation/Icon/Generated/IconArrowFitTopRight20';
-import IconArrowFitTopRight24 from '@foundation/Icon/Generated/IconArrowFitTopRight24';
-import IconArrowFitTopRight32 from '@foundation/Icon/Generated/IconArrowFitTopRight32';
-import IconArrowHook from '@foundation/Icon/Generated/IconArrowHook';
-import IconArrowHook12 from '@foundation/Icon/Generated/IconArrowHook12';
-import IconArrowHook16 from '@foundation/Icon/Generated/IconArrowHook16';
-import IconArrowHook20 from '@foundation/Icon/Generated/IconArrowHook20';
-import IconArrowHook24 from '@foundation/Icon/Generated/IconArrowHook24';
-import IconArrowHook32 from '@foundation/Icon/Generated/IconArrowHook32';
-import IconArrowLeft from '@foundation/Icon/Generated/IconArrowLeft';
-import IconArrowLeft12 from '@foundation/Icon/Generated/IconArrowLeft12';
-import IconArrowLeft16 from '@foundation/Icon/Generated/IconArrowLeft16';
-import IconArrowLeft20 from '@foundation/Icon/Generated/IconArrowLeft20';
-import IconArrowLeft24 from '@foundation/Icon/Generated/IconArrowLeft24';
-import IconArrowLeft32 from '@foundation/Icon/Generated/IconArrowLeft32';
-import IconArrowLeftToRight from '@foundation/Icon/Generated/IconArrowLeftToRight';
-import IconArrowLeftToRight12 from '@foundation/Icon/Generated/IconArrowLeftToRight12';
-import IconArrowLeftToRight16 from '@foundation/Icon/Generated/IconArrowLeftToRight16';
-import IconArrowLeftToRight20 from '@foundation/Icon/Generated/IconArrowLeftToRight20';
-import IconArrowLeftToRight24 from '@foundation/Icon/Generated/IconArrowLeftToRight24';
-import IconArrowLeftToRight32 from '@foundation/Icon/Generated/IconArrowLeftToRight32';
-import IconArrowMinimize from '@foundation/Icon/Generated/IconArrowMinimize';
-import IconArrowMinimize12 from '@foundation/Icon/Generated/IconArrowMinimize12';
-import IconArrowMinimize16 from '@foundation/Icon/Generated/IconArrowMinimize16';
-import IconArrowMinimize20 from '@foundation/Icon/Generated/IconArrowMinimize20';
-import IconArrowMinimize24 from '@foundation/Icon/Generated/IconArrowMinimize24';
-import IconArrowMinimize32 from '@foundation/Icon/Generated/IconArrowMinimize32';
-import IconArrowMove from '@foundation/Icon/Generated/IconArrowMove';
-import IconArrowMove12 from '@foundation/Icon/Generated/IconArrowMove12';
-import IconArrowMove16 from '@foundation/Icon/Generated/IconArrowMove16';
-import IconArrowMove20 from '@foundation/Icon/Generated/IconArrowMove20';
-import IconArrowMove24 from '@foundation/Icon/Generated/IconArrowMove24';
-import IconArrowMove32 from '@foundation/Icon/Generated/IconArrowMove32';
-import IconArrowOutExternal from '@foundation/Icon/Generated/IconArrowOutExternal';
-import IconArrowOutExternal12 from '@foundation/Icon/Generated/IconArrowOutExternal12';
-import IconArrowOutExternal16 from '@foundation/Icon/Generated/IconArrowOutExternal16';
-import IconArrowOutExternal20 from '@foundation/Icon/Generated/IconArrowOutExternal20';
-import IconArrowOutExternal24 from '@foundation/Icon/Generated/IconArrowOutExternal24';
-import IconArrowOutExternal32 from '@foundation/Icon/Generated/IconArrowOutExternal32';
-import IconArrowOutLogout from '@foundation/Icon/Generated/IconArrowOutLogout';
-import IconArrowOutLogout12 from '@foundation/Icon/Generated/IconArrowOutLogout12';
-import IconArrowOutLogout16 from '@foundation/Icon/Generated/IconArrowOutLogout16';
-import IconArrowOutLogout20 from '@foundation/Icon/Generated/IconArrowOutLogout20';
-import IconArrowOutLogout24 from '@foundation/Icon/Generated/IconArrowOutLogout24';
-import IconArrowOutLogout32 from '@foundation/Icon/Generated/IconArrowOutLogout32';
-import IconArrowOutShare from '@foundation/Icon/Generated/IconArrowOutShare';
-import IconArrowOutShare12 from '@foundation/Icon/Generated/IconArrowOutShare12';
-import IconArrowOutShare16 from '@foundation/Icon/Generated/IconArrowOutShare16';
-import IconArrowOutShare20 from '@foundation/Icon/Generated/IconArrowOutShare20';
-import IconArrowOutShare24 from '@foundation/Icon/Generated/IconArrowOutShare24';
-import IconArrowOutShare32 from '@foundation/Icon/Generated/IconArrowOutShare32';
-import IconArrowRight from '@foundation/Icon/Generated/IconArrowRight';
-import IconArrowRight12 from '@foundation/Icon/Generated/IconArrowRight12';
-import IconArrowRight16 from '@foundation/Icon/Generated/IconArrowRight16';
-import IconArrowRight20 from '@foundation/Icon/Generated/IconArrowRight20';
-import IconArrowRight24 from '@foundation/Icon/Generated/IconArrowRight24';
-import IconArrowRight32 from '@foundation/Icon/Generated/IconArrowRight32';
-import IconArrowRightToLeft from '@foundation/Icon/Generated/IconArrowRightToLeft';
-import IconArrowRightToLeft12 from '@foundation/Icon/Generated/IconArrowRightToLeft12';
-import IconArrowRightToLeft16 from '@foundation/Icon/Generated/IconArrowRightToLeft16';
-import IconArrowRightToLeft20 from '@foundation/Icon/Generated/IconArrowRightToLeft20';
-import IconArrowRightToLeft24 from '@foundation/Icon/Generated/IconArrowRightToLeft24';
-import IconArrowRightToLeft32 from '@foundation/Icon/Generated/IconArrowRightToLeft32';
-import IconArrowRoundAntiClockwise from '@foundation/Icon/Generated/IconArrowRoundAntiClockwise';
-import IconArrowRoundAntiClockwise12 from '@foundation/Icon/Generated/IconArrowRoundAntiClockwise12';
-import IconArrowRoundAntiClockwise16 from '@foundation/Icon/Generated/IconArrowRoundAntiClockwise16';
-import IconArrowRoundAntiClockwise20 from '@foundation/Icon/Generated/IconArrowRoundAntiClockwise20';
-import IconArrowRoundAntiClockwise24 from '@foundation/Icon/Generated/IconArrowRoundAntiClockwise24';
-import IconArrowRoundAntiClockwise32 from '@foundation/Icon/Generated/IconArrowRoundAntiClockwise32';
-import IconArrowRoundClockwise from '@foundation/Icon/Generated/IconArrowRoundClockwise';
-import IconArrowRoundClockwise12 from '@foundation/Icon/Generated/IconArrowRoundClockwise12';
-import IconArrowRoundClockwise16 from '@foundation/Icon/Generated/IconArrowRoundClockwise16';
-import IconArrowRoundClockwise20 from '@foundation/Icon/Generated/IconArrowRoundClockwise20';
-import IconArrowRoundClockwise24 from '@foundation/Icon/Generated/IconArrowRoundClockwise24';
-import IconArrowRoundClockwise32 from '@foundation/Icon/Generated/IconArrowRoundClockwise32';
-import IconArrowScaleHorizontal from '@foundation/Icon/Generated/IconArrowScaleHorizontal';
-import IconArrowScaleHorizontal12 from '@foundation/Icon/Generated/IconArrowScaleHorizontal12';
-import IconArrowScaleHorizontal16 from '@foundation/Icon/Generated/IconArrowScaleHorizontal16';
-import IconArrowScaleHorizontal20 from '@foundation/Icon/Generated/IconArrowScaleHorizontal20';
-import IconArrowScaleHorizontal24 from '@foundation/Icon/Generated/IconArrowScaleHorizontal24';
-import IconArrowScaleHorizontal32 from '@foundation/Icon/Generated/IconArrowScaleHorizontal32';
-import IconArrowScaleVertical from '@foundation/Icon/Generated/IconArrowScaleVertical';
-import IconArrowScaleVertical12 from '@foundation/Icon/Generated/IconArrowScaleVertical12';
-import IconArrowScaleVertical16 from '@foundation/Icon/Generated/IconArrowScaleVertical16';
-import IconArrowScaleVertical20 from '@foundation/Icon/Generated/IconArrowScaleVertical20';
-import IconArrowScaleVertical24 from '@foundation/Icon/Generated/IconArrowScaleVertical24';
-import IconArrowScaleVertical32 from '@foundation/Icon/Generated/IconArrowScaleVertical32';
-import IconArrowStretchBox from '@foundation/Icon/Generated/IconArrowStretchBox';
-import IconArrowStretchBox12 from '@foundation/Icon/Generated/IconArrowStretchBox12';
-import IconArrowStretchBox16 from '@foundation/Icon/Generated/IconArrowStretchBox16';
-import IconArrowStretchBox20 from '@foundation/Icon/Generated/IconArrowStretchBox20';
-import IconArrowStretchBox24 from '@foundation/Icon/Generated/IconArrowStretchBox24';
-import IconArrowStretchBox32 from '@foundation/Icon/Generated/IconArrowStretchBox32';
-import IconArrowSwap from '@foundation/Icon/Generated/IconArrowSwap';
-import IconArrowSwap12 from '@foundation/Icon/Generated/IconArrowSwap12';
-import IconArrowSwap16 from '@foundation/Icon/Generated/IconArrowSwap16';
-import IconArrowSwap20 from '@foundation/Icon/Generated/IconArrowSwap20';
-import IconArrowSwap24 from '@foundation/Icon/Generated/IconArrowSwap24';
-import IconArrowSwap32 from '@foundation/Icon/Generated/IconArrowSwap32';
-import IconArrowSync from '@foundation/Icon/Generated/IconArrowSync';
-import IconArrowSync12 from '@foundation/Icon/Generated/IconArrowSync12';
-import IconArrowSync16 from '@foundation/Icon/Generated/IconArrowSync16';
-import IconArrowSync20 from '@foundation/Icon/Generated/IconArrowSync20';
-import IconArrowSync24 from '@foundation/Icon/Generated/IconArrowSync24';
-import IconArrowSync32 from '@foundation/Icon/Generated/IconArrowSync32';
-import IconArrowTopToBottom from '@foundation/Icon/Generated/IconArrowTopToBottom';
-import IconArrowTopToBottom12 from '@foundation/Icon/Generated/IconArrowTopToBottom12';
-import IconArrowTopToBottom16 from '@foundation/Icon/Generated/IconArrowTopToBottom16';
-import IconArrowTopToBottom20 from '@foundation/Icon/Generated/IconArrowTopToBottom20';
-import IconArrowTopToBottom24 from '@foundation/Icon/Generated/IconArrowTopToBottom24';
-import IconArrowTopToBottom32 from '@foundation/Icon/Generated/IconArrowTopToBottom32';
-import IconArrowUp from '@foundation/Icon/Generated/IconArrowUp';
-import IconArrowUp12 from '@foundation/Icon/Generated/IconArrowUp12';
-import IconArrowUp16 from '@foundation/Icon/Generated/IconArrowUp16';
-import IconArrowUp20 from '@foundation/Icon/Generated/IconArrowUp20';
-import IconArrowUp24 from '@foundation/Icon/Generated/IconArrowUp24';
-import IconArrowUp32 from '@foundation/Icon/Generated/IconArrowUp32';
-import IconBar from '@foundation/Icon/Generated/IconBar';
-import IconBar12 from '@foundation/Icon/Generated/IconBar12';
-import IconBar16 from '@foundation/Icon/Generated/IconBar16';
-import IconBar20 from '@foundation/Icon/Generated/IconBar20';
-import IconBar24 from '@foundation/Icon/Generated/IconBar24';
-import IconBar32 from '@foundation/Icon/Generated/IconBar32';
-import IconBarNarrow from '@foundation/Icon/Generated/IconBarNarrow';
-import IconBarNarrow12 from '@foundation/Icon/Generated/IconBarNarrow12';
-import IconBarNarrow16 from '@foundation/Icon/Generated/IconBarNarrow16';
-import IconBarNarrow20 from '@foundation/Icon/Generated/IconBarNarrow20';
-import IconBarNarrow24 from '@foundation/Icon/Generated/IconBarNarrow24';
-import IconBarNarrow32 from '@foundation/Icon/Generated/IconBarNarrow32';
-import IconBasketball from '@foundation/Icon/Generated/IconBasketball';
-import IconBasketball12 from '@foundation/Icon/Generated/IconBasketball12';
-import IconBasketball16 from '@foundation/Icon/Generated/IconBasketball16';
-import IconBasketball20 from '@foundation/Icon/Generated/IconBasketball20';
-import IconBasketball24 from '@foundation/Icon/Generated/IconBasketball24';
-import IconBasketball32 from '@foundation/Icon/Generated/IconBasketball32';
-import IconBell from '@foundation/Icon/Generated/IconBell';
-import IconBell12 from '@foundation/Icon/Generated/IconBell12';
-import IconBell16 from '@foundation/Icon/Generated/IconBell16';
-import IconBell20 from '@foundation/Icon/Generated/IconBell20';
-import IconBell24 from '@foundation/Icon/Generated/IconBell24';
-import IconBell32 from '@foundation/Icon/Generated/IconBell32';
-import IconBookmark from '@foundation/Icon/Generated/IconBookmark';
-import IconBookmark12 from '@foundation/Icon/Generated/IconBookmark12';
-import IconBookmark12Filled from '@foundation/Icon/Generated/IconBookmark12Filled';
-import IconBookmark16 from '@foundation/Icon/Generated/IconBookmark16';
-import IconBookmark16Filled from '@foundation/Icon/Generated/IconBookmark16Filled';
-import IconBookmark20 from '@foundation/Icon/Generated/IconBookmark20';
-import IconBookmark20Filled from '@foundation/Icon/Generated/IconBookmark20Filled';
-import IconBookmark24 from '@foundation/Icon/Generated/IconBookmark24';
-import IconBookmark24Filled from '@foundation/Icon/Generated/IconBookmark24Filled';
-import IconBookmark32 from '@foundation/Icon/Generated/IconBookmark32';
-import IconBookmark32Filled from '@foundation/Icon/Generated/IconBookmark32Filled';
-import IconBracketsCurly from '@foundation/Icon/Generated/IconBracketsCurly';
-import IconBracketsCurly12 from '@foundation/Icon/Generated/IconBracketsCurly12';
-import IconBracketsCurly16 from '@foundation/Icon/Generated/IconBracketsCurly16';
-import IconBracketsCurly20 from '@foundation/Icon/Generated/IconBracketsCurly20';
-import IconBracketsCurly24 from '@foundation/Icon/Generated/IconBracketsCurly24';
-import IconBracketsCurly32 from '@foundation/Icon/Generated/IconBracketsCurly32';
-import IconBuilding from '@foundation/Icon/Generated/IconBuilding';
-import IconBuilding12 from '@foundation/Icon/Generated/IconBuilding12';
-import IconBuilding16 from '@foundation/Icon/Generated/IconBuilding16';
-import IconBuilding20 from '@foundation/Icon/Generated/IconBuilding20';
-import IconBuilding24 from '@foundation/Icon/Generated/IconBuilding24';
-import IconBuilding32 from '@foundation/Icon/Generated/IconBuilding32';
-import IconBuildingBlock from '@foundation/Icon/Generated/IconBuildingBlock';
-import IconBuildingBlock12 from '@foundation/Icon/Generated/IconBuildingBlock12';
-import IconBuildingBlock16 from '@foundation/Icon/Generated/IconBuildingBlock16';
-import IconBuildingBlock20 from '@foundation/Icon/Generated/IconBuildingBlock20';
-import IconBuildingBlock24 from '@foundation/Icon/Generated/IconBuildingBlock24';
-import IconBuildingBlock32 from '@foundation/Icon/Generated/IconBuildingBlock32';
-import IconBuildings from '@foundation/Icon/Generated/IconBuildings';
-import IconBuildings12 from '@foundation/Icon/Generated/IconBuildings12';
-import IconBuildings16 from '@foundation/Icon/Generated/IconBuildings16';
-import IconBuildings20 from '@foundation/Icon/Generated/IconBuildings20';
-import IconBuildings24 from '@foundation/Icon/Generated/IconBuildings24';
-import IconBuildings32 from '@foundation/Icon/Generated/IconBuildings32';
-import IconButton from '@foundation/Icon/Generated/IconButton';
-import IconButton12 from '@foundation/Icon/Generated/IconButton12';
-import IconButton16 from '@foundation/Icon/Generated/IconButton16';
-import IconButton20 from '@foundation/Icon/Generated/IconButton20';
-import IconButton24 from '@foundation/Icon/Generated/IconButton24';
-import IconButton32 from '@foundation/Icon/Generated/IconButton32';
-import IconCalendar from '@foundation/Icon/Generated/IconCalendar';
-import IconCalendar12 from '@foundation/Icon/Generated/IconCalendar12';
-import IconCalendar16 from '@foundation/Icon/Generated/IconCalendar16';
-import IconCalendar20 from '@foundation/Icon/Generated/IconCalendar20';
-import IconCalendar24 from '@foundation/Icon/Generated/IconCalendar24';
-import IconCalendar32 from '@foundation/Icon/Generated/IconCalendar32';
-import IconCalendarChecked from '@foundation/Icon/Generated/IconCalendarChecked';
-import IconCalendarChecked12 from '@foundation/Icon/Generated/IconCalendarChecked12';
-import IconCalendarChecked16 from '@foundation/Icon/Generated/IconCalendarChecked16';
-import IconCalendarChecked20 from '@foundation/Icon/Generated/IconCalendarChecked20';
-import IconCalendarChecked24 from '@foundation/Icon/Generated/IconCalendarChecked24';
-import IconCalendarChecked32 from '@foundation/Icon/Generated/IconCalendarChecked32';
-import IconCap from '@foundation/Icon/Generated/IconCap';
-import IconCap12 from '@foundation/Icon/Generated/IconCap12';
-import IconCap16 from '@foundation/Icon/Generated/IconCap16';
-import IconCap20 from '@foundation/Icon/Generated/IconCap20';
-import IconCap24 from '@foundation/Icon/Generated/IconCap24';
-import IconCap32 from '@foundation/Icon/Generated/IconCap32';
-import IconCard from '@foundation/Icon/Generated/IconCard';
-import IconCard12 from '@foundation/Icon/Generated/IconCard12';
-import IconCard16 from '@foundation/Icon/Generated/IconCard16';
-import IconCard20 from '@foundation/Icon/Generated/IconCard20';
-import IconCard24 from '@foundation/Icon/Generated/IconCard24';
-import IconCard32 from '@foundation/Icon/Generated/IconCard32';
-import IconCardLayoutLarge from '@foundation/Icon/Generated/IconCardLayoutLarge';
-import IconCardLayoutLarge12 from '@foundation/Icon/Generated/IconCardLayoutLarge12';
-import IconCardLayoutLarge16 from '@foundation/Icon/Generated/IconCardLayoutLarge16';
-import IconCardLayoutLarge20 from '@foundation/Icon/Generated/IconCardLayoutLarge20';
-import IconCardLayoutLarge24 from '@foundation/Icon/Generated/IconCardLayoutLarge24';
-import IconCardLayoutLarge32 from '@foundation/Icon/Generated/IconCardLayoutLarge32';
-import IconCardLayoutMedium from '@foundation/Icon/Generated/IconCardLayoutMedium';
-import IconCardLayoutMedium12 from '@foundation/Icon/Generated/IconCardLayoutMedium12';
-import IconCardLayoutMedium16 from '@foundation/Icon/Generated/IconCardLayoutMedium16';
-import IconCardLayoutMedium20 from '@foundation/Icon/Generated/IconCardLayoutMedium20';
-import IconCardLayoutMedium24 from '@foundation/Icon/Generated/IconCardLayoutMedium24';
-import IconCardLayoutMedium32 from '@foundation/Icon/Generated/IconCardLayoutMedium32';
-import IconCardLayoutSmall from '@foundation/Icon/Generated/IconCardLayoutSmall';
-import IconCardLayoutSmall12 from '@foundation/Icon/Generated/IconCardLayoutSmall12';
-import IconCardLayoutSmall16 from '@foundation/Icon/Generated/IconCardLayoutSmall16';
-import IconCardLayoutSmall20 from '@foundation/Icon/Generated/IconCardLayoutSmall20';
-import IconCardLayoutSmall24 from '@foundation/Icon/Generated/IconCardLayoutSmall24';
-import IconCardLayoutSmall32 from '@foundation/Icon/Generated/IconCardLayoutSmall32';
-import IconCaretDown from '@foundation/Icon/Generated/IconCaretDown';
-import IconCaretDown12 from '@foundation/Icon/Generated/IconCaretDown12';
-import IconCaretDown16 from '@foundation/Icon/Generated/IconCaretDown16';
-import IconCaretDown20 from '@foundation/Icon/Generated/IconCaretDown20';
-import IconCaretDown24 from '@foundation/Icon/Generated/IconCaretDown24';
-import IconCaretDown32 from '@foundation/Icon/Generated/IconCaretDown32';
-import IconCaretLeft from '@foundation/Icon/Generated/IconCaretLeft';
-import IconCaretLeft12 from '@foundation/Icon/Generated/IconCaretLeft12';
-import IconCaretLeft16 from '@foundation/Icon/Generated/IconCaretLeft16';
-import IconCaretLeft20 from '@foundation/Icon/Generated/IconCaretLeft20';
-import IconCaretLeft24 from '@foundation/Icon/Generated/IconCaretLeft24';
-import IconCaretLeft32 from '@foundation/Icon/Generated/IconCaretLeft32';
-import IconCaretLeftDouble from '@foundation/Icon/Generated/IconCaretLeftDouble';
-import IconCaretLeftDouble12 from '@foundation/Icon/Generated/IconCaretLeftDouble12';
-import IconCaretLeftDouble16 from '@foundation/Icon/Generated/IconCaretLeftDouble16';
-import IconCaretLeftDouble20 from '@foundation/Icon/Generated/IconCaretLeftDouble20';
-import IconCaretLeftDouble24 from '@foundation/Icon/Generated/IconCaretLeftDouble24';
-import IconCaretLeftDouble32 from '@foundation/Icon/Generated/IconCaretLeftDouble32';
-import IconCaretRight from '@foundation/Icon/Generated/IconCaretRight';
-import IconCaretRight12 from '@foundation/Icon/Generated/IconCaretRight12';
-import IconCaretRight16 from '@foundation/Icon/Generated/IconCaretRight16';
-import IconCaretRight20 from '@foundation/Icon/Generated/IconCaretRight20';
-import IconCaretRight24 from '@foundation/Icon/Generated/IconCaretRight24';
-import IconCaretRight32 from '@foundation/Icon/Generated/IconCaretRight32';
-import IconCaretRightDouble from '@foundation/Icon/Generated/IconCaretRightDouble';
-import IconCaretRightDouble12 from '@foundation/Icon/Generated/IconCaretRightDouble12';
-import IconCaretRightDouble16 from '@foundation/Icon/Generated/IconCaretRightDouble16';
-import IconCaretRightDouble20 from '@foundation/Icon/Generated/IconCaretRightDouble20';
-import IconCaretRightDouble24 from '@foundation/Icon/Generated/IconCaretRightDouble24';
-import IconCaretRightDouble32 from '@foundation/Icon/Generated/IconCaretRightDouble32';
-import IconCaretUp from '@foundation/Icon/Generated/IconCaretUp';
-import IconCaretUp12 from '@foundation/Icon/Generated/IconCaretUp12';
-import IconCaretUp16 from '@foundation/Icon/Generated/IconCaretUp16';
-import IconCaretUp20 from '@foundation/Icon/Generated/IconCaretUp20';
-import IconCaretUp24 from '@foundation/Icon/Generated/IconCaretUp24';
-import IconCaretUp32 from '@foundation/Icon/Generated/IconCaretUp32';
-import IconCarousel from '@foundation/Icon/Generated/IconCarousel';
-import IconCarousel12 from '@foundation/Icon/Generated/IconCarousel12';
-import IconCarousel16 from '@foundation/Icon/Generated/IconCarousel16';
-import IconCarousel20 from '@foundation/Icon/Generated/IconCarousel20';
-import IconCarousel24 from '@foundation/Icon/Generated/IconCarousel24';
-import IconCarousel32 from '@foundation/Icon/Generated/IconCarousel32';
-import IconChartBars from '@foundation/Icon/Generated/IconChartBars';
-import IconChartBars12 from '@foundation/Icon/Generated/IconChartBars12';
-import IconChartBars16 from '@foundation/Icon/Generated/IconChartBars16';
-import IconChartBars20 from '@foundation/Icon/Generated/IconChartBars20';
-import IconChartBars24 from '@foundation/Icon/Generated/IconChartBars24';
-import IconChartBars32 from '@foundation/Icon/Generated/IconChartBars32';
-import IconChartLine from '@foundation/Icon/Generated/IconChartLine';
-import IconChartLine12 from '@foundation/Icon/Generated/IconChartLine12';
-import IconChartLine16 from '@foundation/Icon/Generated/IconChartLine16';
-import IconChartLine20 from '@foundation/Icon/Generated/IconChartLine20';
-import IconChartLine24 from '@foundation/Icon/Generated/IconChartLine24';
-import IconChartLine32 from '@foundation/Icon/Generated/IconChartLine32';
-import IconChartPie from '@foundation/Icon/Generated/IconChartPie';
-import IconChartPie12 from '@foundation/Icon/Generated/IconChartPie12';
-import IconChartPie16 from '@foundation/Icon/Generated/IconChartPie16';
-import IconChartPie20 from '@foundation/Icon/Generated/IconChartPie20';
-import IconChartPie24 from '@foundation/Icon/Generated/IconChartPie24';
-import IconChartPie32 from '@foundation/Icon/Generated/IconChartPie32';
-import IconCheckMark from '@foundation/Icon/Generated/IconCheckMark';
-import IconCheckMark12 from '@foundation/Icon/Generated/IconCheckMark12';
-import IconCheckMark16 from '@foundation/Icon/Generated/IconCheckMark16';
-import IconCheckMark20 from '@foundation/Icon/Generated/IconCheckMark20';
-import IconCheckMark24 from '@foundation/Icon/Generated/IconCheckMark24';
-import IconCheckMark32 from '@foundation/Icon/Generated/IconCheckMark32';
-import IconCheckMarkCircle from '@foundation/Icon/Generated/IconCheckMarkCircle';
-import IconCheckMarkCircle12 from '@foundation/Icon/Generated/IconCheckMarkCircle12';
-import IconCheckMarkCircle12Filled from '@foundation/Icon/Generated/IconCheckMarkCircle12Filled';
-import IconCheckMarkCircle16 from '@foundation/Icon/Generated/IconCheckMarkCircle16';
-import IconCheckMarkCircle16Filled from '@foundation/Icon/Generated/IconCheckMarkCircle16Filled';
-import IconCheckMarkCircle20 from '@foundation/Icon/Generated/IconCheckMarkCircle20';
-import IconCheckMarkCircle20Filled from '@foundation/Icon/Generated/IconCheckMarkCircle20Filled';
-import IconCheckMarkCircle24 from '@foundation/Icon/Generated/IconCheckMarkCircle24';
-import IconCheckMarkCircle24Filled from '@foundation/Icon/Generated/IconCheckMarkCircle24Filled';
-import IconCheckMarkCircle32 from '@foundation/Icon/Generated/IconCheckMarkCircle32';
-import IconCheckMarkCircle32Filled from '@foundation/Icon/Generated/IconCheckMarkCircle32Filled';
-import IconCheckbox from '@foundation/Icon/Generated/IconCheckbox';
-import IconCheckbox12 from '@foundation/Icon/Generated/IconCheckbox12';
-import IconCheckbox16 from '@foundation/Icon/Generated/IconCheckbox16';
-import IconCheckbox20 from '@foundation/Icon/Generated/IconCheckbox20';
-import IconCheckbox24 from '@foundation/Icon/Generated/IconCheckbox24';
-import IconCheckbox32 from '@foundation/Icon/Generated/IconCheckbox32';
-import IconCheckmarkCircleBox from '@foundation/Icon/Generated/IconCheckmarkCircleBox';
-import IconCheckmarkCircleBox12 from '@foundation/Icon/Generated/IconCheckmarkCircleBox12';
-import IconCheckmarkCircleBox16 from '@foundation/Icon/Generated/IconCheckmarkCircleBox16';
-import IconCheckmarkCircleBox20 from '@foundation/Icon/Generated/IconCheckmarkCircleBox20';
-import IconCheckmarkCircleBox24 from '@foundation/Icon/Generated/IconCheckmarkCircleBox24';
-import IconCheckmarkCircleBox32 from '@foundation/Icon/Generated/IconCheckmarkCircleBox32';
-import IconCircle from '@foundation/Icon/Generated/IconCircle';
-import IconCircle12 from '@foundation/Icon/Generated/IconCircle12';
-import IconCircle16 from '@foundation/Icon/Generated/IconCircle16';
-import IconCircle20 from '@foundation/Icon/Generated/IconCircle20';
-import IconCircle24 from '@foundation/Icon/Generated/IconCircle24';
-import IconCircle32 from '@foundation/Icon/Generated/IconCircle32';
-import IconClipboard from '@foundation/Icon/Generated/IconClipboard';
-import IconClipboard12 from '@foundation/Icon/Generated/IconClipboard12';
-import IconClipboard16 from '@foundation/Icon/Generated/IconClipboard16';
-import IconClipboard20 from '@foundation/Icon/Generated/IconClipboard20';
-import IconClipboard24 from '@foundation/Icon/Generated/IconClipboard24';
-import IconClipboard32 from '@foundation/Icon/Generated/IconClipboard32';
-import IconClipboardAt from '@foundation/Icon/Generated/IconClipboardAt';
-import IconClipboardAt12 from '@foundation/Icon/Generated/IconClipboardAt12';
-import IconClipboardAt16 from '@foundation/Icon/Generated/IconClipboardAt16';
-import IconClipboardAt20 from '@foundation/Icon/Generated/IconClipboardAt20';
-import IconClipboardAt24 from '@foundation/Icon/Generated/IconClipboardAt24';
-import IconClipboardAt32 from '@foundation/Icon/Generated/IconClipboardAt32';
-import IconClock from '@foundation/Icon/Generated/IconClock';
-import IconClock12 from '@foundation/Icon/Generated/IconClock12';
-import IconClock16 from '@foundation/Icon/Generated/IconClock16';
-import IconClock20 from '@foundation/Icon/Generated/IconClock20';
-import IconClock24 from '@foundation/Icon/Generated/IconClock24';
-import IconClock32 from '@foundation/Icon/Generated/IconClock32';
-import IconClockAlternative from '@foundation/Icon/Generated/IconClockAlternative';
-import IconClockAlternative1 from '@foundation/Icon/Generated/IconClockAlternative1';
-import IconClockAlternative112 from '@foundation/Icon/Generated/IconClockAlternative112';
-import IconClockAlternative116 from '@foundation/Icon/Generated/IconClockAlternative116';
-import IconClockAlternative12 from '@foundation/Icon/Generated/IconClockAlternative12';
-import IconClockAlternative120 from '@foundation/Icon/Generated/IconClockAlternative120';
-import IconClockAlternative124 from '@foundation/Icon/Generated/IconClockAlternative124';
-import IconClockAlternative132 from '@foundation/Icon/Generated/IconClockAlternative132';
-import IconClockAlternative16 from '@foundation/Icon/Generated/IconClockAlternative16';
-import IconClockAlternative20 from '@foundation/Icon/Generated/IconClockAlternative20';
-import IconClockAlternative24 from '@foundation/Icon/Generated/IconClockAlternative24';
-import IconClockAlternative32 from '@foundation/Icon/Generated/IconClockAlternative32';
-import IconClockList from '@foundation/Icon/Generated/IconClockList';
-import IconClockList12 from '@foundation/Icon/Generated/IconClockList12';
-import IconClockList16 from '@foundation/Icon/Generated/IconClockList16';
-import IconClockList20 from '@foundation/Icon/Generated/IconClockList20';
-import IconClockList24 from '@foundation/Icon/Generated/IconClockList24';
-import IconClockList32 from '@foundation/Icon/Generated/IconClockList32';
-import IconCloudArrowDown from '@foundation/Icon/Generated/IconCloudArrowDown';
-import IconCloudArrowDown12 from '@foundation/Icon/Generated/IconCloudArrowDown12';
-import IconCloudArrowDown16 from '@foundation/Icon/Generated/IconCloudArrowDown16';
-import IconCloudArrowDown20 from '@foundation/Icon/Generated/IconCloudArrowDown20';
-import IconCloudArrowDown24 from '@foundation/Icon/Generated/IconCloudArrowDown24';
-import IconCloudArrowDown32 from '@foundation/Icon/Generated/IconCloudArrowDown32';
-import IconCloudArrowUp from '@foundation/Icon/Generated/IconCloudArrowUp';
-import IconCloudArrowUp12 from '@foundation/Icon/Generated/IconCloudArrowUp12';
-import IconCloudArrowUp16 from '@foundation/Icon/Generated/IconCloudArrowUp16';
-import IconCloudArrowUp20 from '@foundation/Icon/Generated/IconCloudArrowUp20';
-import IconCloudArrowUp24 from '@foundation/Icon/Generated/IconCloudArrowUp24';
-import IconCloudArrowUp32 from '@foundation/Icon/Generated/IconCloudArrowUp32';
-import IconCode from '@foundation/Icon/Generated/IconCode';
-import IconCode12 from '@foundation/Icon/Generated/IconCode12';
-import IconCode16 from '@foundation/Icon/Generated/IconCode16';
-import IconCode20 from '@foundation/Icon/Generated/IconCode20';
-import IconCode24 from '@foundation/Icon/Generated/IconCode24';
-import IconCode32 from '@foundation/Icon/Generated/IconCode32';
-import IconCodeBox from '@foundation/Icon/Generated/IconCodeBox';
-import IconCodeBox12 from '@foundation/Icon/Generated/IconCodeBox12';
-import IconCodeBox16 from '@foundation/Icon/Generated/IconCodeBox16';
-import IconCodeBox20 from '@foundation/Icon/Generated/IconCodeBox20';
-import IconCodeBox24 from '@foundation/Icon/Generated/IconCodeBox24';
-import IconCodeBox32 from '@foundation/Icon/Generated/IconCodeBox32';
-import IconCodeDocument from '@foundation/Icon/Generated/IconCodeDocument';
-import IconCodeDocument12 from '@foundation/Icon/Generated/IconCodeDocument12';
-import IconCodeDocument16 from '@foundation/Icon/Generated/IconCodeDocument16';
-import IconCodeDocument20 from '@foundation/Icon/Generated/IconCodeDocument20';
-import IconCodeDocument24 from '@foundation/Icon/Generated/IconCodeDocument24';
-import IconCodeDocument32 from '@foundation/Icon/Generated/IconCodeDocument32';
-import IconCodeFrame from '@foundation/Icon/Generated/IconCodeFrame';
-import IconCodeFrame12 from '@foundation/Icon/Generated/IconCodeFrame12';
-import IconCodeFrame16 from '@foundation/Icon/Generated/IconCodeFrame16';
-import IconCodeFrame20 from '@foundation/Icon/Generated/IconCodeFrame20';
-import IconCodeFrame24 from '@foundation/Icon/Generated/IconCodeFrame24';
-import IconCodeFrame32 from '@foundation/Icon/Generated/IconCodeFrame32';
-import IconCog from '@foundation/Icon/Generated/IconCog';
-import IconCog12 from '@foundation/Icon/Generated/IconCog12';
-import IconCog16 from '@foundation/Icon/Generated/IconCog16';
-import IconCog20 from '@foundation/Icon/Generated/IconCog20';
-import IconCog24 from '@foundation/Icon/Generated/IconCog24';
-import IconCog32 from '@foundation/Icon/Generated/IconCog32';
-import IconColorFan from '@foundation/Icon/Generated/IconColorFan';
-import IconColorFan12 from '@foundation/Icon/Generated/IconColorFan12';
-import IconColorFan16 from '@foundation/Icon/Generated/IconColorFan16';
-import IconColorFan20 from '@foundation/Icon/Generated/IconColorFan20';
-import IconColorFan24 from '@foundation/Icon/Generated/IconColorFan24';
-import IconColorFan32 from '@foundation/Icon/Generated/IconColorFan32';
-import IconCommandLineBox from '@foundation/Icon/Generated/IconCommandLineBox';
-import IconCommandLineBox12 from '@foundation/Icon/Generated/IconCommandLineBox12';
-import IconCommandLineBox16 from '@foundation/Icon/Generated/IconCommandLineBox16';
-import IconCommandLineBox20 from '@foundation/Icon/Generated/IconCommandLineBox20';
-import IconCommandLineBox24 from '@foundation/Icon/Generated/IconCommandLineBox24';
-import IconCommandLineBox32 from '@foundation/Icon/Generated/IconCommandLineBox32';
-import IconContentBlock from '@foundation/Icon/Generated/IconContentBlock';
-import IconContentBlock12 from '@foundation/Icon/Generated/IconContentBlock12';
-import IconContentBlock16 from '@foundation/Icon/Generated/IconContentBlock16';
-import IconContentBlock20 from '@foundation/Icon/Generated/IconContentBlock20';
-import IconContentBlock24 from '@foundation/Icon/Generated/IconContentBlock24';
-import IconContentBlock32 from '@foundation/Icon/Generated/IconContentBlock32';
-import IconController from '@foundation/Icon/Generated/IconController';
-import IconController12 from '@foundation/Icon/Generated/IconController12';
-import IconController16 from '@foundation/Icon/Generated/IconController16';
-import IconController20 from '@foundation/Icon/Generated/IconController20';
-import IconController24 from '@foundation/Icon/Generated/IconController24';
-import IconController32 from '@foundation/Icon/Generated/IconController32';
-import IconCountry from '@foundation/Icon/Generated/IconCountry';
-import IconCountry12 from '@foundation/Icon/Generated/IconCountry12';
-import IconCountry16 from '@foundation/Icon/Generated/IconCountry16';
-import IconCountry20 from '@foundation/Icon/Generated/IconCountry20';
-import IconCountry24 from '@foundation/Icon/Generated/IconCountry24';
-import IconCountry32 from '@foundation/Icon/Generated/IconCountry32';
-import IconCrop from '@foundation/Icon/Generated/IconCrop';
-import IconCrop12 from '@foundation/Icon/Generated/IconCrop12';
-import IconCrop16 from '@foundation/Icon/Generated/IconCrop16';
-import IconCrop20 from '@foundation/Icon/Generated/IconCrop20';
-import IconCrop24 from '@foundation/Icon/Generated/IconCrop24';
-import IconCrop32 from '@foundation/Icon/Generated/IconCrop32';
-import IconCross from '@foundation/Icon/Generated/IconCross';
-import IconCross12 from '@foundation/Icon/Generated/IconCross12';
-import IconCross16 from '@foundation/Icon/Generated/IconCross16';
-import IconCross20 from '@foundation/Icon/Generated/IconCross20';
-import IconCross24 from '@foundation/Icon/Generated/IconCross24';
-import IconCross32 from '@foundation/Icon/Generated/IconCross32';
-import IconCrossCircle from '@foundation/Icon/Generated/IconCrossCircle';
-import IconCrossCircle12 from '@foundation/Icon/Generated/IconCrossCircle12';
-import IconCrossCircle12Filled from '@foundation/Icon/Generated/IconCrossCircle12Filled';
-import IconCrossCircle16 from '@foundation/Icon/Generated/IconCrossCircle16';
-import IconCrossCircle16Filled from '@foundation/Icon/Generated/IconCrossCircle16Filled';
-import IconCrossCircle20 from '@foundation/Icon/Generated/IconCrossCircle20';
-import IconCrossCircle20Filled from '@foundation/Icon/Generated/IconCrossCircle20Filled';
-import IconCrossCircle24 from '@foundation/Icon/Generated/IconCrossCircle24';
-import IconCrossCircle24Filled from '@foundation/Icon/Generated/IconCrossCircle24Filled';
-import IconCrossCircle32 from '@foundation/Icon/Generated/IconCrossCircle32';
-import IconCrossCircle32Filled from '@foundation/Icon/Generated/IconCrossCircle32Filled';
-import IconCrossCircleBox from '@foundation/Icon/Generated/IconCrossCircleBox';
-import IconCrossCircleBox12 from '@foundation/Icon/Generated/IconCrossCircleBox12';
-import IconCrossCircleBox16 from '@foundation/Icon/Generated/IconCrossCircleBox16';
-import IconCrossCircleBox20 from '@foundation/Icon/Generated/IconCrossCircleBox20';
-import IconCrossCircleBox24 from '@foundation/Icon/Generated/IconCrossCircleBox24';
-import IconCrossCircleBox32 from '@foundation/Icon/Generated/IconCrossCircleBox32';
-import IconCursorClick from '@foundation/Icon/Generated/IconCursorClick';
-import IconCursorClick12 from '@foundation/Icon/Generated/IconCursorClick12';
-import IconCursorClick16 from '@foundation/Icon/Generated/IconCursorClick16';
-import IconCursorClick20 from '@foundation/Icon/Generated/IconCursorClick20';
-import IconCursorClick24 from '@foundation/Icon/Generated/IconCursorClick24';
-import IconCursorClick32 from '@foundation/Icon/Generated/IconCursorClick32';
-import IconDiagramTree from '@foundation/Icon/Generated/IconDiagramTree';
-import IconDiagramTree12 from '@foundation/Icon/Generated/IconDiagramTree12';
-import IconDiagramTree16 from '@foundation/Icon/Generated/IconDiagramTree16';
-import IconDiagramTree20 from '@foundation/Icon/Generated/IconDiagramTree20';
-import IconDiagramTree24 from '@foundation/Icon/Generated/IconDiagramTree24';
-import IconDiagramTree32 from '@foundation/Icon/Generated/IconDiagramTree32';
-import IconDividerBlank from '@foundation/Icon/Generated/IconDividerBlank';
-import IconDividerBlank12 from '@foundation/Icon/Generated/IconDividerBlank12';
-import IconDividerBlank16 from '@foundation/Icon/Generated/IconDividerBlank16';
-import IconDividerBlank20 from '@foundation/Icon/Generated/IconDividerBlank20';
-import IconDividerBlank24 from '@foundation/Icon/Generated/IconDividerBlank24';
-import IconDividerBlank32 from '@foundation/Icon/Generated/IconDividerBlank32';
-import IconDividerDashed from '@foundation/Icon/Generated/IconDividerDashed';
-import IconDividerDashed12 from '@foundation/Icon/Generated/IconDividerDashed12';
-import IconDividerDashed16 from '@foundation/Icon/Generated/IconDividerDashed16';
-import IconDividerDashed20 from '@foundation/Icon/Generated/IconDividerDashed20';
-import IconDividerDashed24 from '@foundation/Icon/Generated/IconDividerDashed24';
-import IconDividerDashed32 from '@foundation/Icon/Generated/IconDividerDashed32';
-import IconDividerDotted from '@foundation/Icon/Generated/IconDividerDotted';
-import IconDividerDotted12 from '@foundation/Icon/Generated/IconDividerDotted12';
-import IconDividerDotted16 from '@foundation/Icon/Generated/IconDividerDotted16';
-import IconDividerDotted20 from '@foundation/Icon/Generated/IconDividerDotted20';
-import IconDividerDotted24 from '@foundation/Icon/Generated/IconDividerDotted24';
-import IconDividerDotted32 from '@foundation/Icon/Generated/IconDividerDotted32';
-import IconDividerSolid from '@foundation/Icon/Generated/IconDividerSolid';
-import IconDividerSolid12 from '@foundation/Icon/Generated/IconDividerSolid12';
-import IconDividerSolid16 from '@foundation/Icon/Generated/IconDividerSolid16';
-import IconDividerSolid20 from '@foundation/Icon/Generated/IconDividerSolid20';
-import IconDividerSolid24 from '@foundation/Icon/Generated/IconDividerSolid24';
-import IconDividerSolid32 from '@foundation/Icon/Generated/IconDividerSolid32';
-import IconDividerSolidInset from '@foundation/Icon/Generated/IconDividerSolidInset';
-import IconDividerSolidInset12 from '@foundation/Icon/Generated/IconDividerSolidInset12';
-import IconDividerSolidInset16 from '@foundation/Icon/Generated/IconDividerSolidInset16';
-import IconDividerSolidInset20 from '@foundation/Icon/Generated/IconDividerSolidInset20';
-import IconDividerSolidInset24 from '@foundation/Icon/Generated/IconDividerSolidInset24';
-import IconDividerSolidInset32 from '@foundation/Icon/Generated/IconDividerSolidInset32';
-import IconDoAndDont from '@foundation/Icon/Generated/IconDoAndDont';
-import IconDoAndDont12 from '@foundation/Icon/Generated/IconDoAndDont12';
-import IconDoAndDont16 from '@foundation/Icon/Generated/IconDoAndDont16';
-import IconDoAndDont20 from '@foundation/Icon/Generated/IconDoAndDont20';
-import IconDoAndDont24 from '@foundation/Icon/Generated/IconDoAndDont24';
-import IconDoAndDont32 from '@foundation/Icon/Generated/IconDoAndDont32';
-import IconDoAndDontBox from '@foundation/Icon/Generated/IconDoAndDontBox';
-import IconDoAndDontBox12 from '@foundation/Icon/Generated/IconDoAndDontBox12';
-import IconDoAndDontBox16 from '@foundation/Icon/Generated/IconDoAndDontBox16';
-import IconDoAndDontBox20 from '@foundation/Icon/Generated/IconDoAndDontBox20';
-import IconDoAndDontBox24 from '@foundation/Icon/Generated/IconDoAndDontBox24';
-import IconDoAndDontBox32 from '@foundation/Icon/Generated/IconDoAndDontBox32';
-import IconDocument from '@foundation/Icon/Generated/IconDocument';
-import IconDocument12 from '@foundation/Icon/Generated/IconDocument12';
-import IconDocument16 from '@foundation/Icon/Generated/IconDocument16';
-import IconDocument20 from '@foundation/Icon/Generated/IconDocument20';
-import IconDocument24 from '@foundation/Icon/Generated/IconDocument24';
-import IconDocument32 from '@foundation/Icon/Generated/IconDocument32';
-import IconDocumentArrowRight from '@foundation/Icon/Generated/IconDocumentArrowRight';
-import IconDocumentArrowRight12 from '@foundation/Icon/Generated/IconDocumentArrowRight12';
-import IconDocumentArrowRight16 from '@foundation/Icon/Generated/IconDocumentArrowRight16';
-import IconDocumentArrowRight20 from '@foundation/Icon/Generated/IconDocumentArrowRight20';
-import IconDocumentArrowRight24 from '@foundation/Icon/Generated/IconDocumentArrowRight24';
-import IconDocumentArrowRight32 from '@foundation/Icon/Generated/IconDocumentArrowRight32';
-import IconDocumentBadge from '@foundation/Icon/Generated/IconDocumentBadge';
-import IconDocumentBadge12 from '@foundation/Icon/Generated/IconDocumentBadge12';
-import IconDocumentBadge16 from '@foundation/Icon/Generated/IconDocumentBadge16';
-import IconDocumentBadge20 from '@foundation/Icon/Generated/IconDocumentBadge20';
-import IconDocumentBadge24 from '@foundation/Icon/Generated/IconDocumentBadge24';
-import IconDocumentBadge32 from '@foundation/Icon/Generated/IconDocumentBadge32';
-import IconDocumentCheck from '@foundation/Icon/Generated/IconDocumentCheck';
-import IconDocumentCheck12 from '@foundation/Icon/Generated/IconDocumentCheck12';
-import IconDocumentCheck16 from '@foundation/Icon/Generated/IconDocumentCheck16';
-import IconDocumentCheck20 from '@foundation/Icon/Generated/IconDocumentCheck20';
-import IconDocumentCheck24 from '@foundation/Icon/Generated/IconDocumentCheck24';
-import IconDocumentCheck32 from '@foundation/Icon/Generated/IconDocumentCheck32';
-import IconDocumentCode from '@foundation/Icon/Generated/IconDocumentCode';
-import IconDocumentCode12 from '@foundation/Icon/Generated/IconDocumentCode12';
-import IconDocumentCode16 from '@foundation/Icon/Generated/IconDocumentCode16';
-import IconDocumentCode20 from '@foundation/Icon/Generated/IconDocumentCode20';
-import IconDocumentCode24 from '@foundation/Icon/Generated/IconDocumentCode24';
-import IconDocumentCode32 from '@foundation/Icon/Generated/IconDocumentCode32';
-import IconDocumentList from '@foundation/Icon/Generated/IconDocumentList';
-import IconDocumentList12 from '@foundation/Icon/Generated/IconDocumentList12';
-import IconDocumentList16 from '@foundation/Icon/Generated/IconDocumentList16';
-import IconDocumentList20 from '@foundation/Icon/Generated/IconDocumentList20';
-import IconDocumentList24 from '@foundation/Icon/Generated/IconDocumentList24';
-import IconDocumentList32 from '@foundation/Icon/Generated/IconDocumentList32';
-import IconDocumentStack from '@foundation/Icon/Generated/IconDocumentStack';
-import IconDocumentStack12 from '@foundation/Icon/Generated/IconDocumentStack12';
-import IconDocumentStack16 from '@foundation/Icon/Generated/IconDocumentStack16';
-import IconDocumentStack20 from '@foundation/Icon/Generated/IconDocumentStack20';
-import IconDocumentStack24 from '@foundation/Icon/Generated/IconDocumentStack24';
-import IconDocumentStack32 from '@foundation/Icon/Generated/IconDocumentStack32';
-import IconDocumentText from '@foundation/Icon/Generated/IconDocumentText';
-import IconDocumentText12 from '@foundation/Icon/Generated/IconDocumentText12';
-import IconDocumentText16 from '@foundation/Icon/Generated/IconDocumentText16';
-import IconDocumentText20 from '@foundation/Icon/Generated/IconDocumentText20';
-import IconDocumentText24 from '@foundation/Icon/Generated/IconDocumentText24';
-import IconDocumentText32 from '@foundation/Icon/Generated/IconDocumentText32';
-import IconDotsHorizontal from '@foundation/Icon/Generated/IconDotsHorizontal';
-import IconDotsHorizontal12 from '@foundation/Icon/Generated/IconDotsHorizontal12';
-import IconDotsHorizontal16 from '@foundation/Icon/Generated/IconDotsHorizontal16';
-import IconDotsHorizontal20 from '@foundation/Icon/Generated/IconDotsHorizontal20';
-import IconDotsHorizontal24 from '@foundation/Icon/Generated/IconDotsHorizontal24';
-import IconDotsHorizontal32 from '@foundation/Icon/Generated/IconDotsHorizontal32';
-import IconDotsVertical from '@foundation/Icon/Generated/IconDotsVertical';
-import IconDotsVertical12 from '@foundation/Icon/Generated/IconDotsVertical12';
-import IconDotsVertical16 from '@foundation/Icon/Generated/IconDotsVertical16';
-import IconDotsVertical20 from '@foundation/Icon/Generated/IconDotsVertical20';
-import IconDotsVertical24 from '@foundation/Icon/Generated/IconDotsVertical24';
-import IconDotsVertical32 from '@foundation/Icon/Generated/IconDotsVertical32';
-import IconDroplet from '@foundation/Icon/Generated/IconDroplet';
-import IconDroplet12 from '@foundation/Icon/Generated/IconDroplet12';
-import IconDroplet16 from '@foundation/Icon/Generated/IconDroplet16';
-import IconDroplet20 from '@foundation/Icon/Generated/IconDroplet20';
-import IconDroplet24 from '@foundation/Icon/Generated/IconDroplet24';
-import IconDroplet32 from '@foundation/Icon/Generated/IconDroplet32';
-import IconDrops from '@foundation/Icon/Generated/IconDrops';
-import IconDrops12 from '@foundation/Icon/Generated/IconDrops12';
-import IconDrops16 from '@foundation/Icon/Generated/IconDrops16';
-import IconDrops20 from '@foundation/Icon/Generated/IconDrops20';
-import IconDrops24 from '@foundation/Icon/Generated/IconDrops24';
-import IconDrops32 from '@foundation/Icon/Generated/IconDrops32';
-import IconEnvelope from '@foundation/Icon/Generated/IconEnvelope';
-import IconEnvelope12 from '@foundation/Icon/Generated/IconEnvelope12';
-import IconEnvelope16 from '@foundation/Icon/Generated/IconEnvelope16';
-import IconEnvelope20 from '@foundation/Icon/Generated/IconEnvelope20';
-import IconEnvelope24 from '@foundation/Icon/Generated/IconEnvelope24';
-import IconEnvelope32 from '@foundation/Icon/Generated/IconEnvelope32';
-import IconEraser from '@foundation/Icon/Generated/IconEraser';
-import IconEraser12 from '@foundation/Icon/Generated/IconEraser12';
-import IconEraser16 from '@foundation/Icon/Generated/IconEraser16';
-import IconEraser20 from '@foundation/Icon/Generated/IconEraser20';
-import IconEraser24 from '@foundation/Icon/Generated/IconEraser24';
-import IconEraser32 from '@foundation/Icon/Generated/IconEraser32';
-import IconExclamationMark from '@foundation/Icon/Generated/IconExclamationMark';
-import IconExclamationMark12 from '@foundation/Icon/Generated/IconExclamationMark12';
-import IconExclamationMark16 from '@foundation/Icon/Generated/IconExclamationMark16';
-import IconExclamationMark20 from '@foundation/Icon/Generated/IconExclamationMark20';
-import IconExclamationMark24 from '@foundation/Icon/Generated/IconExclamationMark24';
-import IconExclamationMark32 from '@foundation/Icon/Generated/IconExclamationMark32';
-import IconExclamationMarkCircle from '@foundation/Icon/Generated/IconExclamationMarkCircle';
-import IconExclamationMarkCircle12 from '@foundation/Icon/Generated/IconExclamationMarkCircle12';
-import IconExclamationMarkCircle12Filled from '@foundation/Icon/Generated/IconExclamationMarkCircle12Filled';
-import IconExclamationMarkCircle16 from '@foundation/Icon/Generated/IconExclamationMarkCircle16';
-import IconExclamationMarkCircle16Filled from '@foundation/Icon/Generated/IconExclamationMarkCircle16Filled';
-import IconExclamationMarkCircle20 from '@foundation/Icon/Generated/IconExclamationMarkCircle20';
-import IconExclamationMarkCircle20Filled from '@foundation/Icon/Generated/IconExclamationMarkCircle20Filled';
-import IconExclamationMarkCircle24 from '@foundation/Icon/Generated/IconExclamationMarkCircle24';
-import IconExclamationMarkCircle24Filled from '@foundation/Icon/Generated/IconExclamationMarkCircle24Filled';
-import IconExclamationMarkCircle32 from '@foundation/Icon/Generated/IconExclamationMarkCircle32';
-import IconExclamationMarkCircle32Filled from '@foundation/Icon/Generated/IconExclamationMarkCircle32Filled';
-import IconExclamationMarkTriangle from '@foundation/Icon/Generated/IconExclamationMarkTriangle';
-import IconExclamationMarkTriangle12 from '@foundation/Icon/Generated/IconExclamationMarkTriangle12';
-import IconExclamationMarkTriangle16 from '@foundation/Icon/Generated/IconExclamationMarkTriangle16';
-import IconExclamationMarkTriangle20 from '@foundation/Icon/Generated/IconExclamationMarkTriangle20';
-import IconExclamationMarkTriangle24 from '@foundation/Icon/Generated/IconExclamationMarkTriangle24';
-import IconExclamationMarkTriangle32 from '@foundation/Icon/Generated/IconExclamationMarkTriangle32';
-import IconEye from '@foundation/Icon/Generated/IconEye';
-import IconEye12 from '@foundation/Icon/Generated/IconEye12';
-import IconEye16 from '@foundation/Icon/Generated/IconEye16';
-import IconEye20 from '@foundation/Icon/Generated/IconEye20';
-import IconEye24 from '@foundation/Icon/Generated/IconEye24';
-import IconEye32 from '@foundation/Icon/Generated/IconEye32';
-import IconEyeOff from '@foundation/Icon/Generated/IconEyeOff';
-import IconEyeOff12 from '@foundation/Icon/Generated/IconEyeOff12';
-import IconEyeOff16 from '@foundation/Icon/Generated/IconEyeOff16';
-import IconEyeOff20 from '@foundation/Icon/Generated/IconEyeOff20';
-import IconEyeOff24 from '@foundation/Icon/Generated/IconEyeOff24';
-import IconEyeOff32 from '@foundation/Icon/Generated/IconEyeOff32';
-import IconFaceExtraHappy from '@foundation/Icon/Generated/IconFaceExtraHappy';
-import IconFaceExtraHappy12 from '@foundation/Icon/Generated/IconFaceExtraHappy12';
-import IconFaceExtraHappy16 from '@foundation/Icon/Generated/IconFaceExtraHappy16';
-import IconFaceExtraHappy20 from '@foundation/Icon/Generated/IconFaceExtraHappy20';
-import IconFaceExtraHappy24 from '@foundation/Icon/Generated/IconFaceExtraHappy24';
-import IconFaceExtraHappy32 from '@foundation/Icon/Generated/IconFaceExtraHappy32';
-import IconFaceHappy from '@foundation/Icon/Generated/IconFaceHappy';
-import IconFaceHappy12 from '@foundation/Icon/Generated/IconFaceHappy12';
-import IconFaceHappy12Filled from '@foundation/Icon/Generated/IconFaceHappy12Filled';
-import IconFaceHappy16 from '@foundation/Icon/Generated/IconFaceHappy16';
-import IconFaceHappy16Filled from '@foundation/Icon/Generated/IconFaceHappy16Filled';
-import IconFaceHappy20 from '@foundation/Icon/Generated/IconFaceHappy20';
-import IconFaceHappy20Filled from '@foundation/Icon/Generated/IconFaceHappy20Filled';
-import IconFaceHappy24 from '@foundation/Icon/Generated/IconFaceHappy24';
-import IconFaceHappy24Filled from '@foundation/Icon/Generated/IconFaceHappy24Filled';
-import IconFaceHappy32 from '@foundation/Icon/Generated/IconFaceHappy32';
-import IconFaceHappy32Filled from '@foundation/Icon/Generated/IconFaceHappy32Filled';
-import IconFaceNeutral from '@foundation/Icon/Generated/IconFaceNeutral';
-import IconFaceNeutral12 from '@foundation/Icon/Generated/IconFaceNeutral12';
-import IconFaceNeutral12Filled from '@foundation/Icon/Generated/IconFaceNeutral12Filled';
-import IconFaceNeutral16 from '@foundation/Icon/Generated/IconFaceNeutral16';
-import IconFaceNeutral16Filled from '@foundation/Icon/Generated/IconFaceNeutral16Filled';
-import IconFaceNeutral20 from '@foundation/Icon/Generated/IconFaceNeutral20';
-import IconFaceNeutral20Filled from '@foundation/Icon/Generated/IconFaceNeutral20Filled';
-import IconFaceNeutral24 from '@foundation/Icon/Generated/IconFaceNeutral24';
-import IconFaceNeutral24Filled from '@foundation/Icon/Generated/IconFaceNeutral24Filled';
-import IconFaceNeutral32 from '@foundation/Icon/Generated/IconFaceNeutral32';
-import IconFaceNeutral32Filled from '@foundation/Icon/Generated/IconFaceNeutral32Filled';
-import IconFaceSad from '@foundation/Icon/Generated/IconFaceSad';
-import IconFaceSad12 from '@foundation/Icon/Generated/IconFaceSad12';
-import IconFaceSad12Filled from '@foundation/Icon/Generated/IconFaceSad12Filled';
-import IconFaceSad16 from '@foundation/Icon/Generated/IconFaceSad16';
-import IconFaceSad16Filled from '@foundation/Icon/Generated/IconFaceSad16Filled';
-import IconFaceSad20 from '@foundation/Icon/Generated/IconFaceSad20';
-import IconFaceSad20Filled from '@foundation/Icon/Generated/IconFaceSad20Filled';
-import IconFaceSad24 from '@foundation/Icon/Generated/IconFaceSad24';
-import IconFaceSad24Filled from '@foundation/Icon/Generated/IconFaceSad24Filled';
-import IconFaceSad32 from '@foundation/Icon/Generated/IconFaceSad32';
-import IconFaceSad32Filled from '@foundation/Icon/Generated/IconFaceSad32Filled';
-import IconFixedHeight from '@foundation/Icon/Generated/IconFixedHeight';
-import IconFixedHeight12 from '@foundation/Icon/Generated/IconFixedHeight12';
-import IconFixedHeight16 from '@foundation/Icon/Generated/IconFixedHeight16';
-import IconFixedHeight20 from '@foundation/Icon/Generated/IconFixedHeight20';
-import IconFixedHeight24 from '@foundation/Icon/Generated/IconFixedHeight24';
-import IconFixedHeight32 from '@foundation/Icon/Generated/IconFixedHeight32';
-import IconFixedWidth from '@foundation/Icon/Generated/IconFixedWidth';
-import IconFixedWidth12 from '@foundation/Icon/Generated/IconFixedWidth12';
-import IconFixedWidth16 from '@foundation/Icon/Generated/IconFixedWidth16';
-import IconFixedWidth20 from '@foundation/Icon/Generated/IconFixedWidth20';
-import IconFixedWidth24 from '@foundation/Icon/Generated/IconFixedWidth24';
-import IconFixedWidth32 from '@foundation/Icon/Generated/IconFixedWidth32';
-import IconFlag from '@foundation/Icon/Generated/IconFlag';
-import IconFlag12 from '@foundation/Icon/Generated/IconFlag12';
-import IconFlag16 from '@foundation/Icon/Generated/IconFlag16';
-import IconFlag20 from '@foundation/Icon/Generated/IconFlag20';
-import IconFlag24 from '@foundation/Icon/Generated/IconFlag24';
-import IconFlag32 from '@foundation/Icon/Generated/IconFlag32';
-import IconFocalPoint from '@foundation/Icon/Generated/IconFocalPoint';
-import IconFocalPoint12 from '@foundation/Icon/Generated/IconFocalPoint12';
-import IconFocalPoint16 from '@foundation/Icon/Generated/IconFocalPoint16';
-import IconFocalPoint20 from '@foundation/Icon/Generated/IconFocalPoint20';
-import IconFocalPoint24 from '@foundation/Icon/Generated/IconFocalPoint24';
-import IconFocalPoint32 from '@foundation/Icon/Generated/IconFocalPoint32';
-import IconFolder from '@foundation/Icon/Generated/IconFolder';
-import IconFolder12 from '@foundation/Icon/Generated/IconFolder12';
-import IconFolder16 from '@foundation/Icon/Generated/IconFolder16';
-import IconFolder20 from '@foundation/Icon/Generated/IconFolder20';
-import IconFolder24 from '@foundation/Icon/Generated/IconFolder24';
-import IconFolder32 from '@foundation/Icon/Generated/IconFolder32';
-import IconFolderArrowUp from '@foundation/Icon/Generated/IconFolderArrowUp';
-import IconFolderArrowUp12 from '@foundation/Icon/Generated/IconFolderArrowUp12';
-import IconFolderArrowUp16 from '@foundation/Icon/Generated/IconFolderArrowUp16';
-import IconFolderArrowUp20 from '@foundation/Icon/Generated/IconFolderArrowUp20';
-import IconFolderArrowUp24 from '@foundation/Icon/Generated/IconFolderArrowUp24';
-import IconFolderArrowUp32 from '@foundation/Icon/Generated/IconFolderArrowUp32';
-import IconFooter from '@foundation/Icon/Generated/IconFooter';
-import IconFooter12 from '@foundation/Icon/Generated/IconFooter12';
-import IconFooter16 from '@foundation/Icon/Generated/IconFooter16';
-import IconFooter20 from '@foundation/Icon/Generated/IconFooter20';
-import IconFooter24 from '@foundation/Icon/Generated/IconFooter24';
-import IconFooter32 from '@foundation/Icon/Generated/IconFooter32';
-import IconFunnel from '@foundation/Icon/Generated/IconFunnel';
-import IconFunnel12 from '@foundation/Icon/Generated/IconFunnel12';
-import IconFunnel16 from '@foundation/Icon/Generated/IconFunnel16';
-import IconFunnel20 from '@foundation/Icon/Generated/IconFunnel20';
-import IconFunnel24 from '@foundation/Icon/Generated/IconFunnel24';
-import IconFunnel32 from '@foundation/Icon/Generated/IconFunnel32';
-import IconGrabHandle from '@foundation/Icon/Generated/IconGrabHandle';
-import IconGrabHandle12 from '@foundation/Icon/Generated/IconGrabHandle12';
-import IconGrabHandle16 from '@foundation/Icon/Generated/IconGrabHandle16';
-import IconGrabHandle20 from '@foundation/Icon/Generated/IconGrabHandle20';
-import IconGrabHandle24 from '@foundation/Icon/Generated/IconGrabHandle24';
-import IconGrabHandle32 from '@foundation/Icon/Generated/IconGrabHandle32';
-import IconGradient from '@foundation/Icon/Generated/IconGradient';
-import IconGradient12 from '@foundation/Icon/Generated/IconGradient12';
-import IconGradient16 from '@foundation/Icon/Generated/IconGradient16';
-import IconGradient20 from '@foundation/Icon/Generated/IconGradient20';
-import IconGradient24 from '@foundation/Icon/Generated/IconGradient24';
-import IconGradient32 from '@foundation/Icon/Generated/IconGradient32';
-import IconGridRegular from '@foundation/Icon/Generated/IconGridRegular';
-import IconGridRegular12 from '@foundation/Icon/Generated/IconGridRegular12';
-import IconGridRegular12Filled from '@foundation/Icon/Generated/IconGridRegular12Filled';
-import IconGridRegular16 from '@foundation/Icon/Generated/IconGridRegular16';
-import IconGridRegular16Filled from '@foundation/Icon/Generated/IconGridRegular16Filled';
-import IconGridRegular20 from '@foundation/Icon/Generated/IconGridRegular20';
-import IconGridRegular20Filled from '@foundation/Icon/Generated/IconGridRegular20Filled';
-import IconGridRegular24 from '@foundation/Icon/Generated/IconGridRegular24';
-import IconGridRegular24Filled from '@foundation/Icon/Generated/IconGridRegular24Filled';
-import IconGridRegular32 from '@foundation/Icon/Generated/IconGridRegular32';
-import IconGridRegular32Filled from '@foundation/Icon/Generated/IconGridRegular32Filled';
-import IconH from '@foundation/Icon/Generated/IconH';
-import IconH12 from '@foundation/Icon/Generated/IconH12';
-import IconH16 from '@foundation/Icon/Generated/IconH16';
-import IconH20 from '@foundation/Icon/Generated/IconH20';
-import IconH24 from '@foundation/Icon/Generated/IconH24';
-import IconH32 from '@foundation/Icon/Generated/IconH32';
-import IconHeading from '@foundation/Icon/Generated/IconHeading';
-import IconHeading12 from '@foundation/Icon/Generated/IconHeading12';
-import IconHeading16 from '@foundation/Icon/Generated/IconHeading16';
-import IconHeading20 from '@foundation/Icon/Generated/IconHeading20';
-import IconHeading24 from '@foundation/Icon/Generated/IconHeading24';
-import IconHeading32 from '@foundation/Icon/Generated/IconHeading32';
-import IconHeartCircle from '@foundation/Icon/Generated/IconHeartCircle';
-import IconHeartCircle12 from '@foundation/Icon/Generated/IconHeartCircle12';
-import IconHeartCircle16 from '@foundation/Icon/Generated/IconHeartCircle16';
-import IconHeartCircle20 from '@foundation/Icon/Generated/IconHeartCircle20';
-import IconHeartCircle24 from '@foundation/Icon/Generated/IconHeartCircle24';
-import IconHeartCircle32 from '@foundation/Icon/Generated/IconHeartCircle32';
-import IconHeartCircleStack from '@foundation/Icon/Generated/IconHeartCircleStack';
-import IconHeartCircleStack12 from '@foundation/Icon/Generated/IconHeartCircleStack12';
-import IconHeartCircleStack16 from '@foundation/Icon/Generated/IconHeartCircleStack16';
-import IconHeartCircleStack20 from '@foundation/Icon/Generated/IconHeartCircleStack20';
-import IconHeartCircleStack24 from '@foundation/Icon/Generated/IconHeartCircleStack24';
-import IconHeartCircleStack32 from '@foundation/Icon/Generated/IconHeartCircleStack32';
-import IconHighlighter from '@foundation/Icon/Generated/IconHighlighter';
-import IconHighlighter12 from '@foundation/Icon/Generated/IconHighlighter12';
-import IconHighlighter16 from '@foundation/Icon/Generated/IconHighlighter16';
-import IconHighlighter20 from '@foundation/Icon/Generated/IconHighlighter20';
-import IconHighlighter24 from '@foundation/Icon/Generated/IconHighlighter24';
-import IconHighlighter32 from '@foundation/Icon/Generated/IconHighlighter32';
-import IconHome from '@foundation/Icon/Generated/IconHome';
-import IconHome12 from '@foundation/Icon/Generated/IconHome12';
-import IconHome16 from '@foundation/Icon/Generated/IconHome16';
-import IconHome20 from '@foundation/Icon/Generated/IconHome20';
-import IconHome24 from '@foundation/Icon/Generated/IconHome24';
-import IconHome32 from '@foundation/Icon/Generated/IconHome32';
-import IconIcon from '@foundation/Icon/Generated/IconIcon';
-import IconIcon12 from '@foundation/Icon/Generated/IconIcon12';
-import IconIcon16 from '@foundation/Icon/Generated/IconIcon16';
-import IconIcon20 from '@foundation/Icon/Generated/IconIcon20';
-import IconIcon24 from '@foundation/Icon/Generated/IconIcon24';
-import IconIcon32 from '@foundation/Icon/Generated/IconIcon32';
-import IconImage from '@foundation/Icon/Generated/IconImage';
-import IconImage12 from '@foundation/Icon/Generated/IconImage12';
-import IconImage16 from '@foundation/Icon/Generated/IconImage16';
-import IconImage20 from '@foundation/Icon/Generated/IconImage20';
-import IconImage24 from '@foundation/Icon/Generated/IconImage24';
-import IconImage32 from '@foundation/Icon/Generated/IconImage32';
-import IconImageStack from '@foundation/Icon/Generated/IconImageStack';
-import IconImageStack12 from '@foundation/Icon/Generated/IconImageStack12';
-import IconImageStack16 from '@foundation/Icon/Generated/IconImageStack16';
-import IconImageStack20 from '@foundation/Icon/Generated/IconImageStack20';
-import IconImageStack24 from '@foundation/Icon/Generated/IconImageStack24';
-import IconImageStack32 from '@foundation/Icon/Generated/IconImageStack32';
-import IconImageStackPlus from '@foundation/Icon/Generated/IconImageStackPlus';
-import IconImageStackPlus12 from '@foundation/Icon/Generated/IconImageStackPlus12';
-import IconImageStackPlus16 from '@foundation/Icon/Generated/IconImageStackPlus16';
-import IconImageStackPlus20 from '@foundation/Icon/Generated/IconImageStackPlus20';
-import IconImageStackPlus24 from '@foundation/Icon/Generated/IconImageStackPlus24';
-import IconImageStackPlus32 from '@foundation/Icon/Generated/IconImageStackPlus32';
-import IconImageWithText from '@foundation/Icon/Generated/IconImageWithText';
-import IconImageWithText12 from '@foundation/Icon/Generated/IconImageWithText12';
-import IconImageWithText16 from '@foundation/Icon/Generated/IconImageWithText16';
-import IconImageWithText20 from '@foundation/Icon/Generated/IconImageWithText20';
-import IconImageWithText24 from '@foundation/Icon/Generated/IconImageWithText24';
-import IconImageWithText32 from '@foundation/Icon/Generated/IconImageWithText32';
-import IconInfo from '@foundation/Icon/Generated/IconInfo';
-import IconInfo12 from '@foundation/Icon/Generated/IconInfo12';
-import IconInfo16 from '@foundation/Icon/Generated/IconInfo16';
-import IconInfo20 from '@foundation/Icon/Generated/IconInfo20';
-import IconInfo24 from '@foundation/Icon/Generated/IconInfo24';
-import IconInfo32 from '@foundation/Icon/Generated/IconInfo32';
-import IconIntercom from '@foundation/Icon/Generated/IconIntercom';
-import IconIntercom12 from '@foundation/Icon/Generated/IconIntercom12';
-import IconIntercom16 from '@foundation/Icon/Generated/IconIntercom16';
-import IconIntercom20 from '@foundation/Icon/Generated/IconIntercom20';
-import IconIntercom24 from '@foundation/Icon/Generated/IconIntercom24';
-import IconIntercom32 from '@foundation/Icon/Generated/IconIntercom32';
-import IconJohanna from '@foundation/Icon/Generated/IconJohanna';
-import IconJohanna12 from '@foundation/Icon/Generated/IconJohanna12';
-import IconJohanna16 from '@foundation/Icon/Generated/IconJohanna16';
-import IconJohanna20 from '@foundation/Icon/Generated/IconJohanna20';
-import IconJohanna24 from '@foundation/Icon/Generated/IconJohanna24';
-import IconJohanna32 from '@foundation/Icon/Generated/IconJohanna32';
-import IconLabel from '@foundation/Icon/Generated/IconLabel';
-import IconLabel12 from '@foundation/Icon/Generated/IconLabel12';
-import IconLabel16 from '@foundation/Icon/Generated/IconLabel16';
-import IconLabel20 from '@foundation/Icon/Generated/IconLabel20';
-import IconLabel24 from '@foundation/Icon/Generated/IconLabel24';
-import IconLabel32 from '@foundation/Icon/Generated/IconLabel32';
-import IconLanguage from '@foundation/Icon/Generated/IconLanguage';
-import IconLanguage12 from '@foundation/Icon/Generated/IconLanguage12';
-import IconLanguage16 from '@foundation/Icon/Generated/IconLanguage16';
-import IconLanguage20 from '@foundation/Icon/Generated/IconLanguage20';
-import IconLanguage24 from '@foundation/Icon/Generated/IconLanguage24';
-import IconLanguage32 from '@foundation/Icon/Generated/IconLanguage32';
-import IconLayers from '@foundation/Icon/Generated/IconLayers';
-import IconLayers12 from '@foundation/Icon/Generated/IconLayers12';
-import IconLayers16 from '@foundation/Icon/Generated/IconLayers16';
-import IconLayers20 from '@foundation/Icon/Generated/IconLayers20';
-import IconLayers24 from '@foundation/Icon/Generated/IconLayers24';
-import IconLayers32 from '@foundation/Icon/Generated/IconLayers32';
-import IconLayersSingle from '@foundation/Icon/Generated/IconLayersSingle';
-import IconLayersSingle12 from '@foundation/Icon/Generated/IconLayersSingle12';
-import IconLayersSingle16 from '@foundation/Icon/Generated/IconLayersSingle16';
-import IconLayersSingle20 from '@foundation/Icon/Generated/IconLayersSingle20';
-import IconLayersSingle24 from '@foundation/Icon/Generated/IconLayersSingle24';
-import IconLayersSingle32 from '@foundation/Icon/Generated/IconLayersSingle32';
-import IconLeaf from '@foundation/Icon/Generated/IconLeaf';
-import IconLeaf12 from '@foundation/Icon/Generated/IconLeaf12';
-import IconLeaf16 from '@foundation/Icon/Generated/IconLeaf16';
-import IconLeaf20 from '@foundation/Icon/Generated/IconLeaf20';
-import IconLeaf24 from '@foundation/Icon/Generated/IconLeaf24';
-import IconLeaf32 from '@foundation/Icon/Generated/IconLeaf32';
-import IconLightbulb from '@foundation/Icon/Generated/IconLightbulb';
-import IconLightbulb12 from '@foundation/Icon/Generated/IconLightbulb12';
-import IconLightbulb16 from '@foundation/Icon/Generated/IconLightbulb16';
-import IconLightbulb20 from '@foundation/Icon/Generated/IconLightbulb20';
-import IconLightbulb24 from '@foundation/Icon/Generated/IconLightbulb24';
-import IconLightbulb32 from '@foundation/Icon/Generated/IconLightbulb32';
-import IconLightning from '@foundation/Icon/Generated/IconLightning';
-import IconLightning12 from '@foundation/Icon/Generated/IconLightning12';
-import IconLightning12Filled from '@foundation/Icon/Generated/IconLightning12Filled';
-import IconLightning16 from '@foundation/Icon/Generated/IconLightning16';
-import IconLightning16Filled from '@foundation/Icon/Generated/IconLightning16Filled';
-import IconLightning20 from '@foundation/Icon/Generated/IconLightning20';
-import IconLightning20Filled from '@foundation/Icon/Generated/IconLightning20Filled';
-import IconLightning24 from '@foundation/Icon/Generated/IconLightning24';
-import IconLightning24Filled from '@foundation/Icon/Generated/IconLightning24Filled';
-import IconLightning32 from '@foundation/Icon/Generated/IconLightning32';
-import IconLightning32Filled from '@foundation/Icon/Generated/IconLightning32Filled';
-import IconLines from '@foundation/Icon/Generated/IconLines';
-import IconLines12 from '@foundation/Icon/Generated/IconLines12';
-import IconLines16 from '@foundation/Icon/Generated/IconLines16';
-import IconLines20 from '@foundation/Icon/Generated/IconLines20';
-import IconLines24 from '@foundation/Icon/Generated/IconLines24';
-import IconLines32 from '@foundation/Icon/Generated/IconLines32';
-import IconLinesTriangular from '@foundation/Icon/Generated/IconLinesTriangular';
-import IconLinesTriangular12 from '@foundation/Icon/Generated/IconLinesTriangular12';
-import IconLinesTriangular16 from '@foundation/Icon/Generated/IconLinesTriangular16';
-import IconLinesTriangular20 from '@foundation/Icon/Generated/IconLinesTriangular20';
-import IconLinesTriangular24 from '@foundation/Icon/Generated/IconLinesTriangular24';
-import IconLinesTriangular32 from '@foundation/Icon/Generated/IconLinesTriangular32';
-import IconLink from '@foundation/Icon/Generated/IconLink';
-import IconLink12 from '@foundation/Icon/Generated/IconLink12';
-import IconLink16 from '@foundation/Icon/Generated/IconLink16';
-import IconLink20 from '@foundation/Icon/Generated/IconLink20';
-import IconLink24 from '@foundation/Icon/Generated/IconLink24';
-import IconLink32 from '@foundation/Icon/Generated/IconLink32';
-import IconLinkBox from '@foundation/Icon/Generated/IconLinkBox';
-import IconLinkBox12 from '@foundation/Icon/Generated/IconLinkBox12';
-import IconLinkBox16 from '@foundation/Icon/Generated/IconLinkBox16';
-import IconLinkBox20 from '@foundation/Icon/Generated/IconLinkBox20';
-import IconLinkBox24 from '@foundation/Icon/Generated/IconLinkBox24';
-import IconLinkBox32 from '@foundation/Icon/Generated/IconLinkBox32';
-import IconLinkOpen from '@foundation/Icon/Generated/IconLinkOpen';
-import IconLinkOpen12 from '@foundation/Icon/Generated/IconLinkOpen12';
-import IconLinkOpen16 from '@foundation/Icon/Generated/IconLinkOpen16';
-import IconLinkOpen20 from '@foundation/Icon/Generated/IconLinkOpen20';
-import IconLinkOpen24 from '@foundation/Icon/Generated/IconLinkOpen24';
-import IconLinkOpen32 from '@foundation/Icon/Generated/IconLinkOpen32';
-import IconListBullet from '@foundation/Icon/Generated/IconListBullet';
-import IconListBullet12 from '@foundation/Icon/Generated/IconListBullet12';
-import IconListBullet16 from '@foundation/Icon/Generated/IconListBullet16';
-import IconListBullet20 from '@foundation/Icon/Generated/IconListBullet20';
-import IconListBullet24 from '@foundation/Icon/Generated/IconListBullet24';
-import IconListBullet32 from '@foundation/Icon/Generated/IconListBullet32';
-import IconListCheck from '@foundation/Icon/Generated/IconListCheck';
-import IconListCheck12 from '@foundation/Icon/Generated/IconListCheck12';
-import IconListCheck16 from '@foundation/Icon/Generated/IconListCheck16';
-import IconListCheck20 from '@foundation/Icon/Generated/IconListCheck20';
-import IconListCheck24 from '@foundation/Icon/Generated/IconListCheck24';
-import IconListCheck32 from '@foundation/Icon/Generated/IconListCheck32';
-import IconListIndented from '@foundation/Icon/Generated/IconListIndented';
-import IconListIndented12 from '@foundation/Icon/Generated/IconListIndented12';
-import IconListIndented16 from '@foundation/Icon/Generated/IconListIndented16';
-import IconListIndented20 from '@foundation/Icon/Generated/IconListIndented20';
-import IconListIndented24 from '@foundation/Icon/Generated/IconListIndented24';
-import IconListIndented32 from '@foundation/Icon/Generated/IconListIndented32';
-import IconListNumbers from '@foundation/Icon/Generated/IconListNumbers';
-import IconListNumbers12 from '@foundation/Icon/Generated/IconListNumbers12';
-import IconListNumbers16 from '@foundation/Icon/Generated/IconListNumbers16';
-import IconListNumbers20 from '@foundation/Icon/Generated/IconListNumbers20';
-import IconListNumbers24 from '@foundation/Icon/Generated/IconListNumbers24';
-import IconListNumbers32 from '@foundation/Icon/Generated/IconListNumbers32';
-import IconLockClosed from '@foundation/Icon/Generated/IconLockClosed';
-import IconLockClosed12 from '@foundation/Icon/Generated/IconLockClosed12';
-import IconLockClosed16 from '@foundation/Icon/Generated/IconLockClosed16';
-import IconLockClosed20 from '@foundation/Icon/Generated/IconLockClosed20';
-import IconLockClosed24 from '@foundation/Icon/Generated/IconLockClosed24';
-import IconLockClosed32 from '@foundation/Icon/Generated/IconLockClosed32';
-import IconLockOpen from '@foundation/Icon/Generated/IconLockOpen';
-import IconLockOpen12 from '@foundation/Icon/Generated/IconLockOpen12';
-import IconLockOpen12Filled from '@foundation/Icon/Generated/IconLockOpen12Filled';
-import IconLockOpen16 from '@foundation/Icon/Generated/IconLockOpen16';
-import IconLockOpen16Filled from '@foundation/Icon/Generated/IconLockOpen16Filled';
-import IconLockOpen20 from '@foundation/Icon/Generated/IconLockOpen20';
-import IconLockOpen20Filled from '@foundation/Icon/Generated/IconLockOpen20Filled';
-import IconLockOpen24 from '@foundation/Icon/Generated/IconLockOpen24';
-import IconLockOpen24Filled from '@foundation/Icon/Generated/IconLockOpen24Filled';
-import IconLockOpen32 from '@foundation/Icon/Generated/IconLockOpen32';
-import IconLockOpen32Filled from '@foundation/Icon/Generated/IconLockOpen32Filled';
-import IconLollipop from '@foundation/Icon/Generated/IconLollipop';
-import IconLollipop12 from '@foundation/Icon/Generated/IconLollipop12';
-import IconLollipop16 from '@foundation/Icon/Generated/IconLollipop16';
-import IconLollipop20 from '@foundation/Icon/Generated/IconLollipop20';
-import IconLollipop24 from '@foundation/Icon/Generated/IconLollipop24';
-import IconLollipop32 from '@foundation/Icon/Generated/IconLollipop32';
-import IconMagnifier from '@foundation/Icon/Generated/IconMagnifier';
-import IconMagnifier12 from '@foundation/Icon/Generated/IconMagnifier12';
-import IconMagnifier16 from '@foundation/Icon/Generated/IconMagnifier16';
-import IconMagnifier20 from '@foundation/Icon/Generated/IconMagnifier20';
-import IconMagnifier24 from '@foundation/Icon/Generated/IconMagnifier24';
-import IconMagnifier32 from '@foundation/Icon/Generated/IconMagnifier32';
-import IconMagnifierPlus from '@foundation/Icon/Generated/IconMagnifierPlus';
-import IconMagnifierPlus12 from '@foundation/Icon/Generated/IconMagnifierPlus12';
-import IconMagnifierPlus16 from '@foundation/Icon/Generated/IconMagnifierPlus16';
-import IconMagnifierPlus20 from '@foundation/Icon/Generated/IconMagnifierPlus20';
-import IconMagnifierPlus24 from '@foundation/Icon/Generated/IconMagnifierPlus24';
-import IconMagnifierPlus32 from '@foundation/Icon/Generated/IconMagnifierPlus32';
-import IconMapPointer from '@foundation/Icon/Generated/IconMapPointer';
-import IconMapPointer12 from '@foundation/Icon/Generated/IconMapPointer12';
-import IconMapPointer16 from '@foundation/Icon/Generated/IconMapPointer16';
-import IconMapPointer20 from '@foundation/Icon/Generated/IconMapPointer20';
-import IconMapPointer24 from '@foundation/Icon/Generated/IconMapPointer24';
-import IconMapPointer32 from '@foundation/Icon/Generated/IconMapPointer32';
-import IconMarkArea from '@foundation/Icon/Generated/IconMarkArea';
-import IconMarkArea12 from '@foundation/Icon/Generated/IconMarkArea12';
-import IconMarkArea16 from '@foundation/Icon/Generated/IconMarkArea16';
-import IconMarkArea20 from '@foundation/Icon/Generated/IconMarkArea20';
-import IconMarkArea24 from '@foundation/Icon/Generated/IconMarkArea24';
-import IconMarkArea32 from '@foundation/Icon/Generated/IconMarkArea32';
-import IconMasonry from '@foundation/Icon/Generated/IconMasonry';
-import IconMasonry12 from '@foundation/Icon/Generated/IconMasonry12';
-import IconMasonry16 from '@foundation/Icon/Generated/IconMasonry16';
-import IconMasonry20 from '@foundation/Icon/Generated/IconMasonry20';
-import IconMasonry24 from '@foundation/Icon/Generated/IconMasonry24';
-import IconMasonry32 from '@foundation/Icon/Generated/IconMasonry32';
-import IconMediaObject from '@foundation/Icon/Generated/IconMediaObject';
-import IconMediaObject12 from '@foundation/Icon/Generated/IconMediaObject12';
-import IconMediaObject16 from '@foundation/Icon/Generated/IconMediaObject16';
-import IconMediaObject20 from '@foundation/Icon/Generated/IconMediaObject20';
-import IconMediaObject24 from '@foundation/Icon/Generated/IconMediaObject24';
-import IconMediaObject32 from '@foundation/Icon/Generated/IconMediaObject32';
-import IconMediaObjectIcon from '@foundation/Icon/Generated/IconMediaObjectIcon';
-import IconMediaObjectIcon12 from '@foundation/Icon/Generated/IconMediaObjectIcon12';
-import IconMediaObjectIcon16 from '@foundation/Icon/Generated/IconMediaObjectIcon16';
-import IconMediaObjectIcon20 from '@foundation/Icon/Generated/IconMediaObjectIcon20';
-import IconMediaObjectIcon24 from '@foundation/Icon/Generated/IconMediaObjectIcon24';
-import IconMediaObjectIcon32 from '@foundation/Icon/Generated/IconMediaObjectIcon32';
-import IconMediaObjectImage from '@foundation/Icon/Generated/IconMediaObjectImage';
-import IconMediaObjectImage12 from '@foundation/Icon/Generated/IconMediaObjectImage12';
-import IconMediaObjectImage16 from '@foundation/Icon/Generated/IconMediaObjectImage16';
-import IconMediaObjectImage20 from '@foundation/Icon/Generated/IconMediaObjectImage20';
-import IconMediaObjectImage24 from '@foundation/Icon/Generated/IconMediaObjectImage24';
-import IconMediaObjectImage32 from '@foundation/Icon/Generated/IconMediaObjectImage32';
-import IconMediaObjectRatio1To1 from '@foundation/Icon/Generated/IconMediaObjectRatio1To1';
-import IconMediaObjectRatio1To112 from '@foundation/Icon/Generated/IconMediaObjectRatio1To112';
-import IconMediaObjectRatio1To116 from '@foundation/Icon/Generated/IconMediaObjectRatio1To116';
-import IconMediaObjectRatio1To120 from '@foundation/Icon/Generated/IconMediaObjectRatio1To120';
-import IconMediaObjectRatio1To124 from '@foundation/Icon/Generated/IconMediaObjectRatio1To124';
-import IconMediaObjectRatio1To132 from '@foundation/Icon/Generated/IconMediaObjectRatio1To132';
-import IconMediaObjectRatio1To2 from '@foundation/Icon/Generated/IconMediaObjectRatio1To2';
-import IconMediaObjectRatio1To212 from '@foundation/Icon/Generated/IconMediaObjectRatio1To212';
-import IconMediaObjectRatio1To216 from '@foundation/Icon/Generated/IconMediaObjectRatio1To216';
-import IconMediaObjectRatio1To220 from '@foundation/Icon/Generated/IconMediaObjectRatio1To220';
-import IconMediaObjectRatio1To224 from '@foundation/Icon/Generated/IconMediaObjectRatio1To224';
-import IconMediaObjectRatio1To232 from '@foundation/Icon/Generated/IconMediaObjectRatio1To232';
-import IconMediaObjectRatio1To3 from '@foundation/Icon/Generated/IconMediaObjectRatio1To3';
-import IconMediaObjectRatio1To312 from '@foundation/Icon/Generated/IconMediaObjectRatio1To312';
-import IconMediaObjectRatio1To316 from '@foundation/Icon/Generated/IconMediaObjectRatio1To316';
-import IconMediaObjectRatio1To320 from '@foundation/Icon/Generated/IconMediaObjectRatio1To320';
-import IconMediaObjectRatio1To324 from '@foundation/Icon/Generated/IconMediaObjectRatio1To324';
-import IconMediaObjectRatio1To332 from '@foundation/Icon/Generated/IconMediaObjectRatio1To332';
-import IconMediaObjectRatio2To1 from '@foundation/Icon/Generated/IconMediaObjectRatio2To1';
-import IconMediaObjectRatio2To112 from '@foundation/Icon/Generated/IconMediaObjectRatio2To112';
-import IconMediaObjectRatio2To116 from '@foundation/Icon/Generated/IconMediaObjectRatio2To116';
-import IconMediaObjectRatio2To120 from '@foundation/Icon/Generated/IconMediaObjectRatio2To120';
-import IconMediaObjectRatio2To124 from '@foundation/Icon/Generated/IconMediaObjectRatio2To124';
-import IconMediaObjectRatio2To132 from '@foundation/Icon/Generated/IconMediaObjectRatio2To132';
-import IconMediaObjectRatio3To1 from '@foundation/Icon/Generated/IconMediaObjectRatio3To1';
-import IconMediaObjectRatio3To112 from '@foundation/Icon/Generated/IconMediaObjectRatio3To112';
-import IconMediaObjectRatio3To116 from '@foundation/Icon/Generated/IconMediaObjectRatio3To116';
-import IconMediaObjectRatio3To120 from '@foundation/Icon/Generated/IconMediaObjectRatio3To120';
-import IconMediaObjectRatio3To124 from '@foundation/Icon/Generated/IconMediaObjectRatio3To124';
-import IconMediaObjectRatio3To132 from '@foundation/Icon/Generated/IconMediaObjectRatio3To132';
-import IconMediaObjectTextBottom from '@foundation/Icon/Generated/IconMediaObjectTextBottom';
-import IconMediaObjectTextBottom12 from '@foundation/Icon/Generated/IconMediaObjectTextBottom12';
-import IconMediaObjectTextBottom16 from '@foundation/Icon/Generated/IconMediaObjectTextBottom16';
-import IconMediaObjectTextBottom20 from '@foundation/Icon/Generated/IconMediaObjectTextBottom20';
-import IconMediaObjectTextBottom24 from '@foundation/Icon/Generated/IconMediaObjectTextBottom24';
-import IconMediaObjectTextBottom32 from '@foundation/Icon/Generated/IconMediaObjectTextBottom32';
-import IconMediaObjectTextLeft from '@foundation/Icon/Generated/IconMediaObjectTextLeft';
-import IconMediaObjectTextLeft12 from '@foundation/Icon/Generated/IconMediaObjectTextLeft12';
-import IconMediaObjectTextLeft16 from '@foundation/Icon/Generated/IconMediaObjectTextLeft16';
-import IconMediaObjectTextLeft20 from '@foundation/Icon/Generated/IconMediaObjectTextLeft20';
-import IconMediaObjectTextLeft24 from '@foundation/Icon/Generated/IconMediaObjectTextLeft24';
-import IconMediaObjectTextLeft32 from '@foundation/Icon/Generated/IconMediaObjectTextLeft32';
-import IconMediaObjectTextRight from '@foundation/Icon/Generated/IconMediaObjectTextRight';
-import IconMediaObjectTextRight12 from '@foundation/Icon/Generated/IconMediaObjectTextRight12';
-import IconMediaObjectTextRight16 from '@foundation/Icon/Generated/IconMediaObjectTextRight16';
-import IconMediaObjectTextRight20 from '@foundation/Icon/Generated/IconMediaObjectTextRight20';
-import IconMediaObjectTextRight24 from '@foundation/Icon/Generated/IconMediaObjectTextRight24';
-import IconMediaObjectTextRight32 from '@foundation/Icon/Generated/IconMediaObjectTextRight32';
-import IconMediaObjectTextTop from '@foundation/Icon/Generated/IconMediaObjectTextTop';
-import IconMediaObjectTextTop12 from '@foundation/Icon/Generated/IconMediaObjectTextTop12';
-import IconMediaObjectTextTop16 from '@foundation/Icon/Generated/IconMediaObjectTextTop16';
-import IconMediaObjectTextTop20 from '@foundation/Icon/Generated/IconMediaObjectTextTop20';
-import IconMediaObjectTextTop24 from '@foundation/Icon/Generated/IconMediaObjectTextTop24';
-import IconMediaObjectTextTop32 from '@foundation/Icon/Generated/IconMediaObjectTextTop32';
-import IconMediaObjectVideo from '@foundation/Icon/Generated/IconMediaObjectVideo';
-import IconMediaObjectVideo12 from '@foundation/Icon/Generated/IconMediaObjectVideo12';
-import IconMediaObjectVideo16 from '@foundation/Icon/Generated/IconMediaObjectVideo16';
-import IconMediaObjectVideo20 from '@foundation/Icon/Generated/IconMediaObjectVideo20';
-import IconMediaObjectVideo24 from '@foundation/Icon/Generated/IconMediaObjectVideo24';
-import IconMediaObjectVideo32 from '@foundation/Icon/Generated/IconMediaObjectVideo32';
-import IconMegaMenu from '@foundation/Icon/Generated/IconMegaMenu';
-import IconMegaMenu12 from '@foundation/Icon/Generated/IconMegaMenu12';
-import IconMegaMenu16 from '@foundation/Icon/Generated/IconMegaMenu16';
-import IconMegaMenu20 from '@foundation/Icon/Generated/IconMegaMenu20';
-import IconMegaMenu24 from '@foundation/Icon/Generated/IconMegaMenu24';
-import IconMegaMenu32 from '@foundation/Icon/Generated/IconMegaMenu32';
-import IconMegaphone from '@foundation/Icon/Generated/IconMegaphone';
-import IconMegaphone12 from '@foundation/Icon/Generated/IconMegaphone12';
-import IconMegaphone16 from '@foundation/Icon/Generated/IconMegaphone16';
-import IconMegaphone20 from '@foundation/Icon/Generated/IconMegaphone20';
-import IconMegaphone24 from '@foundation/Icon/Generated/IconMegaphone24';
-import IconMegaphone32 from '@foundation/Icon/Generated/IconMegaphone32';
-import IconMinus from '@foundation/Icon/Generated/IconMinus';
-import IconMinus12 from '@foundation/Icon/Generated/IconMinus12';
-import IconMinus16 from '@foundation/Icon/Generated/IconMinus16';
-import IconMinus20 from '@foundation/Icon/Generated/IconMinus20';
-import IconMinus24 from '@foundation/Icon/Generated/IconMinus24';
-import IconMinus32 from '@foundation/Icon/Generated/IconMinus32';
-import IconMinusCircle from '@foundation/Icon/Generated/IconMinusCircle';
-import IconMinusCircle12 from '@foundation/Icon/Generated/IconMinusCircle12';
-import IconMinusCircle16 from '@foundation/Icon/Generated/IconMinusCircle16';
-import IconMinusCircle20 from '@foundation/Icon/Generated/IconMinusCircle20';
-import IconMinusCircle24 from '@foundation/Icon/Generated/IconMinusCircle24';
-import IconMinusCircle32 from '@foundation/Icon/Generated/IconMinusCircle32';
-import IconMoney from '@foundation/Icon/Generated/IconMoney';
-import IconMoney12 from '@foundation/Icon/Generated/IconMoney12';
-import IconMoney16 from '@foundation/Icon/Generated/IconMoney16';
-import IconMoney20 from '@foundation/Icon/Generated/IconMoney20';
-import IconMoney24 from '@foundation/Icon/Generated/IconMoney24';
-import IconMoney32 from '@foundation/Icon/Generated/IconMoney32';
-import IconMoneyCircle from '@foundation/Icon/Generated/IconMoneyCircle';
-import IconMoneyCircle12Filled from '@foundation/Icon/Generated/IconMoneyCircle12Filled';
-import IconMoneyCircle16Filled from '@foundation/Icon/Generated/IconMoneyCircle16Filled';
-import IconMoneyCircle20Filled from '@foundation/Icon/Generated/IconMoneyCircle20Filled';
-import IconMoneyCircle24Filled from '@foundation/Icon/Generated/IconMoneyCircle24Filled';
-import IconMoneyCircle32Filled from '@foundation/Icon/Generated/IconMoneyCircle32Filled';
-import IconMusicNote from '@foundation/Icon/Generated/IconMusicNote';
-import IconMusicNote12 from '@foundation/Icon/Generated/IconMusicNote12';
-import IconMusicNote16 from '@foundation/Icon/Generated/IconMusicNote16';
-import IconMusicNote20 from '@foundation/Icon/Generated/IconMusicNote20';
-import IconMusicNote24 from '@foundation/Icon/Generated/IconMusicNote24';
-import IconMusicNote32 from '@foundation/Icon/Generated/IconMusicNote32';
-import IconNook from '@foundation/Icon/Generated/IconNook';
-import IconNook12 from '@foundation/Icon/Generated/IconNook12';
-import IconNook16 from '@foundation/Icon/Generated/IconNook16';
-import IconNook20 from '@foundation/Icon/Generated/IconNook20';
-import IconNook24 from '@foundation/Icon/Generated/IconNook24';
-import IconNook32 from '@foundation/Icon/Generated/IconNook32';
-import IconOrientationBoxes from '@foundation/Icon/Generated/IconOrientationBoxes';
-import IconOrientationBoxes12 from '@foundation/Icon/Generated/IconOrientationBoxes12';
-import IconOrientationBoxes16 from '@foundation/Icon/Generated/IconOrientationBoxes16';
-import IconOrientationBoxes20 from '@foundation/Icon/Generated/IconOrientationBoxes20';
-import IconOrientationBoxes24 from '@foundation/Icon/Generated/IconOrientationBoxes24';
-import IconOrientationBoxes32 from '@foundation/Icon/Generated/IconOrientationBoxes32';
-import IconPaintbrush from '@foundation/Icon/Generated/IconPaintbrush';
-import IconPaintbrush12 from '@foundation/Icon/Generated/IconPaintbrush12';
-import IconPaintbrush16 from '@foundation/Icon/Generated/IconPaintbrush16';
-import IconPaintbrush20 from '@foundation/Icon/Generated/IconPaintbrush20';
-import IconPaintbrush24 from '@foundation/Icon/Generated/IconPaintbrush24';
-import IconPaintbrush32 from '@foundation/Icon/Generated/IconPaintbrush32';
-import IconPaperclip from '@foundation/Icon/Generated/IconPaperclip';
-import IconPaperclip12 from '@foundation/Icon/Generated/IconPaperclip12';
-import IconPaperclip16 from '@foundation/Icon/Generated/IconPaperclip16';
-import IconPaperclip20 from '@foundation/Icon/Generated/IconPaperclip20';
-import IconPaperclip24 from '@foundation/Icon/Generated/IconPaperclip24';
-import IconPaperclip32 from '@foundation/Icon/Generated/IconPaperclip32';
-import IconPaperplane from '@foundation/Icon/Generated/IconPaperplane';
-import IconPaperplane12 from '@foundation/Icon/Generated/IconPaperplane12';
-import IconPaperplane16 from '@foundation/Icon/Generated/IconPaperplane16';
-import IconPaperplane20 from '@foundation/Icon/Generated/IconPaperplane20';
-import IconPaperplane24 from '@foundation/Icon/Generated/IconPaperplane24';
-import IconPaperplane32 from '@foundation/Icon/Generated/IconPaperplane32';
-import IconPause from '@foundation/Icon/Generated/IconPause';
-import IconPause12 from '@foundation/Icon/Generated/IconPause12';
-import IconPause12Filled from '@foundation/Icon/Generated/IconPause12Filled';
-import IconPause16 from '@foundation/Icon/Generated/IconPause16';
-import IconPause16Filled from '@foundation/Icon/Generated/IconPause16Filled';
-import IconPause20 from '@foundation/Icon/Generated/IconPause20';
-import IconPause20Filled from '@foundation/Icon/Generated/IconPause20Filled';
-import IconPause24 from '@foundation/Icon/Generated/IconPause24';
-import IconPause24Filled from '@foundation/Icon/Generated/IconPause24Filled';
-import IconPause32 from '@foundation/Icon/Generated/IconPause32';
-import IconPause32Filled from '@foundation/Icon/Generated/IconPause32Filled';
-import IconPen from '@foundation/Icon/Generated/IconPen';
-import IconPen12 from '@foundation/Icon/Generated/IconPen12';
-import IconPen16 from '@foundation/Icon/Generated/IconPen16';
-import IconPen20 from '@foundation/Icon/Generated/IconPen20';
-import IconPen24 from '@foundation/Icon/Generated/IconPen24';
-import IconPen32 from '@foundation/Icon/Generated/IconPen32';
-import IconPeople from '@foundation/Icon/Generated/IconPeople';
-import IconPeople12 from '@foundation/Icon/Generated/IconPeople12';
-import IconPeople16 from '@foundation/Icon/Generated/IconPeople16';
-import IconPeople20 from '@foundation/Icon/Generated/IconPeople20';
-import IconPeople24 from '@foundation/Icon/Generated/IconPeople24';
-import IconPeople32 from '@foundation/Icon/Generated/IconPeople32';
-import IconPerson from '@foundation/Icon/Generated/IconPerson';
-import IconPerson12 from '@foundation/Icon/Generated/IconPerson12';
-import IconPerson16 from '@foundation/Icon/Generated/IconPerson16';
-import IconPerson20 from '@foundation/Icon/Generated/IconPerson20';
-import IconPerson24 from '@foundation/Icon/Generated/IconPerson24';
-import IconPerson32 from '@foundation/Icon/Generated/IconPerson32';
-import IconPersonAnonymous from '@foundation/Icon/Generated/IconPersonAnonymous';
-import IconPersonAnonymous12 from '@foundation/Icon/Generated/IconPersonAnonymous12';
-import IconPersonAnonymous16 from '@foundation/Icon/Generated/IconPersonAnonymous16';
-import IconPersonAnonymous20 from '@foundation/Icon/Generated/IconPersonAnonymous20';
-import IconPersonAnonymous24 from '@foundation/Icon/Generated/IconPersonAnonymous24';
-import IconPersonAnonymous32 from '@foundation/Icon/Generated/IconPersonAnonymous32';
-import IconPersonSlash from '@foundation/Icon/Generated/IconPersonSlash';
-import IconPersonSlash12 from '@foundation/Icon/Generated/IconPersonSlash12';
-import IconPersonSlash16 from '@foundation/Icon/Generated/IconPersonSlash16';
-import IconPersonSlash20 from '@foundation/Icon/Generated/IconPersonSlash20';
-import IconPersonSlash24 from '@foundation/Icon/Generated/IconPersonSlash24';
-import IconPersonSlash32 from '@foundation/Icon/Generated/IconPersonSlash32';
-import IconPills from '@foundation/Icon/Generated/IconPills';
-import IconPills12 from '@foundation/Icon/Generated/IconPills12';
-import IconPills16 from '@foundation/Icon/Generated/IconPills16';
-import IconPills20 from '@foundation/Icon/Generated/IconPills20';
-import IconPills24 from '@foundation/Icon/Generated/IconPills24';
-import IconPills32 from '@foundation/Icon/Generated/IconPills32';
-import IconPin from '@foundation/Icon/Generated/IconPin';
-import IconPin12 from '@foundation/Icon/Generated/IconPin12';
-import IconPin12Filled from '@foundation/Icon/Generated/IconPin12Filled';
-import IconPin16 from '@foundation/Icon/Generated/IconPin16';
-import IconPin16Filled from '@foundation/Icon/Generated/IconPin16Filled';
-import IconPin20 from '@foundation/Icon/Generated/IconPin20';
-import IconPin20Filled from '@foundation/Icon/Generated/IconPin20Filled';
-import IconPin24 from '@foundation/Icon/Generated/IconPin24';
-import IconPin24Filled from '@foundation/Icon/Generated/IconPin24Filled';
-import IconPin32 from '@foundation/Icon/Generated/IconPin32';
-import IconPin32Filled from '@foundation/Icon/Generated/IconPin32Filled';
-import IconPlay from '@foundation/Icon/Generated/IconPlay';
-import IconPlay12 from '@foundation/Icon/Generated/IconPlay12';
-import IconPlay12Filled from '@foundation/Icon/Generated/IconPlay12Filled';
-import IconPlay16 from '@foundation/Icon/Generated/IconPlay16';
-import IconPlay16Filled from '@foundation/Icon/Generated/IconPlay16Filled';
-import IconPlay20 from '@foundation/Icon/Generated/IconPlay20';
-import IconPlay20Filled from '@foundation/Icon/Generated/IconPlay20Filled';
-import IconPlay24 from '@foundation/Icon/Generated/IconPlay24';
-import IconPlay24Filled from '@foundation/Icon/Generated/IconPlay24Filled';
-import IconPlay32 from '@foundation/Icon/Generated/IconPlay32';
-import IconPlay32Filled from '@foundation/Icon/Generated/IconPlay32Filled';
-import IconPlayBox from '@foundation/Icon/Generated/IconPlayBox';
-import IconPlayBox12 from '@foundation/Icon/Generated/IconPlayBox12';
-import IconPlayBox16 from '@foundation/Icon/Generated/IconPlayBox16';
-import IconPlayBox20 from '@foundation/Icon/Generated/IconPlayBox20';
-import IconPlayBox24 from '@foundation/Icon/Generated/IconPlayBox24';
-import IconPlayBox32 from '@foundation/Icon/Generated/IconPlayBox32';
-import IconPlayCircle from '@foundation/Icon/Generated/IconPlayCircle';
-import IconPlayCircle12 from '@foundation/Icon/Generated/IconPlayCircle12';
-import IconPlayCircle16 from '@foundation/Icon/Generated/IconPlayCircle16';
-import IconPlayCircle20 from '@foundation/Icon/Generated/IconPlayCircle20';
-import IconPlayCircle24 from '@foundation/Icon/Generated/IconPlayCircle24';
-import IconPlayCircle32 from '@foundation/Icon/Generated/IconPlayCircle32';
-import IconPlayFrame from '@foundation/Icon/Generated/IconPlayFrame';
-import IconPlayFrame12 from '@foundation/Icon/Generated/IconPlayFrame12';
-import IconPlayFrame16 from '@foundation/Icon/Generated/IconPlayFrame16';
-import IconPlayFrame20 from '@foundation/Icon/Generated/IconPlayFrame20';
-import IconPlayFrame24 from '@foundation/Icon/Generated/IconPlayFrame24';
-import IconPlayFrame32 from '@foundation/Icon/Generated/IconPlayFrame32';
-import IconPlus from '@foundation/Icon/Generated/IconPlus';
-import IconPlus12 from '@foundation/Icon/Generated/IconPlus12';
-import IconPlus16 from '@foundation/Icon/Generated/IconPlus16';
-import IconPlus20 from '@foundation/Icon/Generated/IconPlus20';
-import IconPlus24 from '@foundation/Icon/Generated/IconPlus24';
-import IconPlus32 from '@foundation/Icon/Generated/IconPlus32';
-import IconPlusBoxStack from '@foundation/Icon/Generated/IconPlusBoxStack';
-import IconPlusBoxStack12 from '@foundation/Icon/Generated/IconPlusBoxStack12';
-import IconPlusBoxStack16 from '@foundation/Icon/Generated/IconPlusBoxStack16';
-import IconPlusBoxStack20 from '@foundation/Icon/Generated/IconPlusBoxStack20';
-import IconPlusBoxStack24 from '@foundation/Icon/Generated/IconPlusBoxStack24';
-import IconPlusBoxStack32 from '@foundation/Icon/Generated/IconPlusBoxStack32';
-import IconPlusCircle from '@foundation/Icon/Generated/IconPlusCircle';
-import IconPlusCircle12 from '@foundation/Icon/Generated/IconPlusCircle12';
-import IconPlusCircle16 from '@foundation/Icon/Generated/IconPlusCircle16';
-import IconPlusCircle20 from '@foundation/Icon/Generated/IconPlusCircle20';
-import IconPlusCircle24 from '@foundation/Icon/Generated/IconPlusCircle24';
-import IconPlusCircle32 from '@foundation/Icon/Generated/IconPlusCircle32';
-import IconPlusFrame from '@foundation/Icon/Generated/IconPlusFrame';
-import IconPlusFrame12 from '@foundation/Icon/Generated/IconPlusFrame12';
-import IconPlusFrame16 from '@foundation/Icon/Generated/IconPlusFrame16';
-import IconPlusFrame20 from '@foundation/Icon/Generated/IconPlusFrame20';
-import IconPlusFrame24 from '@foundation/Icon/Generated/IconPlusFrame24';
-import IconPlusFrame32 from '@foundation/Icon/Generated/IconPlusFrame32';
-import IconPointout from '@foundation/Icon/Generated/IconPointout';
-import IconPointout12 from '@foundation/Icon/Generated/IconPointout12';
-import IconPointout16 from '@foundation/Icon/Generated/IconPointout16';
-import IconPointout20 from '@foundation/Icon/Generated/IconPointout20';
-import IconPointout24 from '@foundation/Icon/Generated/IconPointout24';
-import IconPointout32 from '@foundation/Icon/Generated/IconPointout32';
-import IconPolygon from '@foundation/Icon/Generated/IconPolygon';
-import IconPolygon12 from '@foundation/Icon/Generated/IconPolygon12';
-import IconPolygon16 from '@foundation/Icon/Generated/IconPolygon16';
-import IconPolygon20 from '@foundation/Icon/Generated/IconPolygon20';
-import IconPolygon24 from '@foundation/Icon/Generated/IconPolygon24';
-import IconPolygon32 from '@foundation/Icon/Generated/IconPolygon32';
-import IconPuzzle from '@foundation/Icon/Generated/IconPuzzle';
-import IconPuzzle12 from '@foundation/Icon/Generated/IconPuzzle12';
-import IconPuzzle16 from '@foundation/Icon/Generated/IconPuzzle16';
-import IconPuzzle20 from '@foundation/Icon/Generated/IconPuzzle20';
-import IconPuzzle24 from '@foundation/Icon/Generated/IconPuzzle24';
-import IconPuzzle32 from '@foundation/Icon/Generated/IconPuzzle32';
-import IconQuestionMark from '@foundation/Icon/Generated/IconQuestionMark';
-import IconQuestionMark12 from '@foundation/Icon/Generated/IconQuestionMark12';
-import IconQuestionMark16 from '@foundation/Icon/Generated/IconQuestionMark16';
-import IconQuestionMark20 from '@foundation/Icon/Generated/IconQuestionMark20';
-import IconQuestionMark24 from '@foundation/Icon/Generated/IconQuestionMark24';
-import IconQuestionMark32 from '@foundation/Icon/Generated/IconQuestionMark32';
-import IconQuestionMarkCircle from '@foundation/Icon/Generated/IconQuestionMarkCircle';
-import IconQuestionMarkCircle12 from '@foundation/Icon/Generated/IconQuestionMarkCircle12';
-import IconQuestionMarkCircle16 from '@foundation/Icon/Generated/IconQuestionMarkCircle16';
-import IconQuestionMarkCircle20 from '@foundation/Icon/Generated/IconQuestionMarkCircle20';
-import IconQuestionMarkCircle24 from '@foundation/Icon/Generated/IconQuestionMarkCircle24';
-import IconQuestionMarkCircle32 from '@foundation/Icon/Generated/IconQuestionMarkCircle32';
-import IconRectangleLandscape from '@foundation/Icon/Generated/IconRectangleLandscape';
-import IconRectangleLandscape12 from '@foundation/Icon/Generated/IconRectangleLandscape12';
-import IconRectangleLandscape16 from '@foundation/Icon/Generated/IconRectangleLandscape16';
-import IconRectangleLandscape20 from '@foundation/Icon/Generated/IconRectangleLandscape20';
-import IconRectangleLandscape24 from '@foundation/Icon/Generated/IconRectangleLandscape24';
-import IconRectangleLandscape32 from '@foundation/Icon/Generated/IconRectangleLandscape32';
-import IconRectanglePortraitSquare from '@foundation/Icon/Generated/IconRectanglePortraitSquare';
-import IconRectanglePortraitSquare12 from '@foundation/Icon/Generated/IconRectanglePortraitSquare12';
-import IconRectanglePortraitSquare16 from '@foundation/Icon/Generated/IconRectanglePortraitSquare16';
-import IconRectanglePortraitSquare20 from '@foundation/Icon/Generated/IconRectanglePortraitSquare20';
-import IconRectanglePortraitSquare24 from '@foundation/Icon/Generated/IconRectanglePortraitSquare24';
-import IconRectanglePortraitSquare32 from '@foundation/Icon/Generated/IconRectanglePortraitSquare32';
-import IconReport from '@foundation/Icon/Generated/IconReport';
-import IconReport12 from '@foundation/Icon/Generated/IconReport12';
-import IconReport16 from '@foundation/Icon/Generated/IconReport16';
-import IconReport20 from '@foundation/Icon/Generated/IconReport20';
-import IconReport24 from '@foundation/Icon/Generated/IconReport24';
-import IconReport32 from '@foundation/Icon/Generated/IconReport32';
-import IconRocket from '@foundation/Icon/Generated/IconRocket';
-import IconRocket12 from '@foundation/Icon/Generated/IconRocket12';
-import IconRocket16 from '@foundation/Icon/Generated/IconRocket16';
-import IconRocket20 from '@foundation/Icon/Generated/IconRocket20';
-import IconRocket24 from '@foundation/Icon/Generated/IconRocket24';
-import IconRocket32 from '@foundation/Icon/Generated/IconRocket32';
-import IconRuler from '@foundation/Icon/Generated/IconRuler';
-import IconRuler12 from '@foundation/Icon/Generated/IconRuler12';
-import IconRuler16 from '@foundation/Icon/Generated/IconRuler16';
-import IconRuler20 from '@foundation/Icon/Generated/IconRuler20';
-import IconRuler24 from '@foundation/Icon/Generated/IconRuler24';
-import IconRuler32 from '@foundation/Icon/Generated/IconRuler32';
-import IconScaleLarge from '@foundation/Icon/Generated/IconScaleLarge';
-import IconScaleLarge12 from '@foundation/Icon/Generated/IconScaleLarge12';
-import IconScaleLarge16 from '@foundation/Icon/Generated/IconScaleLarge16';
-import IconScaleLarge20 from '@foundation/Icon/Generated/IconScaleLarge20';
-import IconScaleLarge24 from '@foundation/Icon/Generated/IconScaleLarge24';
-import IconScaleLarge32 from '@foundation/Icon/Generated/IconScaleLarge32';
-import IconScaleSmall from '@foundation/Icon/Generated/IconScaleSmall';
-import IconScaleSmall12 from '@foundation/Icon/Generated/IconScaleSmall12';
-import IconScaleSmall16 from '@foundation/Icon/Generated/IconScaleSmall16';
-import IconScaleSmall20 from '@foundation/Icon/Generated/IconScaleSmall20';
-import IconScaleSmall24 from '@foundation/Icon/Generated/IconScaleSmall24';
-import IconScaleSmall32 from '@foundation/Icon/Generated/IconScaleSmall32';
-import IconScissors from '@foundation/Icon/Generated/IconScissors';
-import IconScissors12 from '@foundation/Icon/Generated/IconScissors12';
-import IconScissors16 from '@foundation/Icon/Generated/IconScissors16';
-import IconScissors20 from '@foundation/Icon/Generated/IconScissors20';
-import IconScissors24 from '@foundation/Icon/Generated/IconScissors24';
-import IconScissors32 from '@foundation/Icon/Generated/IconScissors32';
-import IconSide from '@foundation/Icon/Generated/IconSide';
-import IconSide12 from '@foundation/Icon/Generated/IconSide12';
-import IconSide16 from '@foundation/Icon/Generated/IconSide16';
-import IconSide20 from '@foundation/Icon/Generated/IconSide20';
-import IconSide24 from '@foundation/Icon/Generated/IconSide24';
-import IconSide32 from '@foundation/Icon/Generated/IconSide32';
-import IconSidebarLeft from '@foundation/Icon/Generated/IconSidebarLeft';
-import IconSidebarLeft12 from '@foundation/Icon/Generated/IconSidebarLeft12';
-import IconSidebarLeft16 from '@foundation/Icon/Generated/IconSidebarLeft16';
-import IconSidebarLeft20 from '@foundation/Icon/Generated/IconSidebarLeft20';
-import IconSidebarLeft24 from '@foundation/Icon/Generated/IconSidebarLeft24';
-import IconSidebarLeft32 from '@foundation/Icon/Generated/IconSidebarLeft32';
-import IconSidebarLeftHide from '@foundation/Icon/Generated/IconSidebarLeftHide';
-import IconSidebarLeftHide12 from '@foundation/Icon/Generated/IconSidebarLeftHide12';
-import IconSidebarLeftHide12Filled from '@foundation/Icon/Generated/IconSidebarLeftHide12Filled';
-import IconSidebarLeftHide16 from '@foundation/Icon/Generated/IconSidebarLeftHide16';
-import IconSidebarLeftHide16Filled from '@foundation/Icon/Generated/IconSidebarLeftHide16Filled';
-import IconSidebarLeftHide20 from '@foundation/Icon/Generated/IconSidebarLeftHide20';
-import IconSidebarLeftHide20Filled from '@foundation/Icon/Generated/IconSidebarLeftHide20Filled';
-import IconSidebarLeftHide24 from '@foundation/Icon/Generated/IconSidebarLeftHide24';
-import IconSidebarLeftHide24Filled from '@foundation/Icon/Generated/IconSidebarLeftHide24Filled';
-import IconSidebarLeftHide32 from '@foundation/Icon/Generated/IconSidebarLeftHide32';
-import IconSidebarLeftHide32Filled from '@foundation/Icon/Generated/IconSidebarLeftHide32Filled';
-import IconSidebarLeftShow from '@foundation/Icon/Generated/IconSidebarLeftShow';
-import IconSidebarLeftShow12 from '@foundation/Icon/Generated/IconSidebarLeftShow12';
-import IconSidebarLeftShow12Filled from '@foundation/Icon/Generated/IconSidebarLeftShow12Filled';
-import IconSidebarLeftShow16 from '@foundation/Icon/Generated/IconSidebarLeftShow16';
-import IconSidebarLeftShow16Filled from '@foundation/Icon/Generated/IconSidebarLeftShow16Filled';
-import IconSidebarLeftShow20 from '@foundation/Icon/Generated/IconSidebarLeftShow20';
-import IconSidebarLeftShow20Filled from '@foundation/Icon/Generated/IconSidebarLeftShow20Filled';
-import IconSidebarLeftShow24 from '@foundation/Icon/Generated/IconSidebarLeftShow24';
-import IconSidebarLeftShow24Filled from '@foundation/Icon/Generated/IconSidebarLeftShow24Filled';
-import IconSidebarLeftShow32 from '@foundation/Icon/Generated/IconSidebarLeftShow32';
-import IconSidebarLeftShow32Filled from '@foundation/Icon/Generated/IconSidebarLeftShow32Filled';
-import IconSidebarRight from '@foundation/Icon/Generated/IconSidebarRight';
-import IconSidebarRight12 from '@foundation/Icon/Generated/IconSidebarRight12';
-import IconSidebarRight16 from '@foundation/Icon/Generated/IconSidebarRight16';
-import IconSidebarRight20 from '@foundation/Icon/Generated/IconSidebarRight20';
-import IconSidebarRight24 from '@foundation/Icon/Generated/IconSidebarRight24';
-import IconSidebarRight32 from '@foundation/Icon/Generated/IconSidebarRight32';
-import IconSidebarRightHide from '@foundation/Icon/Generated/IconSidebarRightHide';
-import IconSidebarRightHide12 from '@foundation/Icon/Generated/IconSidebarRightHide12';
-import IconSidebarRightHide12Filled from '@foundation/Icon/Generated/IconSidebarRightHide12Filled';
-import IconSidebarRightHide16 from '@foundation/Icon/Generated/IconSidebarRightHide16';
-import IconSidebarRightHide16Filled from '@foundation/Icon/Generated/IconSidebarRightHide16Filled';
-import IconSidebarRightHide20 from '@foundation/Icon/Generated/IconSidebarRightHide20';
-import IconSidebarRightHide20Filled from '@foundation/Icon/Generated/IconSidebarRightHide20Filled';
-import IconSidebarRightHide24 from '@foundation/Icon/Generated/IconSidebarRightHide24';
-import IconSidebarRightHide24Filled from '@foundation/Icon/Generated/IconSidebarRightHide24Filled';
-import IconSidebarRightHide32 from '@foundation/Icon/Generated/IconSidebarRightHide32';
-import IconSidebarRightHide32Filled from '@foundation/Icon/Generated/IconSidebarRightHide32Filled';
-import IconSidebarRightShow from '@foundation/Icon/Generated/IconSidebarRightShow';
-import IconSidebarRightShow12 from '@foundation/Icon/Generated/IconSidebarRightShow12';
-import IconSidebarRightShow12Filled from '@foundation/Icon/Generated/IconSidebarRightShow12Filled';
-import IconSidebarRightShow16 from '@foundation/Icon/Generated/IconSidebarRightShow16';
-import IconSidebarRightShow16Filled from '@foundation/Icon/Generated/IconSidebarRightShow16Filled';
-import IconSidebarRightShow20 from '@foundation/Icon/Generated/IconSidebarRightShow20';
-import IconSidebarRightShow20Filled from '@foundation/Icon/Generated/IconSidebarRightShow20Filled';
-import IconSidebarRightShow24 from '@foundation/Icon/Generated/IconSidebarRightShow24';
-import IconSidebarRightShow24Filled from '@foundation/Icon/Generated/IconSidebarRightShow24Filled';
-import IconSidebarRightShow32 from '@foundation/Icon/Generated/IconSidebarRightShow32';
-import IconSidebarRightShow32Filled from '@foundation/Icon/Generated/IconSidebarRightShow32Filled';
-import IconSketch from '@foundation/Icon/Generated/IconSketch';
-import IconSketch12 from '@foundation/Icon/Generated/IconSketch12';
-import IconSketch16 from '@foundation/Icon/Generated/IconSketch16';
-import IconSketch20 from '@foundation/Icon/Generated/IconSketch20';
-import IconSketch24 from '@foundation/Icon/Generated/IconSketch24';
-import IconSketch32 from '@foundation/Icon/Generated/IconSketch32';
-import IconSkip5SecondsBackward from '@foundation/Icon/Generated/IconSkip5SecondsBackward';
-import IconSkip5SecondsBackward12 from '@foundation/Icon/Generated/IconSkip5SecondsBackward12';
-import IconSkip5SecondsBackward16 from '@foundation/Icon/Generated/IconSkip5SecondsBackward16';
-import IconSkip5SecondsBackward20 from '@foundation/Icon/Generated/IconSkip5SecondsBackward20';
-import IconSkip5SecondsBackward24 from '@foundation/Icon/Generated/IconSkip5SecondsBackward24';
-import IconSkip5SecondsBackward32 from '@foundation/Icon/Generated/IconSkip5SecondsBackward32';
-import IconSkip5SecondsForward from '@foundation/Icon/Generated/IconSkip5SecondsForward';
-import IconSkip5SecondsForward12 from '@foundation/Icon/Generated/IconSkip5SecondsForward12';
-import IconSkip5SecondsForward16 from '@foundation/Icon/Generated/IconSkip5SecondsForward16';
-import IconSkip5SecondsForward20 from '@foundation/Icon/Generated/IconSkip5SecondsForward20';
-import IconSkip5SecondsForward24 from '@foundation/Icon/Generated/IconSkip5SecondsForward24';
-import IconSkip5SecondsForward32 from '@foundation/Icon/Generated/IconSkip5SecondsForward32';
-import IconSoundWave from '@foundation/Icon/Generated/IconSoundWave';
-import IconSoundWave12 from '@foundation/Icon/Generated/IconSoundWave12';
-import IconSoundWave16 from '@foundation/Icon/Generated/IconSoundWave16';
-import IconSoundWave20 from '@foundation/Icon/Generated/IconSoundWave20';
-import IconSoundWave24 from '@foundation/Icon/Generated/IconSoundWave24';
-import IconSoundWave32 from '@foundation/Icon/Generated/IconSoundWave32';
-import IconSpeaker from '@foundation/Icon/Generated/IconSpeaker';
-import IconSpeaker12 from '@foundation/Icon/Generated/IconSpeaker12';
-import IconSpeaker16 from '@foundation/Icon/Generated/IconSpeaker16';
-import IconSpeaker20 from '@foundation/Icon/Generated/IconSpeaker20';
-import IconSpeaker24 from '@foundation/Icon/Generated/IconSpeaker24';
-import IconSpeaker32 from '@foundation/Icon/Generated/IconSpeaker32';
-import IconSpeakerOff from '@foundation/Icon/Generated/IconSpeakerOff';
-import IconSpeakerOff12 from '@foundation/Icon/Generated/IconSpeakerOff12';
-import IconSpeakerOff16 from '@foundation/Icon/Generated/IconSpeakerOff16';
-import IconSpeakerOff20 from '@foundation/Icon/Generated/IconSpeakerOff20';
-import IconSpeakerOff24 from '@foundation/Icon/Generated/IconSpeakerOff24';
-import IconSpeakerOff32 from '@foundation/Icon/Generated/IconSpeakerOff32';
-import IconSpeechBubbleDots from '@foundation/Icon/Generated/IconSpeechBubbleDots';
-import IconSpeechBubbleDots12 from '@foundation/Icon/Generated/IconSpeechBubbleDots12';
-import IconSpeechBubbleDots16 from '@foundation/Icon/Generated/IconSpeechBubbleDots16';
-import IconSpeechBubbleDots20 from '@foundation/Icon/Generated/IconSpeechBubbleDots20';
-import IconSpeechBubbleDots24 from '@foundation/Icon/Generated/IconSpeechBubbleDots24';
-import IconSpeechBubbleDots32 from '@foundation/Icon/Generated/IconSpeechBubbleDots32';
-import IconSpeechBubbleEmpty from '@foundation/Icon/Generated/IconSpeechBubbleEmpty';
-import IconSpeechBubbleEmpty12 from '@foundation/Icon/Generated/IconSpeechBubbleEmpty12';
-import IconSpeechBubbleEmpty16 from '@foundation/Icon/Generated/IconSpeechBubbleEmpty16';
-import IconSpeechBubbleEmpty20 from '@foundation/Icon/Generated/IconSpeechBubbleEmpty20';
-import IconSpeechBubbleEmpty24 from '@foundation/Icon/Generated/IconSpeechBubbleEmpty24';
-import IconSpeechBubbleEmpty32 from '@foundation/Icon/Generated/IconSpeechBubbleEmpty32';
-import IconSpeechBubbleQuote from '@foundation/Icon/Generated/IconSpeechBubbleQuote';
-import IconSpeechBubbleQuote12 from '@foundation/Icon/Generated/IconSpeechBubbleQuote12';
-import IconSpeechBubbleQuote16 from '@foundation/Icon/Generated/IconSpeechBubbleQuote16';
-import IconSpeechBubbleQuote20 from '@foundation/Icon/Generated/IconSpeechBubbleQuote20';
-import IconSpeechBubbleQuote24 from '@foundation/Icon/Generated/IconSpeechBubbleQuote24';
-import IconSpeechBubbleQuote32 from '@foundation/Icon/Generated/IconSpeechBubbleQuote32';
-import IconSpeechBubbleStack from '@foundation/Icon/Generated/IconSpeechBubbleStack';
-import IconSpeechBubbleStack12 from '@foundation/Icon/Generated/IconSpeechBubbleStack12';
-import IconSpeechBubbleStack16 from '@foundation/Icon/Generated/IconSpeechBubbleStack16';
-import IconSpeechBubbleStack20 from '@foundation/Icon/Generated/IconSpeechBubbleStack20';
-import IconSpeechBubbleStack24 from '@foundation/Icon/Generated/IconSpeechBubbleStack24';
-import IconSpeechBubbleStack32 from '@foundation/Icon/Generated/IconSpeechBubbleStack32';
-import IconStackHorizontal from '@foundation/Icon/Generated/IconStackHorizontal';
-import IconStackHorizontal12 from '@foundation/Icon/Generated/IconStackHorizontal12';
-import IconStackHorizontal16 from '@foundation/Icon/Generated/IconStackHorizontal16';
-import IconStackHorizontal20 from '@foundation/Icon/Generated/IconStackHorizontal20';
-import IconStackHorizontal24 from '@foundation/Icon/Generated/IconStackHorizontal24';
-import IconStackHorizontal32 from '@foundation/Icon/Generated/IconStackHorizontal32';
-import IconStackHorizontalAltFour from '@foundation/Icon/Generated/IconStackHorizontalAltFour';
-import IconStackHorizontalAltFour12 from '@foundation/Icon/Generated/IconStackHorizontalAltFour12';
-import IconStackHorizontalAltFour16 from '@foundation/Icon/Generated/IconStackHorizontalAltFour16';
-import IconStackHorizontalAltFour20 from '@foundation/Icon/Generated/IconStackHorizontalAltFour20';
-import IconStackHorizontalAltFour24 from '@foundation/Icon/Generated/IconStackHorizontalAltFour24';
-import IconStackHorizontalAltFour32 from '@foundation/Icon/Generated/IconStackHorizontalAltFour32';
-import IconStackHorizontalAltThree from '@foundation/Icon/Generated/IconStackHorizontalAltThree';
-import IconStackHorizontalAltThree12 from '@foundation/Icon/Generated/IconStackHorizontalAltThree12';
-import IconStackHorizontalAltThree16 from '@foundation/Icon/Generated/IconStackHorizontalAltThree16';
-import IconStackHorizontalAltThree20 from '@foundation/Icon/Generated/IconStackHorizontalAltThree20';
-import IconStackHorizontalAltThree24 from '@foundation/Icon/Generated/IconStackHorizontalAltThree24';
-import IconStackHorizontalAltThree32 from '@foundation/Icon/Generated/IconStackHorizontalAltThree32';
-import IconStackHorizontalAltTwo from '@foundation/Icon/Generated/IconStackHorizontalAltTwo';
-import IconStackHorizontalAltTwo12 from '@foundation/Icon/Generated/IconStackHorizontalAltTwo12';
-import IconStackHorizontalAltTwo16 from '@foundation/Icon/Generated/IconStackHorizontalAltTwo16';
-import IconStackHorizontalAltTwo20 from '@foundation/Icon/Generated/IconStackHorizontalAltTwo20';
-import IconStackHorizontalAltTwo24 from '@foundation/Icon/Generated/IconStackHorizontalAltTwo24';
-import IconStackHorizontalAltTwo32 from '@foundation/Icon/Generated/IconStackHorizontalAltTwo32';
-import IconStackVertical from '@foundation/Icon/Generated/IconStackVertical';
-import IconStackVertical12 from '@foundation/Icon/Generated/IconStackVertical12';
-import IconStackVertical16 from '@foundation/Icon/Generated/IconStackVertical16';
-import IconStackVertical20 from '@foundation/Icon/Generated/IconStackVertical20';
-import IconStackVertical24 from '@foundation/Icon/Generated/IconStackVertical24';
-import IconStackVertical32 from '@foundation/Icon/Generated/IconStackVertical32';
-import IconStar from '@foundation/Icon/Generated/IconStar';
-import IconStar12 from '@foundation/Icon/Generated/IconStar12';
-import IconStar12Filled from '@foundation/Icon/Generated/IconStar12Filled';
-import IconStar16 from '@foundation/Icon/Generated/IconStar16';
-import IconStar16Filled from '@foundation/Icon/Generated/IconStar16Filled';
-import IconStar20 from '@foundation/Icon/Generated/IconStar20';
-import IconStar20Filled from '@foundation/Icon/Generated/IconStar20Filled';
-import IconStar24 from '@foundation/Icon/Generated/IconStar24';
-import IconStar24Filled from '@foundation/Icon/Generated/IconStar24Filled';
-import IconStar32 from '@foundation/Icon/Generated/IconStar32';
-import IconStar32Filled from '@foundation/Icon/Generated/IconStar32Filled';
-import IconStorybook from '@foundation/Icon/Generated/IconStorybook';
-import IconStorybook12 from '@foundation/Icon/Generated/IconStorybook12';
-import IconStorybook16 from '@foundation/Icon/Generated/IconStorybook16';
-import IconStorybook20 from '@foundation/Icon/Generated/IconStorybook20';
-import IconStorybook24 from '@foundation/Icon/Generated/IconStorybook24';
-import IconStorybook32 from '@foundation/Icon/Generated/IconStorybook32';
-import IconStrikethroughBox from '@foundation/Icon/Generated/IconStrikethroughBox';
-import IconStrikethroughBox12 from '@foundation/Icon/Generated/IconStrikethroughBox12';
-import IconStrikethroughBox16 from '@foundation/Icon/Generated/IconStrikethroughBox16';
-import IconStrikethroughBox20 from '@foundation/Icon/Generated/IconStrikethroughBox20';
-import IconStrikethroughBox24 from '@foundation/Icon/Generated/IconStrikethroughBox24';
-import IconStrikethroughBox32 from '@foundation/Icon/Generated/IconStrikethroughBox32';
-import IconSubscript from '@foundation/Icon/Generated/IconSubscript';
-import IconSubscript12 from '@foundation/Icon/Generated/IconSubscript12';
-import IconSubscript16 from '@foundation/Icon/Generated/IconSubscript16';
-import IconSubscript20 from '@foundation/Icon/Generated/IconSubscript20';
-import IconSubscript24 from '@foundation/Icon/Generated/IconSubscript24';
-import IconSubscript32 from '@foundation/Icon/Generated/IconSubscript32';
-import IconSuitcase from '@foundation/Icon/Generated/IconSuitcase';
-import IconSuitcase12 from '@foundation/Icon/Generated/IconSuitcase12';
-import IconSuitcase16 from '@foundation/Icon/Generated/IconSuitcase16';
-import IconSuitcase20 from '@foundation/Icon/Generated/IconSuitcase20';
-import IconSuitcase24 from '@foundation/Icon/Generated/IconSuitcase24';
-import IconSuitcase32 from '@foundation/Icon/Generated/IconSuitcase32';
-import IconSuitcaseTag from '@foundation/Icon/Generated/IconSuitcaseTag';
-import IconSuitcaseTag12 from '@foundation/Icon/Generated/IconSuitcaseTag12';
-import IconSuitcaseTag16 from '@foundation/Icon/Generated/IconSuitcaseTag16';
-import IconSuitcaseTag20 from '@foundation/Icon/Generated/IconSuitcaseTag20';
-import IconSuitcaseTag24 from '@foundation/Icon/Generated/IconSuitcaseTag24';
-import IconSuitcaseTag32 from '@foundation/Icon/Generated/IconSuitcaseTag32';
-import IconSuperscript from '@foundation/Icon/Generated/IconSuperscript';
-import IconSuperscript12 from '@foundation/Icon/Generated/IconSuperscript12';
-import IconSuperscript16 from '@foundation/Icon/Generated/IconSuperscript16';
-import IconSuperscript20 from '@foundation/Icon/Generated/IconSuperscript20';
-import IconSuperscript24 from '@foundation/Icon/Generated/IconSuperscript24';
-import IconSuperscript32 from '@foundation/Icon/Generated/IconSuperscript32';
-import IconTable from '@foundation/Icon/Generated/IconTable';
-import IconTable12 from '@foundation/Icon/Generated/IconTable12';
-import IconTable16 from '@foundation/Icon/Generated/IconTable16';
-import IconTable20 from '@foundation/Icon/Generated/IconTable20';
-import IconTable24 from '@foundation/Icon/Generated/IconTable24';
-import IconTable32 from '@foundation/Icon/Generated/IconTable32';
-import IconTarget from '@foundation/Icon/Generated/IconTarget';
-import IconTarget12 from '@foundation/Icon/Generated/IconTarget12';
-import IconTarget16 from '@foundation/Icon/Generated/IconTarget16';
-import IconTarget20 from '@foundation/Icon/Generated/IconTarget20';
-import IconTarget24 from '@foundation/Icon/Generated/IconTarget24';
-import IconTarget32 from '@foundation/Icon/Generated/IconTarget32';
-import IconTester from '@foundation/Icon/Generated/IconTester';
-import IconTester12 from '@foundation/Icon/Generated/IconTester12';
-import IconTester16 from '@foundation/Icon/Generated/IconTester16';
-import IconTester20 from '@foundation/Icon/Generated/IconTester20';
-import IconTester24 from '@foundation/Icon/Generated/IconTester24';
-import IconTester32 from '@foundation/Icon/Generated/IconTester32';
-import IconTextAlignmentCentre from '@foundation/Icon/Generated/IconTextAlignmentCentre';
-import IconTextAlignmentCentre12 from '@foundation/Icon/Generated/IconTextAlignmentCentre12';
-import IconTextAlignmentCentre16 from '@foundation/Icon/Generated/IconTextAlignmentCentre16';
-import IconTextAlignmentCentre20 from '@foundation/Icon/Generated/IconTextAlignmentCentre20';
-import IconTextAlignmentCentre24 from '@foundation/Icon/Generated/IconTextAlignmentCentre24';
-import IconTextAlignmentCentre32 from '@foundation/Icon/Generated/IconTextAlignmentCentre32';
-import IconTextAlignmentJustify from '@foundation/Icon/Generated/IconTextAlignmentJustify';
-import IconTextAlignmentJustify12 from '@foundation/Icon/Generated/IconTextAlignmentJustify12';
-import IconTextAlignmentJustify16 from '@foundation/Icon/Generated/IconTextAlignmentJustify16';
-import IconTextAlignmentJustify20 from '@foundation/Icon/Generated/IconTextAlignmentJustify20';
-import IconTextAlignmentJustify24 from '@foundation/Icon/Generated/IconTextAlignmentJustify24';
-import IconTextAlignmentJustify32 from '@foundation/Icon/Generated/IconTextAlignmentJustify32';
-import IconTextAlignmentLeft from '@foundation/Icon/Generated/IconTextAlignmentLeft';
-import IconTextAlignmentLeft12 from '@foundation/Icon/Generated/IconTextAlignmentLeft12';
-import IconTextAlignmentLeft16 from '@foundation/Icon/Generated/IconTextAlignmentLeft16';
-import IconTextAlignmentLeft20 from '@foundation/Icon/Generated/IconTextAlignmentLeft20';
-import IconTextAlignmentLeft24 from '@foundation/Icon/Generated/IconTextAlignmentLeft24';
-import IconTextAlignmentLeft32 from '@foundation/Icon/Generated/IconTextAlignmentLeft32';
-import IconTextAlignmentRight from '@foundation/Icon/Generated/IconTextAlignmentRight';
-import IconTextAlignmentRight12 from '@foundation/Icon/Generated/IconTextAlignmentRight12';
-import IconTextAlignmentRight16 from '@foundation/Icon/Generated/IconTextAlignmentRight16';
-import IconTextAlignmentRight20 from '@foundation/Icon/Generated/IconTextAlignmentRight20';
-import IconTextAlignmentRight24 from '@foundation/Icon/Generated/IconTextAlignmentRight24';
-import IconTextAlignmentRight32 from '@foundation/Icon/Generated/IconTextAlignmentRight32';
-import IconTextBoxStack from '@foundation/Icon/Generated/IconTextBoxStack';
-import IconTextBoxStack12 from '@foundation/Icon/Generated/IconTextBoxStack12';
-import IconTextBoxStack16 from '@foundation/Icon/Generated/IconTextBoxStack16';
-import IconTextBoxStack20 from '@foundation/Icon/Generated/IconTextBoxStack20';
-import IconTextBoxStack24 from '@foundation/Icon/Generated/IconTextBoxStack24';
-import IconTextBoxStack32 from '@foundation/Icon/Generated/IconTextBoxStack32';
-import IconTextBrackets from '@foundation/Icon/Generated/IconTextBrackets';
-import IconTextBrackets12 from '@foundation/Icon/Generated/IconTextBrackets12';
-import IconTextBrackets16 from '@foundation/Icon/Generated/IconTextBrackets16';
-import IconTextBrackets20 from '@foundation/Icon/Generated/IconTextBrackets20';
-import IconTextBrackets24 from '@foundation/Icon/Generated/IconTextBrackets24';
-import IconTextBrackets32 from '@foundation/Icon/Generated/IconTextBrackets32';
-import IconTextColumnBreak from '@foundation/Icon/Generated/IconTextColumnBreak';
-import IconTextColumnBreak12 from '@foundation/Icon/Generated/IconTextColumnBreak12';
-import IconTextColumnBreak16 from '@foundation/Icon/Generated/IconTextColumnBreak16';
-import IconTextColumnBreak20 from '@foundation/Icon/Generated/IconTextColumnBreak20';
-import IconTextColumnBreak24 from '@foundation/Icon/Generated/IconTextColumnBreak24';
-import IconTextColumnBreak32 from '@foundation/Icon/Generated/IconTextColumnBreak32';
-import IconTextColumns from '@foundation/Icon/Generated/IconTextColumns';
-import IconTextColumns12 from '@foundation/Icon/Generated/IconTextColumns12';
-import IconTextColumns16 from '@foundation/Icon/Generated/IconTextColumns16';
-import IconTextColumns20 from '@foundation/Icon/Generated/IconTextColumns20';
-import IconTextColumns24 from '@foundation/Icon/Generated/IconTextColumns24';
-import IconTextColumns32 from '@foundation/Icon/Generated/IconTextColumns32';
-import IconTextFormatBold from '@foundation/Icon/Generated/IconTextFormatBold';
-import IconTextFormatBold12 from '@foundation/Icon/Generated/IconTextFormatBold12';
-import IconTextFormatBold16 from '@foundation/Icon/Generated/IconTextFormatBold16';
-import IconTextFormatBold20 from '@foundation/Icon/Generated/IconTextFormatBold20';
-import IconTextFormatBold24 from '@foundation/Icon/Generated/IconTextFormatBold24';
-import IconTextFormatBold32 from '@foundation/Icon/Generated/IconTextFormatBold32';
-import IconTextFormatItalic from '@foundation/Icon/Generated/IconTextFormatItalic';
-import IconTextFormatItalic12 from '@foundation/Icon/Generated/IconTextFormatItalic12';
-import IconTextFormatItalic16 from '@foundation/Icon/Generated/IconTextFormatItalic16';
-import IconTextFormatItalic20 from '@foundation/Icon/Generated/IconTextFormatItalic20';
-import IconTextFormatItalic24 from '@foundation/Icon/Generated/IconTextFormatItalic24';
-import IconTextFormatItalic32 from '@foundation/Icon/Generated/IconTextFormatItalic32';
-import IconTextFormatStrikethrough from '@foundation/Icon/Generated/IconTextFormatStrikethrough';
-import IconTextFormatStrikethrough12 from '@foundation/Icon/Generated/IconTextFormatStrikethrough12';
-import IconTextFormatStrikethrough16 from '@foundation/Icon/Generated/IconTextFormatStrikethrough16';
-import IconTextFormatStrikethrough20 from '@foundation/Icon/Generated/IconTextFormatStrikethrough20';
-import IconTextFormatStrikethrough24 from '@foundation/Icon/Generated/IconTextFormatStrikethrough24';
-import IconTextFormatStrikethrough32 from '@foundation/Icon/Generated/IconTextFormatStrikethrough32';
-import IconTextFormatUnderline from '@foundation/Icon/Generated/IconTextFormatUnderline';
-import IconTextFormatUnderline12 from '@foundation/Icon/Generated/IconTextFormatUnderline12';
-import IconTextFormatUnderline16 from '@foundation/Icon/Generated/IconTextFormatUnderline16';
-import IconTextFormatUnderline20 from '@foundation/Icon/Generated/IconTextFormatUnderline20';
-import IconTextFormatUnderline24 from '@foundation/Icon/Generated/IconTextFormatUnderline24';
-import IconTextFormatUnderline32 from '@foundation/Icon/Generated/IconTextFormatUnderline32';
-import IconTextIndent from '@foundation/Icon/Generated/IconTextIndent';
-import IconTextIndent12 from '@foundation/Icon/Generated/IconTextIndent12';
-import IconTextIndent16 from '@foundation/Icon/Generated/IconTextIndent16';
-import IconTextIndent20 from '@foundation/Icon/Generated/IconTextIndent20';
-import IconTextIndent24 from '@foundation/Icon/Generated/IconTextIndent24';
-import IconTextIndent32 from '@foundation/Icon/Generated/IconTextIndent32';
-import IconTextOutdent from '@foundation/Icon/Generated/IconTextOutdent';
-import IconTextOutdent12 from '@foundation/Icon/Generated/IconTextOutdent12';
-import IconTextOutdent16 from '@foundation/Icon/Generated/IconTextOutdent16';
-import IconTextOutdent20 from '@foundation/Icon/Generated/IconTextOutdent20';
-import IconTextOutdent24 from '@foundation/Icon/Generated/IconTextOutdent24';
-import IconTextOutdent32 from '@foundation/Icon/Generated/IconTextOutdent32';
-import IconTrashBin from '@foundation/Icon/Generated/IconTrashBin';
-import IconTrashBin12 from '@foundation/Icon/Generated/IconTrashBin12';
-import IconTrashBin16 from '@foundation/Icon/Generated/IconTrashBin16';
-import IconTrashBin20 from '@foundation/Icon/Generated/IconTrashBin20';
-import IconTrashBin24 from '@foundation/Icon/Generated/IconTrashBin24';
-import IconTrashBin32 from '@foundation/Icon/Generated/IconTrashBin32';
-import IconTrashBinOpen from '@foundation/Icon/Generated/IconTrashBinOpen';
-import IconTrashBinOpen12 from '@foundation/Icon/Generated/IconTrashBinOpen12';
-import IconTrashBinOpen16 from '@foundation/Icon/Generated/IconTrashBinOpen16';
-import IconTrashBinOpen20 from '@foundation/Icon/Generated/IconTrashBinOpen20';
-import IconTrashBinOpen24 from '@foundation/Icon/Generated/IconTrashBinOpen24';
-import IconTrashBinOpen32 from '@foundation/Icon/Generated/IconTrashBinOpen32';
-import IconTriangle from '@foundation/Icon/Generated/IconTriangle';
-import IconTriangle12 from '@foundation/Icon/Generated/IconTriangle12';
-import IconTriangle16 from '@foundation/Icon/Generated/IconTriangle16';
-import IconTriangle20 from '@foundation/Icon/Generated/IconTriangle20';
-import IconTriangle24 from '@foundation/Icon/Generated/IconTriangle24';
-import IconTriangle32 from '@foundation/Icon/Generated/IconTriangle32';
-import IconTwitter from '@foundation/Icon/Generated/IconTwitter';
-import IconTwitter12 from '@foundation/Icon/Generated/IconTwitter12';
-import IconTwitter16 from '@foundation/Icon/Generated/IconTwitter16';
-import IconTwitter20 from '@foundation/Icon/Generated/IconTwitter20';
-import IconTwitter24 from '@foundation/Icon/Generated/IconTwitter24';
-import IconTwitter32 from '@foundation/Icon/Generated/IconTwitter32';
-import IconTypography from '@foundation/Icon/Generated/IconTypography';
-import IconTypography12 from '@foundation/Icon/Generated/IconTypography12';
-import IconTypography16 from '@foundation/Icon/Generated/IconTypography16';
-import IconTypography20 from '@foundation/Icon/Generated/IconTypography20';
-import IconTypography24 from '@foundation/Icon/Generated/IconTypography24';
-import IconTypography32 from '@foundation/Icon/Generated/IconTypography32';
-import IconTypographyBox from '@foundation/Icon/Generated/IconTypographyBox';
-import IconTypographyBox12 from '@foundation/Icon/Generated/IconTypographyBox12';
-import IconTypographyBox16 from '@foundation/Icon/Generated/IconTypographyBox16';
-import IconTypographyBox20 from '@foundation/Icon/Generated/IconTypographyBox20';
-import IconTypographyBox24 from '@foundation/Icon/Generated/IconTypographyBox24';
-import IconTypographyBox32 from '@foundation/Icon/Generated/IconTypographyBox32';
-import IconTypographyCase from '@foundation/Icon/Generated/IconTypographyCase';
-import IconTypographyCase12 from '@foundation/Icon/Generated/IconTypographyCase12';
-import IconTypographyCase16 from '@foundation/Icon/Generated/IconTypographyCase16';
-import IconTypographyCase20 from '@foundation/Icon/Generated/IconTypographyCase20';
-import IconTypographyCase24 from '@foundation/Icon/Generated/IconTypographyCase24';
-import IconTypographyCase32 from '@foundation/Icon/Generated/IconTypographyCase32';
-import IconTypographyMultiple from '@foundation/Icon/Generated/IconTypographyMultiple';
-import IconTypographyMultiple12 from '@foundation/Icon/Generated/IconTypographyMultiple12';
-import IconTypographyMultiple16 from '@foundation/Icon/Generated/IconTypographyMultiple16';
-import IconTypographyMultiple20 from '@foundation/Icon/Generated/IconTypographyMultiple20';
-import IconTypographyMultiple24 from '@foundation/Icon/Generated/IconTypographyMultiple24';
-import IconTypographyMultiple32 from '@foundation/Icon/Generated/IconTypographyMultiple32';
-import IconVimeo from '@foundation/Icon/Generated/IconVimeo';
-import IconVimeo12 from '@foundation/Icon/Generated/IconVimeo12';
-import IconVimeo16 from '@foundation/Icon/Generated/IconVimeo16';
-import IconVimeo20 from '@foundation/Icon/Generated/IconVimeo20';
-import IconVimeo24 from '@foundation/Icon/Generated/IconVimeo24';
-import IconVimeo32 from '@foundation/Icon/Generated/IconVimeo32';
-import IconWorkflowLanes from '@foundation/Icon/Generated/IconWorkflowLanes';
-import IconWorkflowLanes12 from '@foundation/Icon/Generated/IconWorkflowLanes12';
-import IconWorkflowLanes16 from '@foundation/Icon/Generated/IconWorkflowLanes16';
-import IconWorkflowLanes20 from '@foundation/Icon/Generated/IconWorkflowLanes20';
-import IconWorkflowLanes24 from '@foundation/Icon/Generated/IconWorkflowLanes24';
-import IconWorkflowLanes32 from '@foundation/Icon/Generated/IconWorkflowLanes32';
-import IconWorkflowShapes from '@foundation/Icon/Generated/IconWorkflowShapes';
-import IconWorkflowShapes12 from '@foundation/Icon/Generated/IconWorkflowShapes12';
-import IconWorkflowShapes16 from '@foundation/Icon/Generated/IconWorkflowShapes16';
-import IconWorkflowShapes20 from '@foundation/Icon/Generated/IconWorkflowShapes20';
-import IconWorkflowShapes24 from '@foundation/Icon/Generated/IconWorkflowShapes24';
-import IconWorkflowShapes32 from '@foundation/Icon/Generated/IconWorkflowShapes32';
-
-export const iconsMap: Record<IconEnum, ReactElement> = {
-    [IconEnum.AdobeCreativeCloud]: <IconAdobeCreativeCloud />,
-    [IconEnum.AdobeCreativeCloud12]: <IconAdobeCreativeCloud12 />,
-    [IconEnum.AdobeCreativeCloud16]: <IconAdobeCreativeCloud16 />,
-    [IconEnum.AdobeCreativeCloud20]: <IconAdobeCreativeCloud20 />,
-    [IconEnum.AdobeCreativeCloud24]: <IconAdobeCreativeCloud24 />,
-    [IconEnum.AdobeCreativeCloud32]: <IconAdobeCreativeCloud32 />,
-    [IconEnum.AdobeInDesign]: <IconAdobeInDesign />,
-    [IconEnum.AdobeInDesign12]: <IconAdobeInDesign12 />,
-    [IconEnum.AdobeInDesign16]: <IconAdobeInDesign16 />,
-    [IconEnum.AdobeInDesign20]: <IconAdobeInDesign20 />,
-    [IconEnum.AdobeInDesign24]: <IconAdobeInDesign24 />,
-    [IconEnum.AdobeInDesign32]: <IconAdobeInDesign32 />,
-    [IconEnum.Anchor]: <IconAnchor />,
-    [IconEnum.Anchor12]: <IconAnchor12 />,
-    [IconEnum.Anchor16]: <IconAnchor16 />,
-    [IconEnum.Anchor20]: <IconAnchor20 />,
-    [IconEnum.Anchor24]: <IconAnchor24 />,
-    [IconEnum.Anchor32]: <IconAnchor32 />,
-    [IconEnum.AnnotationLines]: <IconAnnotationLines />,
-    [IconEnum.AnnotationLines12]: <IconAnnotationLines12 />,
-    [IconEnum.AnnotationLines16]: <IconAnnotationLines16 />,
-    [IconEnum.AnnotationLines20]: <IconAnnotationLines20 />,
-    [IconEnum.AnnotationLines24]: <IconAnnotationLines24 />,
-    [IconEnum.AnnotationLines32]: <IconAnnotationLines32 />,
-    [IconEnum.AnnotationNumbers]: <IconAnnotationNumbers />,
-    [IconEnum.AnnotationNumbers12]: <IconAnnotationNumbers12 />,
-    [IconEnum.AnnotationNumbers16]: <IconAnnotationNumbers16 />,
-    [IconEnum.AnnotationNumbers20]: <IconAnnotationNumbers20 />,
-    [IconEnum.AnnotationNumbers24]: <IconAnnotationNumbers24 />,
-    [IconEnum.AnnotationNumbers32]: <IconAnnotationNumbers32 />,
-    [IconEnum.Archive]: <IconArchive />,
-    [IconEnum.Archive12]: <IconArchive12 />,
-    [IconEnum.Archive16]: <IconArchive16 />,
-    [IconEnum.Archive20]: <IconArchive20 />,
-    [IconEnum.Archive24]: <IconArchive24 />,
-    [IconEnum.Archive32]: <IconArchive32 />,
-    [IconEnum.ArrowAlignDown]: <IconArrowAlignDown />,
-    [IconEnum.ArrowAlignDown12]: <IconArrowAlignDown12 />,
-    [IconEnum.ArrowAlignDown16]: <IconArrowAlignDown16 />,
-    [IconEnum.ArrowAlignDown20]: <IconArrowAlignDown20 />,
-    [IconEnum.ArrowAlignDown24]: <IconArrowAlignDown24 />,
-    [IconEnum.ArrowAlignDown32]: <IconArrowAlignDown32 />,
-    [IconEnum.ArrowAlignHorizontalCentre]: <IconArrowAlignHorizontalCentre />,
-    [IconEnum.ArrowAlignHorizontalCentre12]: <IconArrowAlignHorizontalCentre12 />,
-    [IconEnum.ArrowAlignHorizontalCentre16]: <IconArrowAlignHorizontalCentre16 />,
-    [IconEnum.ArrowAlignHorizontalCentre20]: <IconArrowAlignHorizontalCentre20 />,
-    [IconEnum.ArrowAlignHorizontalCentre24]: <IconArrowAlignHorizontalCentre24 />,
-    [IconEnum.ArrowAlignHorizontalCentre32]: <IconArrowAlignHorizontalCentre32 />,
-    [IconEnum.ArrowAlignLeft]: <IconArrowAlignLeft />,
-    [IconEnum.ArrowAlignLeft12]: <IconArrowAlignLeft12 />,
-    [IconEnum.ArrowAlignLeft16]: <IconArrowAlignLeft16 />,
-    [IconEnum.ArrowAlignLeft20]: <IconArrowAlignLeft20 />,
-    [IconEnum.ArrowAlignLeft24]: <IconArrowAlignLeft24 />,
-    [IconEnum.ArrowAlignLeft32]: <IconArrowAlignLeft32 />,
-    [IconEnum.ArrowAlignRight]: <IconArrowAlignRight />,
-    [IconEnum.ArrowAlignRight12]: <IconArrowAlignRight12 />,
-    [IconEnum.ArrowAlignRight16]: <IconArrowAlignRight16 />,
-    [IconEnum.ArrowAlignRight20]: <IconArrowAlignRight20 />,
-    [IconEnum.ArrowAlignRight24]: <IconArrowAlignRight24 />,
-    [IconEnum.ArrowAlignRight32]: <IconArrowAlignRight32 />,
-    [IconEnum.ArrowAlignUp]: <IconArrowAlignUp />,
-    [IconEnum.ArrowAlignUp12]: <IconArrowAlignUp12 />,
-    [IconEnum.ArrowAlignUp16]: <IconArrowAlignUp16 />,
-    [IconEnum.ArrowAlignUp20]: <IconArrowAlignUp20 />,
-    [IconEnum.ArrowAlignUp24]: <IconArrowAlignUp24 />,
-    [IconEnum.ArrowAlignUp32]: <IconArrowAlignUp32 />,
-    [IconEnum.ArrowAlignVerticalCentre]: <IconArrowAlignVerticalCentre />,
-    [IconEnum.ArrowAlignVerticalCentre12]: <IconArrowAlignVerticalCentre12 />,
-    [IconEnum.ArrowAlignVerticalCentre16]: <IconArrowAlignVerticalCentre16 />,
-    [IconEnum.ArrowAlignVerticalCentre20]: <IconArrowAlignVerticalCentre20 />,
-    [IconEnum.ArrowAlignVerticalCentre24]: <IconArrowAlignVerticalCentre24 />,
-    [IconEnum.ArrowAlignVerticalCentre32]: <IconArrowAlignVerticalCentre32 />,
-    [IconEnum.ArrowBidirectional]: <IconArrowBidirectional />,
-    [IconEnum.ArrowBidirectional12]: <IconArrowBidirectional12 />,
-    [IconEnum.ArrowBidirectional16]: <IconArrowBidirectional16 />,
-    [IconEnum.ArrowBidirectional20]: <IconArrowBidirectional20 />,
-    [IconEnum.ArrowBidirectional24]: <IconArrowBidirectional24 />,
-    [IconEnum.ArrowBidirectional32]: <IconArrowBidirectional32 />,
-    [IconEnum.ArrowBidirectionalHorizontal]: <IconArrowBidirectionalHorizontal />,
-    [IconEnum.ArrowBidirectionalHorizontal12]: <IconArrowBidirectionalHorizontal12 />,
-    [IconEnum.ArrowBidirectionalHorizontal16]: <IconArrowBidirectionalHorizontal16 />,
-    [IconEnum.ArrowBidirectionalHorizontal20]: <IconArrowBidirectionalHorizontal20 />,
-    [IconEnum.ArrowBidirectionalHorizontal24]: <IconArrowBidirectionalHorizontal24 />,
-    [IconEnum.ArrowBidirectionalHorizontal32]: <IconArrowBidirectionalHorizontal32 />,
-    [IconEnum.ArrowBottomToTop]: <IconArrowBottomToTop />,
-    [IconEnum.ArrowBottomToTop12]: <IconArrowBottomToTop12 />,
-    [IconEnum.ArrowBottomToTop16]: <IconArrowBottomToTop16 />,
-    [IconEnum.ArrowBottomToTop20]: <IconArrowBottomToTop20 />,
-    [IconEnum.ArrowBottomToTop24]: <IconArrowBottomToTop24 />,
-    [IconEnum.ArrowBottomToTop32]: <IconArrowBottomToTop32 />,
-    [IconEnum.ArrowCircleDown]: <IconArrowCircleDown />,
-    [IconEnum.ArrowCircleDown12]: <IconArrowCircleDown12 />,
-    [IconEnum.ArrowCircleDown16]: <IconArrowCircleDown16 />,
-    [IconEnum.ArrowCircleDown20]: <IconArrowCircleDown20 />,
-    [IconEnum.ArrowCircleDown24]: <IconArrowCircleDown24 />,
-    [IconEnum.ArrowCircleDown32]: <IconArrowCircleDown32 />,
-    [IconEnum.ArrowCircleUp]: <IconArrowCircleUp />,
-    [IconEnum.ArrowCircleUp12]: <IconArrowCircleUp12 />,
-    [IconEnum.ArrowCircleUp16]: <IconArrowCircleUp16 />,
-    [IconEnum.ArrowCircleUp20]: <IconArrowCircleUp20 />,
-    [IconEnum.ArrowCircleUp24]: <IconArrowCircleUp24 />,
-    [IconEnum.ArrowCircleUp32]: <IconArrowCircleUp32 />,
-    [IconEnum.ArrowDown]: <IconArrowDown />,
-    [IconEnum.ArrowDown12]: <IconArrowDown12 />,
-    [IconEnum.ArrowDown16]: <IconArrowDown16 />,
-    [IconEnum.ArrowDown20]: <IconArrowDown20 />,
-    [IconEnum.ArrowDown24]: <IconArrowDown24 />,
-    [IconEnum.ArrowDown32]: <IconArrowDown32 />,
-    [IconEnum.ArrowExpand]: <IconArrowExpand />,
-    [IconEnum.ArrowExpand12]: <IconArrowExpand12 />,
-    [IconEnum.ArrowExpand16]: <IconArrowExpand16 />,
-    [IconEnum.ArrowExpand20]: <IconArrowExpand20 />,
-    [IconEnum.ArrowExpand24]: <IconArrowExpand24 />,
-    [IconEnum.ArrowExpand32]: <IconArrowExpand32 />,
-    [IconEnum.ArrowFitBottomLeft]: <IconArrowFitBottomLeft />,
-    [IconEnum.ArrowFitBottomLeft12]: <IconArrowFitBottomLeft12 />,
-    [IconEnum.ArrowFitBottomLeft16]: <IconArrowFitBottomLeft16 />,
-    [IconEnum.ArrowFitBottomLeft20]: <IconArrowFitBottomLeft20 />,
-    [IconEnum.ArrowFitBottomLeft24]: <IconArrowFitBottomLeft24 />,
-    [IconEnum.ArrowFitBottomLeft32]: <IconArrowFitBottomLeft32 />,
-    [IconEnum.ArrowFitBottomRight]: <IconArrowFitBottomRight />,
-    [IconEnum.ArrowFitBottomRight12]: <IconArrowFitBottomRight12 />,
-    [IconEnum.ArrowFitBottomRight16]: <IconArrowFitBottomRight16 />,
-    [IconEnum.ArrowFitBottomRight20]: <IconArrowFitBottomRight20 />,
-    [IconEnum.ArrowFitBottomRight24]: <IconArrowFitBottomRight24 />,
-    [IconEnum.ArrowFitBottomRight32]: <IconArrowFitBottomRight32 />,
-    [IconEnum.ArrowFitTopLeft]: <IconArrowFitTopLeft />,
-    [IconEnum.ArrowFitTopLeft12]: <IconArrowFitTopLeft12 />,
-    [IconEnum.ArrowFitTopLeft16]: <IconArrowFitTopLeft16 />,
-    [IconEnum.ArrowFitTopLeft20]: <IconArrowFitTopLeft20 />,
-    [IconEnum.ArrowFitTopLeft24]: <IconArrowFitTopLeft24 />,
-    [IconEnum.ArrowFitTopLeft32]: <IconArrowFitTopLeft32 />,
-    [IconEnum.ArrowFitTopRight]: <IconArrowFitTopRight />,
-    [IconEnum.ArrowFitTopRight12]: <IconArrowFitTopRight12 />,
-    [IconEnum.ArrowFitTopRight16]: <IconArrowFitTopRight16 />,
-    [IconEnum.ArrowFitTopRight20]: <IconArrowFitTopRight20 />,
-    [IconEnum.ArrowFitTopRight24]: <IconArrowFitTopRight24 />,
-    [IconEnum.ArrowFitTopRight32]: <IconArrowFitTopRight32 />,
-    [IconEnum.ArrowHook]: <IconArrowHook />,
-    [IconEnum.ArrowHook12]: <IconArrowHook12 />,
-    [IconEnum.ArrowHook16]: <IconArrowHook16 />,
-    [IconEnum.ArrowHook20]: <IconArrowHook20 />,
-    [IconEnum.ArrowHook24]: <IconArrowHook24 />,
-    [IconEnum.ArrowHook32]: <IconArrowHook32 />,
-    [IconEnum.ArrowLeft]: <IconArrowLeft />,
-    [IconEnum.ArrowLeft12]: <IconArrowLeft12 />,
-    [IconEnum.ArrowLeft16]: <IconArrowLeft16 />,
-    [IconEnum.ArrowLeft20]: <IconArrowLeft20 />,
-    [IconEnum.ArrowLeft24]: <IconArrowLeft24 />,
-    [IconEnum.ArrowLeft32]: <IconArrowLeft32 />,
-    [IconEnum.ArrowLeftToRight]: <IconArrowLeftToRight />,
-    [IconEnum.ArrowLeftToRight12]: <IconArrowLeftToRight12 />,
-    [IconEnum.ArrowLeftToRight16]: <IconArrowLeftToRight16 />,
-    [IconEnum.ArrowLeftToRight20]: <IconArrowLeftToRight20 />,
-    [IconEnum.ArrowLeftToRight24]: <IconArrowLeftToRight24 />,
-    [IconEnum.ArrowLeftToRight32]: <IconArrowLeftToRight32 />,
-    [IconEnum.ArrowMinimize]: <IconArrowMinimize />,
-    [IconEnum.ArrowMinimize12]: <IconArrowMinimize12 />,
-    [IconEnum.ArrowMinimize16]: <IconArrowMinimize16 />,
-    [IconEnum.ArrowMinimize20]: <IconArrowMinimize20 />,
-    [IconEnum.ArrowMinimize24]: <IconArrowMinimize24 />,
-    [IconEnum.ArrowMinimize32]: <IconArrowMinimize32 />,
-    [IconEnum.ArrowMove]: <IconArrowMove />,
-    [IconEnum.ArrowMove12]: <IconArrowMove12 />,
-    [IconEnum.ArrowMove16]: <IconArrowMove16 />,
-    [IconEnum.ArrowMove20]: <IconArrowMove20 />,
-    [IconEnum.ArrowMove24]: <IconArrowMove24 />,
-    [IconEnum.ArrowMove32]: <IconArrowMove32 />,
-    [IconEnum.ArrowOutExternal]: <IconArrowOutExternal />,
-    [IconEnum.ArrowOutExternal12]: <IconArrowOutExternal12 />,
-    [IconEnum.ArrowOutExternal16]: <IconArrowOutExternal16 />,
-    [IconEnum.ArrowOutExternal20]: <IconArrowOutExternal20 />,
-    [IconEnum.ArrowOutExternal24]: <IconArrowOutExternal24 />,
-    [IconEnum.ArrowOutExternal32]: <IconArrowOutExternal32 />,
-    [IconEnum.ArrowOutLogout]: <IconArrowOutLogout />,
-    [IconEnum.ArrowOutLogout12]: <IconArrowOutLogout12 />,
-    [IconEnum.ArrowOutLogout16]: <IconArrowOutLogout16 />,
-    [IconEnum.ArrowOutLogout20]: <IconArrowOutLogout20 />,
-    [IconEnum.ArrowOutLogout24]: <IconArrowOutLogout24 />,
-    [IconEnum.ArrowOutLogout32]: <IconArrowOutLogout32 />,
-    [IconEnum.ArrowOutShare]: <IconArrowOutShare />,
-    [IconEnum.ArrowOutShare12]: <IconArrowOutShare12 />,
-    [IconEnum.ArrowOutShare16]: <IconArrowOutShare16 />,
-    [IconEnum.ArrowOutShare20]: <IconArrowOutShare20 />,
-    [IconEnum.ArrowOutShare24]: <IconArrowOutShare24 />,
-    [IconEnum.ArrowOutShare32]: <IconArrowOutShare32 />,
-    [IconEnum.ArrowRight]: <IconArrowRight />,
-    [IconEnum.ArrowRight12]: <IconArrowRight12 />,
-    [IconEnum.ArrowRight16]: <IconArrowRight16 />,
-    [IconEnum.ArrowRight20]: <IconArrowRight20 />,
-    [IconEnum.ArrowRight24]: <IconArrowRight24 />,
-    [IconEnum.ArrowRight32]: <IconArrowRight32 />,
-    [IconEnum.ArrowRightToLeft]: <IconArrowRightToLeft />,
-    [IconEnum.ArrowRightToLeft12]: <IconArrowRightToLeft12 />,
-    [IconEnum.ArrowRightToLeft16]: <IconArrowRightToLeft16 />,
-    [IconEnum.ArrowRightToLeft20]: <IconArrowRightToLeft20 />,
-    [IconEnum.ArrowRightToLeft24]: <IconArrowRightToLeft24 />,
-    [IconEnum.ArrowRightToLeft32]: <IconArrowRightToLeft32 />,
-    [IconEnum.ArrowRoundAntiClockwise]: <IconArrowRoundAntiClockwise />,
-    [IconEnum.ArrowRoundAntiClockwise12]: <IconArrowRoundAntiClockwise12 />,
-    [IconEnum.ArrowRoundAntiClockwise16]: <IconArrowRoundAntiClockwise16 />,
-    [IconEnum.ArrowRoundAntiClockwise20]: <IconArrowRoundAntiClockwise20 />,
-    [IconEnum.ArrowRoundAntiClockwise24]: <IconArrowRoundAntiClockwise24 />,
-    [IconEnum.ArrowRoundAntiClockwise32]: <IconArrowRoundAntiClockwise32 />,
-    [IconEnum.ArrowRoundClockwise]: <IconArrowRoundClockwise />,
-    [IconEnum.ArrowRoundClockwise12]: <IconArrowRoundClockwise12 />,
-    [IconEnum.ArrowRoundClockwise16]: <IconArrowRoundClockwise16 />,
-    [IconEnum.ArrowRoundClockwise20]: <IconArrowRoundClockwise20 />,
-    [IconEnum.ArrowRoundClockwise24]: <IconArrowRoundClockwise24 />,
-    [IconEnum.ArrowRoundClockwise32]: <IconArrowRoundClockwise32 />,
-    [IconEnum.ArrowScaleHorizontal]: <IconArrowScaleHorizontal />,
-    [IconEnum.ArrowScaleHorizontal12]: <IconArrowScaleHorizontal12 />,
-    [IconEnum.ArrowScaleHorizontal16]: <IconArrowScaleHorizontal16 />,
-    [IconEnum.ArrowScaleHorizontal20]: <IconArrowScaleHorizontal20 />,
-    [IconEnum.ArrowScaleHorizontal24]: <IconArrowScaleHorizontal24 />,
-    [IconEnum.ArrowScaleHorizontal32]: <IconArrowScaleHorizontal32 />,
-    [IconEnum.ArrowScaleVertical]: <IconArrowScaleVertical />,
-    [IconEnum.ArrowScaleVertical12]: <IconArrowScaleVertical12 />,
-    [IconEnum.ArrowScaleVertical16]: <IconArrowScaleVertical16 />,
-    [IconEnum.ArrowScaleVertical20]: <IconArrowScaleVertical20 />,
-    [IconEnum.ArrowScaleVertical24]: <IconArrowScaleVertical24 />,
-    [IconEnum.ArrowScaleVertical32]: <IconArrowScaleVertical32 />,
-    [IconEnum.ArrowStretchBox]: <IconArrowStretchBox />,
-    [IconEnum.ArrowStretchBox12]: <IconArrowStretchBox12 />,
-    [IconEnum.ArrowStretchBox16]: <IconArrowStretchBox16 />,
-    [IconEnum.ArrowStretchBox20]: <IconArrowStretchBox20 />,
-    [IconEnum.ArrowStretchBox24]: <IconArrowStretchBox24 />,
-    [IconEnum.ArrowStretchBox32]: <IconArrowStretchBox32 />,
-    [IconEnum.ArrowSwap]: <IconArrowSwap />,
-    [IconEnum.ArrowSwap12]: <IconArrowSwap12 />,
-    [IconEnum.ArrowSwap16]: <IconArrowSwap16 />,
-    [IconEnum.ArrowSwap20]: <IconArrowSwap20 />,
-    [IconEnum.ArrowSwap24]: <IconArrowSwap24 />,
-    [IconEnum.ArrowSwap32]: <IconArrowSwap32 />,
-    [IconEnum.ArrowSync]: <IconArrowSync />,
-    [IconEnum.ArrowSync12]: <IconArrowSync12 />,
-    [IconEnum.ArrowSync16]: <IconArrowSync16 />,
-    [IconEnum.ArrowSync20]: <IconArrowSync20 />,
-    [IconEnum.ArrowSync24]: <IconArrowSync24 />,
-    [IconEnum.ArrowSync32]: <IconArrowSync32 />,
-    [IconEnum.ArrowTopToBottom]: <IconArrowTopToBottom />,
-    [IconEnum.ArrowTopToBottom12]: <IconArrowTopToBottom12 />,
-    [IconEnum.ArrowTopToBottom16]: <IconArrowTopToBottom16 />,
-    [IconEnum.ArrowTopToBottom20]: <IconArrowTopToBottom20 />,
-    [IconEnum.ArrowTopToBottom24]: <IconArrowTopToBottom24 />,
-    [IconEnum.ArrowTopToBottom32]: <IconArrowTopToBottom32 />,
-    [IconEnum.ArrowUp]: <IconArrowUp />,
-    [IconEnum.ArrowUp12]: <IconArrowUp12 />,
-    [IconEnum.ArrowUp16]: <IconArrowUp16 />,
-    [IconEnum.ArrowUp20]: <IconArrowUp20 />,
-    [IconEnum.ArrowUp24]: <IconArrowUp24 />,
-    [IconEnum.ArrowUp32]: <IconArrowUp32 />,
-    [IconEnum.Bar]: <IconBar />,
-    [IconEnum.Bar12]: <IconBar12 />,
-    [IconEnum.Bar16]: <IconBar16 />,
-    [IconEnum.Bar20]: <IconBar20 />,
-    [IconEnum.Bar24]: <IconBar24 />,
-    [IconEnum.Bar32]: <IconBar32 />,
-    [IconEnum.BarNarrow]: <IconBarNarrow />,
-    [IconEnum.BarNarrow12]: <IconBarNarrow12 />,
-    [IconEnum.BarNarrow16]: <IconBarNarrow16 />,
-    [IconEnum.BarNarrow20]: <IconBarNarrow20 />,
-    [IconEnum.BarNarrow24]: <IconBarNarrow24 />,
-    [IconEnum.BarNarrow32]: <IconBarNarrow32 />,
-    [IconEnum.Basketball]: <IconBasketball />,
-    [IconEnum.Basketball12]: <IconBasketball12 />,
-    [IconEnum.Basketball16]: <IconBasketball16 />,
-    [IconEnum.Basketball20]: <IconBasketball20 />,
-    [IconEnum.Basketball24]: <IconBasketball24 />,
-    [IconEnum.Basketball32]: <IconBasketball32 />,
-    [IconEnum.Bell]: <IconBell />,
-    [IconEnum.Bell12]: <IconBell12 />,
-    [IconEnum.Bell16]: <IconBell16 />,
-    [IconEnum.Bell20]: <IconBell20 />,
-    [IconEnum.Bell24]: <IconBell24 />,
-    [IconEnum.Bell32]: <IconBell32 />,
-    [IconEnum.Bookmark]: <IconBookmark />,
-    [IconEnum.Bookmark12]: <IconBookmark12 />,
-    [IconEnum.Bookmark12Filled]: <IconBookmark12Filled />,
-    [IconEnum.Bookmark16]: <IconBookmark16 />,
-    [IconEnum.Bookmark16Filled]: <IconBookmark16Filled />,
-    [IconEnum.Bookmark20]: <IconBookmark20 />,
-    [IconEnum.Bookmark20Filled]: <IconBookmark20Filled />,
-    [IconEnum.Bookmark24]: <IconBookmark24 />,
-    [IconEnum.Bookmark24Filled]: <IconBookmark24Filled />,
-    [IconEnum.Bookmark32]: <IconBookmark32 />,
-    [IconEnum.Bookmark32Filled]: <IconBookmark32Filled />,
-    [IconEnum.BracketsCurly]: <IconBracketsCurly />,
-    [IconEnum.BracketsCurly12]: <IconBracketsCurly12 />,
-    [IconEnum.BracketsCurly16]: <IconBracketsCurly16 />,
-    [IconEnum.BracketsCurly20]: <IconBracketsCurly20 />,
-    [IconEnum.BracketsCurly24]: <IconBracketsCurly24 />,
-    [IconEnum.BracketsCurly32]: <IconBracketsCurly32 />,
-    [IconEnum.Building]: <IconBuilding />,
-    [IconEnum.Building12]: <IconBuilding12 />,
-    [IconEnum.Building16]: <IconBuilding16 />,
-    [IconEnum.Building20]: <IconBuilding20 />,
-    [IconEnum.Building24]: <IconBuilding24 />,
-    [IconEnum.Building32]: <IconBuilding32 />,
-    [IconEnum.BuildingBlock]: <IconBuildingBlock />,
-    [IconEnum.BuildingBlock12]: <IconBuildingBlock12 />,
-    [IconEnum.BuildingBlock16]: <IconBuildingBlock16 />,
-    [IconEnum.BuildingBlock20]: <IconBuildingBlock20 />,
-    [IconEnum.BuildingBlock24]: <IconBuildingBlock24 />,
-    [IconEnum.BuildingBlock32]: <IconBuildingBlock32 />,
-    [IconEnum.Buildings]: <IconBuildings />,
-    [IconEnum.Buildings12]: <IconBuildings12 />,
-    [IconEnum.Buildings16]: <IconBuildings16 />,
-    [IconEnum.Buildings20]: <IconBuildings20 />,
-    [IconEnum.Buildings24]: <IconBuildings24 />,
-    [IconEnum.Buildings32]: <IconBuildings32 />,
-    [IconEnum.Button]: <IconButton />,
-    [IconEnum.Button12]: <IconButton12 />,
-    [IconEnum.Button16]: <IconButton16 />,
-    [IconEnum.Button20]: <IconButton20 />,
-    [IconEnum.Button24]: <IconButton24 />,
-    [IconEnum.Button32]: <IconButton32 />,
-    [IconEnum.Calendar]: <IconCalendar />,
-    [IconEnum.Calendar12]: <IconCalendar12 />,
-    [IconEnum.Calendar16]: <IconCalendar16 />,
-    [IconEnum.Calendar20]: <IconCalendar20 />,
-    [IconEnum.Calendar24]: <IconCalendar24 />,
-    [IconEnum.Calendar32]: <IconCalendar32 />,
-    [IconEnum.CalendarChecked]: <IconCalendarChecked />,
-    [IconEnum.CalendarChecked12]: <IconCalendarChecked12 />,
-    [IconEnum.CalendarChecked16]: <IconCalendarChecked16 />,
-    [IconEnum.CalendarChecked20]: <IconCalendarChecked20 />,
-    [IconEnum.CalendarChecked24]: <IconCalendarChecked24 />,
-    [IconEnum.CalendarChecked32]: <IconCalendarChecked32 />,
-    [IconEnum.Cap]: <IconCap />,
-    [IconEnum.Cap12]: <IconCap12 />,
-    [IconEnum.Cap16]: <IconCap16 />,
-    [IconEnum.Cap20]: <IconCap20 />,
-    [IconEnum.Cap24]: <IconCap24 />,
-    [IconEnum.Cap32]: <IconCap32 />,
-    [IconEnum.Card]: <IconCard />,
-    [IconEnum.Card12]: <IconCard12 />,
-    [IconEnum.Card16]: <IconCard16 />,
-    [IconEnum.Card20]: <IconCard20 />,
-    [IconEnum.Card24]: <IconCard24 />,
-    [IconEnum.Card32]: <IconCard32 />,
-    [IconEnum.CardLayoutLarge]: <IconCardLayoutLarge />,
-    [IconEnum.CardLayoutLarge12]: <IconCardLayoutLarge12 />,
-    [IconEnum.CardLayoutLarge16]: <IconCardLayoutLarge16 />,
-    [IconEnum.CardLayoutLarge20]: <IconCardLayoutLarge20 />,
-    [IconEnum.CardLayoutLarge24]: <IconCardLayoutLarge24 />,
-    [IconEnum.CardLayoutLarge32]: <IconCardLayoutLarge32 />,
-    [IconEnum.CardLayoutMedium]: <IconCardLayoutMedium />,
-    [IconEnum.CardLayoutMedium12]: <IconCardLayoutMedium12 />,
-    [IconEnum.CardLayoutMedium16]: <IconCardLayoutMedium16 />,
-    [IconEnum.CardLayoutMedium20]: <IconCardLayoutMedium20 />,
-    [IconEnum.CardLayoutMedium24]: <IconCardLayoutMedium24 />,
-    [IconEnum.CardLayoutMedium32]: <IconCardLayoutMedium32 />,
-    [IconEnum.CardLayoutSmall]: <IconCardLayoutSmall />,
-    [IconEnum.CardLayoutSmall12]: <IconCardLayoutSmall12 />,
-    [IconEnum.CardLayoutSmall16]: <IconCardLayoutSmall16 />,
-    [IconEnum.CardLayoutSmall20]: <IconCardLayoutSmall20 />,
-    [IconEnum.CardLayoutSmall24]: <IconCardLayoutSmall24 />,
-    [IconEnum.CardLayoutSmall32]: <IconCardLayoutSmall32 />,
-    [IconEnum.CaretDown]: <IconCaretDown />,
-    [IconEnum.CaretDown12]: <IconCaretDown12 />,
-    [IconEnum.CaretDown16]: <IconCaretDown16 />,
-    [IconEnum.CaretDown20]: <IconCaretDown20 />,
-    [IconEnum.CaretDown24]: <IconCaretDown24 />,
-    [IconEnum.CaretDown32]: <IconCaretDown32 />,
-    [IconEnum.CaretLeft]: <IconCaretLeft />,
-    [IconEnum.CaretLeft12]: <IconCaretLeft12 />,
-    [IconEnum.CaretLeft16]: <IconCaretLeft16 />,
-    [IconEnum.CaretLeft20]: <IconCaretLeft20 />,
-    [IconEnum.CaretLeft24]: <IconCaretLeft24 />,
-    [IconEnum.CaretLeft32]: <IconCaretLeft32 />,
-    [IconEnum.CaretLeftDouble]: <IconCaretLeftDouble />,
-    [IconEnum.CaretLeftDouble12]: <IconCaretLeftDouble12 />,
-    [IconEnum.CaretLeftDouble16]: <IconCaretLeftDouble16 />,
-    [IconEnum.CaretLeftDouble20]: <IconCaretLeftDouble20 />,
-    [IconEnum.CaretLeftDouble24]: <IconCaretLeftDouble24 />,
-    [IconEnum.CaretLeftDouble32]: <IconCaretLeftDouble32 />,
-    [IconEnum.CaretRight]: <IconCaretRight />,
-    [IconEnum.CaretRight12]: <IconCaretRight12 />,
-    [IconEnum.CaretRight16]: <IconCaretRight16 />,
-    [IconEnum.CaretRight20]: <IconCaretRight20 />,
-    [IconEnum.CaretRight24]: <IconCaretRight24 />,
-    [IconEnum.CaretRight32]: <IconCaretRight32 />,
-    [IconEnum.CaretRightDouble]: <IconCaretRightDouble />,
-    [IconEnum.CaretRightDouble12]: <IconCaretRightDouble12 />,
-    [IconEnum.CaretRightDouble16]: <IconCaretRightDouble16 />,
-    [IconEnum.CaretRightDouble20]: <IconCaretRightDouble20 />,
-    [IconEnum.CaretRightDouble24]: <IconCaretRightDouble24 />,
-    [IconEnum.CaretRightDouble32]: <IconCaretRightDouble32 />,
-    [IconEnum.CaretUp]: <IconCaretUp />,
-    [IconEnum.CaretUp12]: <IconCaretUp12 />,
-    [IconEnum.CaretUp16]: <IconCaretUp16 />,
-    [IconEnum.CaretUp20]: <IconCaretUp20 />,
-    [IconEnum.CaretUp24]: <IconCaretUp24 />,
-    [IconEnum.CaretUp32]: <IconCaretUp32 />,
-    [IconEnum.Carousel]: <IconCarousel />,
-    [IconEnum.Carousel12]: <IconCarousel12 />,
-    [IconEnum.Carousel16]: <IconCarousel16 />,
-    [IconEnum.Carousel20]: <IconCarousel20 />,
-    [IconEnum.Carousel24]: <IconCarousel24 />,
-    [IconEnum.Carousel32]: <IconCarousel32 />,
-    [IconEnum.ChartBars]: <IconChartBars />,
-    [IconEnum.ChartBars12]: <IconChartBars12 />,
-    [IconEnum.ChartBars16]: <IconChartBars16 />,
-    [IconEnum.ChartBars20]: <IconChartBars20 />,
-    [IconEnum.ChartBars24]: <IconChartBars24 />,
-    [IconEnum.ChartBars32]: <IconChartBars32 />,
-    [IconEnum.ChartLine]: <IconChartLine />,
-    [IconEnum.ChartLine12]: <IconChartLine12 />,
-    [IconEnum.ChartLine16]: <IconChartLine16 />,
-    [IconEnum.ChartLine20]: <IconChartLine20 />,
-    [IconEnum.ChartLine24]: <IconChartLine24 />,
-    [IconEnum.ChartLine32]: <IconChartLine32 />,
-    [IconEnum.ChartPie]: <IconChartPie />,
-    [IconEnum.ChartPie12]: <IconChartPie12 />,
-    [IconEnum.ChartPie16]: <IconChartPie16 />,
-    [IconEnum.ChartPie20]: <IconChartPie20 />,
-    [IconEnum.ChartPie24]: <IconChartPie24 />,
-    [IconEnum.ChartPie32]: <IconChartPie32 />,
-    [IconEnum.CheckMark]: <IconCheckMark />,
-    [IconEnum.CheckMark12]: <IconCheckMark12 />,
-    [IconEnum.CheckMark16]: <IconCheckMark16 />,
-    [IconEnum.CheckMark20]: <IconCheckMark20 />,
-    [IconEnum.CheckMark24]: <IconCheckMark24 />,
-    [IconEnum.CheckMark32]: <IconCheckMark32 />,
-    [IconEnum.CheckMarkCircle]: <IconCheckMarkCircle />,
-    [IconEnum.CheckMarkCircle12]: <IconCheckMarkCircle12 />,
-    [IconEnum.CheckMarkCircle12Filled]: <IconCheckMarkCircle12Filled />,
-    [IconEnum.CheckMarkCircle16]: <IconCheckMarkCircle16 />,
-    [IconEnum.CheckMarkCircle16Filled]: <IconCheckMarkCircle16Filled />,
-    [IconEnum.CheckMarkCircle20]: <IconCheckMarkCircle20 />,
-    [IconEnum.CheckMarkCircle20Filled]: <IconCheckMarkCircle20Filled />,
-    [IconEnum.CheckMarkCircle24]: <IconCheckMarkCircle24 />,
-    [IconEnum.CheckMarkCircle24Filled]: <IconCheckMarkCircle24Filled />,
-    [IconEnum.CheckMarkCircle32]: <IconCheckMarkCircle32 />,
-    [IconEnum.CheckMarkCircle32Filled]: <IconCheckMarkCircle32Filled />,
-    [IconEnum.Checkbox]: <IconCheckbox />,
-    [IconEnum.Checkbox12]: <IconCheckbox12 />,
-    [IconEnum.Checkbox16]: <IconCheckbox16 />,
-    [IconEnum.Checkbox20]: <IconCheckbox20 />,
-    [IconEnum.Checkbox24]: <IconCheckbox24 />,
-    [IconEnum.Checkbox32]: <IconCheckbox32 />,
-    [IconEnum.CheckmarkCircleBox]: <IconCheckmarkCircleBox />,
-    [IconEnum.CheckmarkCircleBox12]: <IconCheckmarkCircleBox12 />,
-    [IconEnum.CheckmarkCircleBox16]: <IconCheckmarkCircleBox16 />,
-    [IconEnum.CheckmarkCircleBox20]: <IconCheckmarkCircleBox20 />,
-    [IconEnum.CheckmarkCircleBox24]: <IconCheckmarkCircleBox24 />,
-    [IconEnum.CheckmarkCircleBox32]: <IconCheckmarkCircleBox32 />,
-    [IconEnum.Circle]: <IconCircle />,
-    [IconEnum.Circle12]: <IconCircle12 />,
-    [IconEnum.Circle16]: <IconCircle16 />,
-    [IconEnum.Circle20]: <IconCircle20 />,
-    [IconEnum.Circle24]: <IconCircle24 />,
-    [IconEnum.Circle32]: <IconCircle32 />,
-    [IconEnum.Clipboard]: <IconClipboard />,
-    [IconEnum.Clipboard12]: <IconClipboard12 />,
-    [IconEnum.Clipboard16]: <IconClipboard16 />,
-    [IconEnum.Clipboard20]: <IconClipboard20 />,
-    [IconEnum.Clipboard24]: <IconClipboard24 />,
-    [IconEnum.Clipboard32]: <IconClipboard32 />,
-    [IconEnum.ClipboardAt]: <IconClipboardAt />,
-    [IconEnum.ClipboardAt12]: <IconClipboardAt12 />,
-    [IconEnum.ClipboardAt16]: <IconClipboardAt16 />,
-    [IconEnum.ClipboardAt20]: <IconClipboardAt20 />,
-    [IconEnum.ClipboardAt24]: <IconClipboardAt24 />,
-    [IconEnum.ClipboardAt32]: <IconClipboardAt32 />,
-    [IconEnum.Clock]: <IconClock />,
-    [IconEnum.Clock12]: <IconClock12 />,
-    [IconEnum.Clock16]: <IconClock16 />,
-    [IconEnum.Clock20]: <IconClock20 />,
-    [IconEnum.Clock24]: <IconClock24 />,
-    [IconEnum.Clock32]: <IconClock32 />,
-    [IconEnum.ClockAlternative]: <IconClockAlternative />,
-    [IconEnum.ClockAlternative1]: <IconClockAlternative1 />,
-    [IconEnum.ClockAlternative112]: <IconClockAlternative112 />,
-    [IconEnum.ClockAlternative116]: <IconClockAlternative116 />,
-    [IconEnum.ClockAlternative12]: <IconClockAlternative12 />,
-    [IconEnum.ClockAlternative120]: <IconClockAlternative120 />,
-    [IconEnum.ClockAlternative124]: <IconClockAlternative124 />,
-    [IconEnum.ClockAlternative132]: <IconClockAlternative132 />,
-    [IconEnum.ClockAlternative16]: <IconClockAlternative16 />,
-    [IconEnum.ClockAlternative20]: <IconClockAlternative20 />,
-    [IconEnum.ClockAlternative24]: <IconClockAlternative24 />,
-    [IconEnum.ClockAlternative32]: <IconClockAlternative32 />,
-    [IconEnum.ClockList]: <IconClockList />,
-    [IconEnum.ClockList12]: <IconClockList12 />,
-    [IconEnum.ClockList16]: <IconClockList16 />,
-    [IconEnum.ClockList20]: <IconClockList20 />,
-    [IconEnum.ClockList24]: <IconClockList24 />,
-    [IconEnum.ClockList32]: <IconClockList32 />,
-    [IconEnum.CloudArrowDown]: <IconCloudArrowDown />,
-    [IconEnum.CloudArrowDown12]: <IconCloudArrowDown12 />,
-    [IconEnum.CloudArrowDown16]: <IconCloudArrowDown16 />,
-    [IconEnum.CloudArrowDown20]: <IconCloudArrowDown20 />,
-    [IconEnum.CloudArrowDown24]: <IconCloudArrowDown24 />,
-    [IconEnum.CloudArrowDown32]: <IconCloudArrowDown32 />,
-    [IconEnum.CloudArrowUp]: <IconCloudArrowUp />,
-    [IconEnum.CloudArrowUp12]: <IconCloudArrowUp12 />,
-    [IconEnum.CloudArrowUp16]: <IconCloudArrowUp16 />,
-    [IconEnum.CloudArrowUp20]: <IconCloudArrowUp20 />,
-    [IconEnum.CloudArrowUp24]: <IconCloudArrowUp24 />,
-    [IconEnum.CloudArrowUp32]: <IconCloudArrowUp32 />,
-    [IconEnum.Code]: <IconCode />,
-    [IconEnum.Code12]: <IconCode12 />,
-    [IconEnum.Code16]: <IconCode16 />,
-    [IconEnum.Code20]: <IconCode20 />,
-    [IconEnum.Code24]: <IconCode24 />,
-    [IconEnum.Code32]: <IconCode32 />,
-    [IconEnum.CodeBox]: <IconCodeBox />,
-    [IconEnum.CodeBox12]: <IconCodeBox12 />,
-    [IconEnum.CodeBox16]: <IconCodeBox16 />,
-    [IconEnum.CodeBox20]: <IconCodeBox20 />,
-    [IconEnum.CodeBox24]: <IconCodeBox24 />,
-    [IconEnum.CodeBox32]: <IconCodeBox32 />,
-    [IconEnum.CodeDocument]: <IconCodeDocument />,
-    [IconEnum.CodeDocument12]: <IconCodeDocument12 />,
-    [IconEnum.CodeDocument16]: <IconCodeDocument16 />,
-    [IconEnum.CodeDocument20]: <IconCodeDocument20 />,
-    [IconEnum.CodeDocument24]: <IconCodeDocument24 />,
-    [IconEnum.CodeDocument32]: <IconCodeDocument32 />,
-    [IconEnum.CodeFrame]: <IconCodeFrame />,
-    [IconEnum.CodeFrame12]: <IconCodeFrame12 />,
-    [IconEnum.CodeFrame16]: <IconCodeFrame16 />,
-    [IconEnum.CodeFrame20]: <IconCodeFrame20 />,
-    [IconEnum.CodeFrame24]: <IconCodeFrame24 />,
-    [IconEnum.CodeFrame32]: <IconCodeFrame32 />,
-    [IconEnum.Cog]: <IconCog />,
-    [IconEnum.Cog12]: <IconCog12 />,
-    [IconEnum.Cog16]: <IconCog16 />,
-    [IconEnum.Cog20]: <IconCog20 />,
-    [IconEnum.Cog24]: <IconCog24 />,
-    [IconEnum.Cog32]: <IconCog32 />,
-    [IconEnum.ColorFan]: <IconColorFan />,
-    [IconEnum.ColorFan12]: <IconColorFan12 />,
-    [IconEnum.ColorFan16]: <IconColorFan16 />,
-    [IconEnum.ColorFan20]: <IconColorFan20 />,
-    [IconEnum.ColorFan24]: <IconColorFan24 />,
-    [IconEnum.ColorFan32]: <IconColorFan32 />,
-    [IconEnum.CommandLineBox]: <IconCommandLineBox />,
-    [IconEnum.CommandLineBox12]: <IconCommandLineBox12 />,
-    [IconEnum.CommandLineBox16]: <IconCommandLineBox16 />,
-    [IconEnum.CommandLineBox20]: <IconCommandLineBox20 />,
-    [IconEnum.CommandLineBox24]: <IconCommandLineBox24 />,
-    [IconEnum.CommandLineBox32]: <IconCommandLineBox32 />,
-    [IconEnum.ContentBlock]: <IconContentBlock />,
-    [IconEnum.ContentBlock12]: <IconContentBlock12 />,
-    [IconEnum.ContentBlock16]: <IconContentBlock16 />,
-    [IconEnum.ContentBlock20]: <IconContentBlock20 />,
-    [IconEnum.ContentBlock24]: <IconContentBlock24 />,
-    [IconEnum.ContentBlock32]: <IconContentBlock32 />,
-    [IconEnum.Controller]: <IconController />,
-    [IconEnum.Controller12]: <IconController12 />,
-    [IconEnum.Controller16]: <IconController16 />,
-    [IconEnum.Controller20]: <IconController20 />,
-    [IconEnum.Controller24]: <IconController24 />,
-    [IconEnum.Controller32]: <IconController32 />,
-    [IconEnum.Country]: <IconCountry />,
-    [IconEnum.Country12]: <IconCountry12 />,
-    [IconEnum.Country16]: <IconCountry16 />,
-    [IconEnum.Country20]: <IconCountry20 />,
-    [IconEnum.Country24]: <IconCountry24 />,
-    [IconEnum.Country32]: <IconCountry32 />,
-    [IconEnum.Crop]: <IconCrop />,
-    [IconEnum.Crop12]: <IconCrop12 />,
-    [IconEnum.Crop16]: <IconCrop16 />,
-    [IconEnum.Crop20]: <IconCrop20 />,
-    [IconEnum.Crop24]: <IconCrop24 />,
-    [IconEnum.Crop32]: <IconCrop32 />,
-    [IconEnum.Cross]: <IconCross />,
-    [IconEnum.Cross12]: <IconCross12 />,
-    [IconEnum.Cross16]: <IconCross16 />,
-    [IconEnum.Cross20]: <IconCross20 />,
-    [IconEnum.Cross24]: <IconCross24 />,
-    [IconEnum.Cross32]: <IconCross32 />,
-    [IconEnum.CrossCircle]: <IconCrossCircle />,
-    [IconEnum.CrossCircle12]: <IconCrossCircle12 />,
-    [IconEnum.CrossCircle12Filled]: <IconCrossCircle12Filled />,
-    [IconEnum.CrossCircle16]: <IconCrossCircle16 />,
-    [IconEnum.CrossCircle16Filled]: <IconCrossCircle16Filled />,
-    [IconEnum.CrossCircle20]: <IconCrossCircle20 />,
-    [IconEnum.CrossCircle20Filled]: <IconCrossCircle20Filled />,
-    [IconEnum.CrossCircle24]: <IconCrossCircle24 />,
-    [IconEnum.CrossCircle24Filled]: <IconCrossCircle24Filled />,
-    [IconEnum.CrossCircle32]: <IconCrossCircle32 />,
-    [IconEnum.CrossCircle32Filled]: <IconCrossCircle32Filled />,
-    [IconEnum.CrossCircleBox]: <IconCrossCircleBox />,
-    [IconEnum.CrossCircleBox12]: <IconCrossCircleBox12 />,
-    [IconEnum.CrossCircleBox16]: <IconCrossCircleBox16 />,
-    [IconEnum.CrossCircleBox20]: <IconCrossCircleBox20 />,
-    [IconEnum.CrossCircleBox24]: <IconCrossCircleBox24 />,
-    [IconEnum.CrossCircleBox32]: <IconCrossCircleBox32 />,
-    [IconEnum.CursorClick]: <IconCursorClick />,
-    [IconEnum.CursorClick12]: <IconCursorClick12 />,
-    [IconEnum.CursorClick16]: <IconCursorClick16 />,
-    [IconEnum.CursorClick20]: <IconCursorClick20 />,
-    [IconEnum.CursorClick24]: <IconCursorClick24 />,
-    [IconEnum.CursorClick32]: <IconCursorClick32 />,
-    [IconEnum.DiagramTree]: <IconDiagramTree />,
-    [IconEnum.DiagramTree12]: <IconDiagramTree12 />,
-    [IconEnum.DiagramTree16]: <IconDiagramTree16 />,
-    [IconEnum.DiagramTree20]: <IconDiagramTree20 />,
-    [IconEnum.DiagramTree24]: <IconDiagramTree24 />,
-    [IconEnum.DiagramTree32]: <IconDiagramTree32 />,
-    [IconEnum.DividerBlank]: <IconDividerBlank />,
-    [IconEnum.DividerBlank12]: <IconDividerBlank12 />,
-    [IconEnum.DividerBlank16]: <IconDividerBlank16 />,
-    [IconEnum.DividerBlank20]: <IconDividerBlank20 />,
-    [IconEnum.DividerBlank24]: <IconDividerBlank24 />,
-    [IconEnum.DividerBlank32]: <IconDividerBlank32 />,
-    [IconEnum.DividerDashed]: <IconDividerDashed />,
-    [IconEnum.DividerDashed12]: <IconDividerDashed12 />,
-    [IconEnum.DividerDashed16]: <IconDividerDashed16 />,
-    [IconEnum.DividerDashed20]: <IconDividerDashed20 />,
-    [IconEnum.DividerDashed24]: <IconDividerDashed24 />,
-    [IconEnum.DividerDashed32]: <IconDividerDashed32 />,
-    [IconEnum.DividerDotted]: <IconDividerDotted />,
-    [IconEnum.DividerDotted12]: <IconDividerDotted12 />,
-    [IconEnum.DividerDotted16]: <IconDividerDotted16 />,
-    [IconEnum.DividerDotted20]: <IconDividerDotted20 />,
-    [IconEnum.DividerDotted24]: <IconDividerDotted24 />,
-    [IconEnum.DividerDotted32]: <IconDividerDotted32 />,
-    [IconEnum.DividerSolid]: <IconDividerSolid />,
-    [IconEnum.DividerSolid12]: <IconDividerSolid12 />,
-    [IconEnum.DividerSolid16]: <IconDividerSolid16 />,
-    [IconEnum.DividerSolid20]: <IconDividerSolid20 />,
-    [IconEnum.DividerSolid24]: <IconDividerSolid24 />,
-    [IconEnum.DividerSolid32]: <IconDividerSolid32 />,
-    [IconEnum.DividerSolidInset]: <IconDividerSolidInset />,
-    [IconEnum.DividerSolidInset12]: <IconDividerSolidInset12 />,
-    [IconEnum.DividerSolidInset16]: <IconDividerSolidInset16 />,
-    [IconEnum.DividerSolidInset20]: <IconDividerSolidInset20 />,
-    [IconEnum.DividerSolidInset24]: <IconDividerSolidInset24 />,
-    [IconEnum.DividerSolidInset32]: <IconDividerSolidInset32 />,
-    [IconEnum.DoAndDont]: <IconDoAndDont />,
-    [IconEnum.DoAndDont12]: <IconDoAndDont12 />,
-    [IconEnum.DoAndDont16]: <IconDoAndDont16 />,
-    [IconEnum.DoAndDont20]: <IconDoAndDont20 />,
-    [IconEnum.DoAndDont24]: <IconDoAndDont24 />,
-    [IconEnum.DoAndDont32]: <IconDoAndDont32 />,
-    [IconEnum.DoAndDontBox]: <IconDoAndDontBox />,
-    [IconEnum.DoAndDontBox12]: <IconDoAndDontBox12 />,
-    [IconEnum.DoAndDontBox16]: <IconDoAndDontBox16 />,
-    [IconEnum.DoAndDontBox20]: <IconDoAndDontBox20 />,
-    [IconEnum.DoAndDontBox24]: <IconDoAndDontBox24 />,
-    [IconEnum.DoAndDontBox32]: <IconDoAndDontBox32 />,
-    [IconEnum.Document]: <IconDocument />,
-    [IconEnum.Document12]: <IconDocument12 />,
-    [IconEnum.Document16]: <IconDocument16 />,
-    [IconEnum.Document20]: <IconDocument20 />,
-    [IconEnum.Document24]: <IconDocument24 />,
-    [IconEnum.Document32]: <IconDocument32 />,
-    [IconEnum.DocumentArrowRight]: <IconDocumentArrowRight />,
-    [IconEnum.DocumentArrowRight12]: <IconDocumentArrowRight12 />,
-    [IconEnum.DocumentArrowRight16]: <IconDocumentArrowRight16 />,
-    [IconEnum.DocumentArrowRight20]: <IconDocumentArrowRight20 />,
-    [IconEnum.DocumentArrowRight24]: <IconDocumentArrowRight24 />,
-    [IconEnum.DocumentArrowRight32]: <IconDocumentArrowRight32 />,
-    [IconEnum.DocumentBadge]: <IconDocumentBadge />,
-    [IconEnum.DocumentBadge12]: <IconDocumentBadge12 />,
-    [IconEnum.DocumentBadge16]: <IconDocumentBadge16 />,
-    [IconEnum.DocumentBadge20]: <IconDocumentBadge20 />,
-    [IconEnum.DocumentBadge24]: <IconDocumentBadge24 />,
-    [IconEnum.DocumentBadge32]: <IconDocumentBadge32 />,
-    [IconEnum.DocumentCheck]: <IconDocumentCheck />,
-    [IconEnum.DocumentCheck12]: <IconDocumentCheck12 />,
-    [IconEnum.DocumentCheck16]: <IconDocumentCheck16 />,
-    [IconEnum.DocumentCheck20]: <IconDocumentCheck20 />,
-    [IconEnum.DocumentCheck24]: <IconDocumentCheck24 />,
-    [IconEnum.DocumentCheck32]: <IconDocumentCheck32 />,
-    [IconEnum.DocumentCode]: <IconDocumentCode />,
-    [IconEnum.DocumentCode12]: <IconDocumentCode12 />,
-    [IconEnum.DocumentCode16]: <IconDocumentCode16 />,
-    [IconEnum.DocumentCode20]: <IconDocumentCode20 />,
-    [IconEnum.DocumentCode24]: <IconDocumentCode24 />,
-    [IconEnum.DocumentCode32]: <IconDocumentCode32 />,
-    [IconEnum.DocumentList]: <IconDocumentList />,
-    [IconEnum.DocumentList12]: <IconDocumentList12 />,
-    [IconEnum.DocumentList16]: <IconDocumentList16 />,
-    [IconEnum.DocumentList20]: <IconDocumentList20 />,
-    [IconEnum.DocumentList24]: <IconDocumentList24 />,
-    [IconEnum.DocumentList32]: <IconDocumentList32 />,
-    [IconEnum.DocumentStack]: <IconDocumentStack />,
-    [IconEnum.DocumentStack12]: <IconDocumentStack12 />,
-    [IconEnum.DocumentStack16]: <IconDocumentStack16 />,
-    [IconEnum.DocumentStack20]: <IconDocumentStack20 />,
-    [IconEnum.DocumentStack24]: <IconDocumentStack24 />,
-    [IconEnum.DocumentStack32]: <IconDocumentStack32 />,
-    [IconEnum.DocumentText]: <IconDocumentText />,
-    [IconEnum.DocumentText12]: <IconDocumentText12 />,
-    [IconEnum.DocumentText16]: <IconDocumentText16 />,
-    [IconEnum.DocumentText20]: <IconDocumentText20 />,
-    [IconEnum.DocumentText24]: <IconDocumentText24 />,
-    [IconEnum.DocumentText32]: <IconDocumentText32 />,
-    [IconEnum.DotsHorizontal]: <IconDotsHorizontal />,
-    [IconEnum.DotsHorizontal12]: <IconDotsHorizontal12 />,
-    [IconEnum.DotsHorizontal16]: <IconDotsHorizontal16 />,
-    [IconEnum.DotsHorizontal20]: <IconDotsHorizontal20 />,
-    [IconEnum.DotsHorizontal24]: <IconDotsHorizontal24 />,
-    [IconEnum.DotsHorizontal32]: <IconDotsHorizontal32 />,
-    [IconEnum.DotsVertical]: <IconDotsVertical />,
-    [IconEnum.DotsVertical12]: <IconDotsVertical12 />,
-    [IconEnum.DotsVertical16]: <IconDotsVertical16 />,
-    [IconEnum.DotsVertical20]: <IconDotsVertical20 />,
-    [IconEnum.DotsVertical24]: <IconDotsVertical24 />,
-    [IconEnum.DotsVertical32]: <IconDotsVertical32 />,
-    [IconEnum.Droplet]: <IconDroplet />,
-    [IconEnum.Droplet12]: <IconDroplet12 />,
-    [IconEnum.Droplet16]: <IconDroplet16 />,
-    [IconEnum.Droplet20]: <IconDroplet20 />,
-    [IconEnum.Droplet24]: <IconDroplet24 />,
-    [IconEnum.Droplet32]: <IconDroplet32 />,
-    [IconEnum.Drops]: <IconDrops />,
-    [IconEnum.Drops12]: <IconDrops12 />,
-    [IconEnum.Drops16]: <IconDrops16 />,
-    [IconEnum.Drops20]: <IconDrops20 />,
-    [IconEnum.Drops24]: <IconDrops24 />,
-    [IconEnum.Drops32]: <IconDrops32 />,
-    [IconEnum.Envelope]: <IconEnvelope />,
-    [IconEnum.Envelope12]: <IconEnvelope12 />,
-    [IconEnum.Envelope16]: <IconEnvelope16 />,
-    [IconEnum.Envelope20]: <IconEnvelope20 />,
-    [IconEnum.Envelope24]: <IconEnvelope24 />,
-    [IconEnum.Envelope32]: <IconEnvelope32 />,
-    [IconEnum.Eraser]: <IconEraser />,
-    [IconEnum.Eraser12]: <IconEraser12 />,
-    [IconEnum.Eraser16]: <IconEraser16 />,
-    [IconEnum.Eraser20]: <IconEraser20 />,
-    [IconEnum.Eraser24]: <IconEraser24 />,
-    [IconEnum.Eraser32]: <IconEraser32 />,
-    [IconEnum.ExclamationMark]: <IconExclamationMark />,
-    [IconEnum.ExclamationMark12]: <IconExclamationMark12 />,
-    [IconEnum.ExclamationMark16]: <IconExclamationMark16 />,
-    [IconEnum.ExclamationMark20]: <IconExclamationMark20 />,
-    [IconEnum.ExclamationMark24]: <IconExclamationMark24 />,
-    [IconEnum.ExclamationMark32]: <IconExclamationMark32 />,
-    [IconEnum.ExclamationMarkCircle]: <IconExclamationMarkCircle />,
-    [IconEnum.ExclamationMarkCircle12]: <IconExclamationMarkCircle12 />,
-    [IconEnum.ExclamationMarkCircle12Filled]: <IconExclamationMarkCircle12Filled />,
-    [IconEnum.ExclamationMarkCircle16]: <IconExclamationMarkCircle16 />,
-    [IconEnum.ExclamationMarkCircle16Filled]: <IconExclamationMarkCircle16Filled />,
-    [IconEnum.ExclamationMarkCircle20]: <IconExclamationMarkCircle20 />,
-    [IconEnum.ExclamationMarkCircle20Filled]: <IconExclamationMarkCircle20Filled />,
-    [IconEnum.ExclamationMarkCircle24]: <IconExclamationMarkCircle24 />,
-    [IconEnum.ExclamationMarkCircle24Filled]: <IconExclamationMarkCircle24Filled />,
-    [IconEnum.ExclamationMarkCircle32]: <IconExclamationMarkCircle32 />,
-    [IconEnum.ExclamationMarkCircle32Filled]: <IconExclamationMarkCircle32Filled />,
-    [IconEnum.ExclamationMarkTriangle]: <IconExclamationMarkTriangle />,
-    [IconEnum.ExclamationMarkTriangle12]: <IconExclamationMarkTriangle12 />,
-    [IconEnum.ExclamationMarkTriangle16]: <IconExclamationMarkTriangle16 />,
-    [IconEnum.ExclamationMarkTriangle20]: <IconExclamationMarkTriangle20 />,
-    [IconEnum.ExclamationMarkTriangle24]: <IconExclamationMarkTriangle24 />,
-    [IconEnum.ExclamationMarkTriangle32]: <IconExclamationMarkTriangle32 />,
-    [IconEnum.Eye]: <IconEye />,
-    [IconEnum.Eye12]: <IconEye12 />,
-    [IconEnum.Eye16]: <IconEye16 />,
-    [IconEnum.Eye20]: <IconEye20 />,
-    [IconEnum.Eye24]: <IconEye24 />,
-    [IconEnum.Eye32]: <IconEye32 />,
-    [IconEnum.EyeOff]: <IconEyeOff />,
-    [IconEnum.EyeOff12]: <IconEyeOff12 />,
-    [IconEnum.EyeOff16]: <IconEyeOff16 />,
-    [IconEnum.EyeOff20]: <IconEyeOff20 />,
-    [IconEnum.EyeOff24]: <IconEyeOff24 />,
-    [IconEnum.EyeOff32]: <IconEyeOff32 />,
-    [IconEnum.FaceExtraHappy]: <IconFaceExtraHappy />,
-    [IconEnum.FaceExtraHappy12]: <IconFaceExtraHappy12 />,
-    [IconEnum.FaceExtraHappy16]: <IconFaceExtraHappy16 />,
-    [IconEnum.FaceExtraHappy20]: <IconFaceExtraHappy20 />,
-    [IconEnum.FaceExtraHappy24]: <IconFaceExtraHappy24 />,
-    [IconEnum.FaceExtraHappy32]: <IconFaceExtraHappy32 />,
-    [IconEnum.FaceHappy]: <IconFaceHappy />,
-    [IconEnum.FaceHappy12]: <IconFaceHappy12 />,
-    [IconEnum.FaceHappy12Filled]: <IconFaceHappy12Filled />,
-    [IconEnum.FaceHappy16]: <IconFaceHappy16 />,
-    [IconEnum.FaceHappy16Filled]: <IconFaceHappy16Filled />,
-    [IconEnum.FaceHappy20]: <IconFaceHappy20 />,
-    [IconEnum.FaceHappy20Filled]: <IconFaceHappy20Filled />,
-    [IconEnum.FaceHappy24]: <IconFaceHappy24 />,
-    [IconEnum.FaceHappy24Filled]: <IconFaceHappy24Filled />,
-    [IconEnum.FaceHappy32]: <IconFaceHappy32 />,
-    [IconEnum.FaceHappy32Filled]: <IconFaceHappy32Filled />,
-    [IconEnum.FaceNeutral]: <IconFaceNeutral />,
-    [IconEnum.FaceNeutral12]: <IconFaceNeutral12 />,
-    [IconEnum.FaceNeutral12Filled]: <IconFaceNeutral12Filled />,
-    [IconEnum.FaceNeutral16]: <IconFaceNeutral16 />,
-    [IconEnum.FaceNeutral16Filled]: <IconFaceNeutral16Filled />,
-    [IconEnum.FaceNeutral20]: <IconFaceNeutral20 />,
-    [IconEnum.FaceNeutral20Filled]: <IconFaceNeutral20Filled />,
-    [IconEnum.FaceNeutral24]: <IconFaceNeutral24 />,
-    [IconEnum.FaceNeutral24Filled]: <IconFaceNeutral24Filled />,
-    [IconEnum.FaceNeutral32]: <IconFaceNeutral32 />,
-    [IconEnum.FaceNeutral32Filled]: <IconFaceNeutral32Filled />,
-    [IconEnum.FaceSad]: <IconFaceSad />,
-    [IconEnum.FaceSad12]: <IconFaceSad12 />,
-    [IconEnum.FaceSad12Filled]: <IconFaceSad12Filled />,
-    [IconEnum.FaceSad16]: <IconFaceSad16 />,
-    [IconEnum.FaceSad16Filled]: <IconFaceSad16Filled />,
-    [IconEnum.FaceSad20]: <IconFaceSad20 />,
-    [IconEnum.FaceSad20Filled]: <IconFaceSad20Filled />,
-    [IconEnum.FaceSad24]: <IconFaceSad24 />,
-    [IconEnum.FaceSad24Filled]: <IconFaceSad24Filled />,
-    [IconEnum.FaceSad32]: <IconFaceSad32 />,
-    [IconEnum.FaceSad32Filled]: <IconFaceSad32Filled />,
-    [IconEnum.FixedHeight]: <IconFixedHeight />,
-    [IconEnum.FixedHeight12]: <IconFixedHeight12 />,
-    [IconEnum.FixedHeight16]: <IconFixedHeight16 />,
-    [IconEnum.FixedHeight20]: <IconFixedHeight20 />,
-    [IconEnum.FixedHeight24]: <IconFixedHeight24 />,
-    [IconEnum.FixedHeight32]: <IconFixedHeight32 />,
-    [IconEnum.FixedWidth]: <IconFixedWidth />,
-    [IconEnum.FixedWidth12]: <IconFixedWidth12 />,
-    [IconEnum.FixedWidth16]: <IconFixedWidth16 />,
-    [IconEnum.FixedWidth20]: <IconFixedWidth20 />,
-    [IconEnum.FixedWidth24]: <IconFixedWidth24 />,
-    [IconEnum.FixedWidth32]: <IconFixedWidth32 />,
-    [IconEnum.Flag]: <IconFlag />,
-    [IconEnum.Flag12]: <IconFlag12 />,
-    [IconEnum.Flag16]: <IconFlag16 />,
-    [IconEnum.Flag20]: <IconFlag20 />,
-    [IconEnum.Flag24]: <IconFlag24 />,
-    [IconEnum.Flag32]: <IconFlag32 />,
-    [IconEnum.FocalPoint]: <IconFocalPoint />,
-    [IconEnum.FocalPoint12]: <IconFocalPoint12 />,
-    [IconEnum.FocalPoint16]: <IconFocalPoint16 />,
-    [IconEnum.FocalPoint20]: <IconFocalPoint20 />,
-    [IconEnum.FocalPoint24]: <IconFocalPoint24 />,
-    [IconEnum.FocalPoint32]: <IconFocalPoint32 />,
-    [IconEnum.Folder]: <IconFolder />,
-    [IconEnum.Folder12]: <IconFolder12 />,
-    [IconEnum.Folder16]: <IconFolder16 />,
-    [IconEnum.Folder20]: <IconFolder20 />,
-    [IconEnum.Folder24]: <IconFolder24 />,
-    [IconEnum.Folder32]: <IconFolder32 />,
-    [IconEnum.FolderArrowUp]: <IconFolderArrowUp />,
-    [IconEnum.FolderArrowUp12]: <IconFolderArrowUp12 />,
-    [IconEnum.FolderArrowUp16]: <IconFolderArrowUp16 />,
-    [IconEnum.FolderArrowUp20]: <IconFolderArrowUp20 />,
-    [IconEnum.FolderArrowUp24]: <IconFolderArrowUp24 />,
-    [IconEnum.FolderArrowUp32]: <IconFolderArrowUp32 />,
-    [IconEnum.Footer]: <IconFooter />,
-    [IconEnum.Footer12]: <IconFooter12 />,
-    [IconEnum.Footer16]: <IconFooter16 />,
-    [IconEnum.Footer20]: <IconFooter20 />,
-    [IconEnum.Footer24]: <IconFooter24 />,
-    [IconEnum.Footer32]: <IconFooter32 />,
-    [IconEnum.Funnel]: <IconFunnel />,
-    [IconEnum.Funnel12]: <IconFunnel12 />,
-    [IconEnum.Funnel16]: <IconFunnel16 />,
-    [IconEnum.Funnel20]: <IconFunnel20 />,
-    [IconEnum.Funnel24]: <IconFunnel24 />,
-    [IconEnum.Funnel32]: <IconFunnel32 />,
-    [IconEnum.GrabHandle]: <IconGrabHandle />,
-    [IconEnum.GrabHandle12]: <IconGrabHandle12 />,
-    [IconEnum.GrabHandle16]: <IconGrabHandle16 />,
-    [IconEnum.GrabHandle20]: <IconGrabHandle20 />,
-    [IconEnum.GrabHandle24]: <IconGrabHandle24 />,
-    [IconEnum.GrabHandle32]: <IconGrabHandle32 />,
-    [IconEnum.Gradient]: <IconGradient />,
-    [IconEnum.Gradient12]: <IconGradient12 />,
-    [IconEnum.Gradient16]: <IconGradient16 />,
-    [IconEnum.Gradient20]: <IconGradient20 />,
-    [IconEnum.Gradient24]: <IconGradient24 />,
-    [IconEnum.Gradient32]: <IconGradient32 />,
-    [IconEnum.GridRegular]: <IconGridRegular />,
-    [IconEnum.GridRegular12]: <IconGridRegular12 />,
-    [IconEnum.GridRegular12Filled]: <IconGridRegular12Filled />,
-    [IconEnum.GridRegular16]: <IconGridRegular16 />,
-    [IconEnum.GridRegular16Filled]: <IconGridRegular16Filled />,
-    [IconEnum.GridRegular20]: <IconGridRegular20 />,
-    [IconEnum.GridRegular20Filled]: <IconGridRegular20Filled />,
-    [IconEnum.GridRegular24]: <IconGridRegular24 />,
-    [IconEnum.GridRegular24Filled]: <IconGridRegular24Filled />,
-    [IconEnum.GridRegular32]: <IconGridRegular32 />,
-    [IconEnum.GridRegular32Filled]: <IconGridRegular32Filled />,
-    [IconEnum.H]: <IconH />,
-    [IconEnum.H12]: <IconH12 />,
-    [IconEnum.H16]: <IconH16 />,
-    [IconEnum.H20]: <IconH20 />,
-    [IconEnum.H24]: <IconH24 />,
-    [IconEnum.H32]: <IconH32 />,
-    [IconEnum.Heading]: <IconHeading />,
-    [IconEnum.Heading12]: <IconHeading12 />,
-    [IconEnum.Heading16]: <IconHeading16 />,
-    [IconEnum.Heading20]: <IconHeading20 />,
-    [IconEnum.Heading24]: <IconHeading24 />,
-    [IconEnum.Heading32]: <IconHeading32 />,
-    [IconEnum.HeartCircle]: <IconHeartCircle />,
-    [IconEnum.HeartCircle12]: <IconHeartCircle12 />,
-    [IconEnum.HeartCircle16]: <IconHeartCircle16 />,
-    [IconEnum.HeartCircle20]: <IconHeartCircle20 />,
-    [IconEnum.HeartCircle24]: <IconHeartCircle24 />,
-    [IconEnum.HeartCircle32]: <IconHeartCircle32 />,
-    [IconEnum.HeartCircleStack]: <IconHeartCircleStack />,
-    [IconEnum.HeartCircleStack12]: <IconHeartCircleStack12 />,
-    [IconEnum.HeartCircleStack16]: <IconHeartCircleStack16 />,
-    [IconEnum.HeartCircleStack20]: <IconHeartCircleStack20 />,
-    [IconEnum.HeartCircleStack24]: <IconHeartCircleStack24 />,
-    [IconEnum.HeartCircleStack32]: <IconHeartCircleStack32 />,
-    [IconEnum.Highlighter]: <IconHighlighter />,
-    [IconEnum.Highlighter12]: <IconHighlighter12 />,
-    [IconEnum.Highlighter16]: <IconHighlighter16 />,
-    [IconEnum.Highlighter20]: <IconHighlighter20 />,
-    [IconEnum.Highlighter24]: <IconHighlighter24 />,
-    [IconEnum.Highlighter32]: <IconHighlighter32 />,
-    [IconEnum.Home]: <IconHome />,
-    [IconEnum.Home12]: <IconHome12 />,
-    [IconEnum.Home16]: <IconHome16 />,
-    [IconEnum.Home20]: <IconHome20 />,
-    [IconEnum.Home24]: <IconHome24 />,
-    [IconEnum.Home32]: <IconHome32 />,
-    [IconEnum.Icon]: <IconIcon />,
-    [IconEnum.Icon12]: <IconIcon12 />,
-    [IconEnum.Icon16]: <IconIcon16 />,
-    [IconEnum.Icon20]: <IconIcon20 />,
-    [IconEnum.Icon24]: <IconIcon24 />,
-    [IconEnum.Icon32]: <IconIcon32 />,
-    [IconEnum.Image]: <IconImage />,
-    [IconEnum.Image12]: <IconImage12 />,
-    [IconEnum.Image16]: <IconImage16 />,
-    [IconEnum.Image20]: <IconImage20 />,
-    [IconEnum.Image24]: <IconImage24 />,
-    [IconEnum.Image32]: <IconImage32 />,
-    [IconEnum.ImageStack]: <IconImageStack />,
-    [IconEnum.ImageStack12]: <IconImageStack12 />,
-    [IconEnum.ImageStack16]: <IconImageStack16 />,
-    [IconEnum.ImageStack20]: <IconImageStack20 />,
-    [IconEnum.ImageStack24]: <IconImageStack24 />,
-    [IconEnum.ImageStack32]: <IconImageStack32 />,
-    [IconEnum.ImageStackPlus]: <IconImageStackPlus />,
-    [IconEnum.ImageStackPlus12]: <IconImageStackPlus12 />,
-    [IconEnum.ImageStackPlus16]: <IconImageStackPlus16 />,
-    [IconEnum.ImageStackPlus20]: <IconImageStackPlus20 />,
-    [IconEnum.ImageStackPlus24]: <IconImageStackPlus24 />,
-    [IconEnum.ImageStackPlus32]: <IconImageStackPlus32 />,
-    [IconEnum.ImageWithText]: <IconImageWithText />,
-    [IconEnum.ImageWithText12]: <IconImageWithText12 />,
-    [IconEnum.ImageWithText16]: <IconImageWithText16 />,
-    [IconEnum.ImageWithText20]: <IconImageWithText20 />,
-    [IconEnum.ImageWithText24]: <IconImageWithText24 />,
-    [IconEnum.ImageWithText32]: <IconImageWithText32 />,
-    [IconEnum.Info]: <IconInfo />,
-    [IconEnum.Info12]: <IconInfo12 />,
-    [IconEnum.Info16]: <IconInfo16 />,
-    [IconEnum.Info20]: <IconInfo20 />,
-    [IconEnum.Info24]: <IconInfo24 />,
-    [IconEnum.Info32]: <IconInfo32 />,
-    [IconEnum.Intercom]: <IconIntercom />,
-    [IconEnum.Intercom12]: <IconIntercom12 />,
-    [IconEnum.Intercom16]: <IconIntercom16 />,
-    [IconEnum.Intercom20]: <IconIntercom20 />,
-    [IconEnum.Intercom24]: <IconIntercom24 />,
-    [IconEnum.Intercom32]: <IconIntercom32 />,
-    [IconEnum.Johanna]: <IconJohanna />,
-    [IconEnum.Johanna12]: <IconJohanna12 />,
-    [IconEnum.Johanna16]: <IconJohanna16 />,
-    [IconEnum.Johanna20]: <IconJohanna20 />,
-    [IconEnum.Johanna24]: <IconJohanna24 />,
-    [IconEnum.Johanna32]: <IconJohanna32 />,
-    [IconEnum.Label]: <IconLabel />,
-    [IconEnum.Label12]: <IconLabel12 />,
-    [IconEnum.Label16]: <IconLabel16 />,
-    [IconEnum.Label20]: <IconLabel20 />,
-    [IconEnum.Label24]: <IconLabel24 />,
-    [IconEnum.Label32]: <IconLabel32 />,
-    [IconEnum.Language]: <IconLanguage />,
-    [IconEnum.Language12]: <IconLanguage12 />,
-    [IconEnum.Language16]: <IconLanguage16 />,
-    [IconEnum.Language20]: <IconLanguage20 />,
-    [IconEnum.Language24]: <IconLanguage24 />,
-    [IconEnum.Language32]: <IconLanguage32 />,
-    [IconEnum.Layers]: <IconLayers />,
-    [IconEnum.Layers12]: <IconLayers12 />,
-    [IconEnum.Layers16]: <IconLayers16 />,
-    [IconEnum.Layers20]: <IconLayers20 />,
-    [IconEnum.Layers24]: <IconLayers24 />,
-    [IconEnum.Layers32]: <IconLayers32 />,
-    [IconEnum.LayersSingle]: <IconLayersSingle />,
-    [IconEnum.LayersSingle12]: <IconLayersSingle12 />,
-    [IconEnum.LayersSingle16]: <IconLayersSingle16 />,
-    [IconEnum.LayersSingle20]: <IconLayersSingle20 />,
-    [IconEnum.LayersSingle24]: <IconLayersSingle24 />,
-    [IconEnum.LayersSingle32]: <IconLayersSingle32 />,
-    [IconEnum.Leaf]: <IconLeaf />,
-    [IconEnum.Leaf12]: <IconLeaf12 />,
-    [IconEnum.Leaf16]: <IconLeaf16 />,
-    [IconEnum.Leaf20]: <IconLeaf20 />,
-    [IconEnum.Leaf24]: <IconLeaf24 />,
-    [IconEnum.Leaf32]: <IconLeaf32 />,
-    [IconEnum.Lightbulb]: <IconLightbulb />,
-    [IconEnum.Lightbulb12]: <IconLightbulb12 />,
-    [IconEnum.Lightbulb16]: <IconLightbulb16 />,
-    [IconEnum.Lightbulb20]: <IconLightbulb20 />,
-    [IconEnum.Lightbulb24]: <IconLightbulb24 />,
-    [IconEnum.Lightbulb32]: <IconLightbulb32 />,
-    [IconEnum.Lightning]: <IconLightning />,
-    [IconEnum.Lightning12]: <IconLightning12 />,
-    [IconEnum.Lightning12Filled]: <IconLightning12Filled />,
-    [IconEnum.Lightning16]: <IconLightning16 />,
-    [IconEnum.Lightning16Filled]: <IconLightning16Filled />,
-    [IconEnum.Lightning20]: <IconLightning20 />,
-    [IconEnum.Lightning20Filled]: <IconLightning20Filled />,
-    [IconEnum.Lightning24]: <IconLightning24 />,
-    [IconEnum.Lightning24Filled]: <IconLightning24Filled />,
-    [IconEnum.Lightning32]: <IconLightning32 />,
-    [IconEnum.Lightning32Filled]: <IconLightning32Filled />,
-    [IconEnum.Lines]: <IconLines />,
-    [IconEnum.Lines12]: <IconLines12 />,
-    [IconEnum.Lines16]: <IconLines16 />,
-    [IconEnum.Lines20]: <IconLines20 />,
-    [IconEnum.Lines24]: <IconLines24 />,
-    [IconEnum.Lines32]: <IconLines32 />,
-    [IconEnum.LinesTriangular]: <IconLinesTriangular />,
-    [IconEnum.LinesTriangular12]: <IconLinesTriangular12 />,
-    [IconEnum.LinesTriangular16]: <IconLinesTriangular16 />,
-    [IconEnum.LinesTriangular20]: <IconLinesTriangular20 />,
-    [IconEnum.LinesTriangular24]: <IconLinesTriangular24 />,
-    [IconEnum.LinesTriangular32]: <IconLinesTriangular32 />,
-    [IconEnum.Link]: <IconLink />,
-    [IconEnum.Link12]: <IconLink12 />,
-    [IconEnum.Link16]: <IconLink16 />,
-    [IconEnum.Link20]: <IconLink20 />,
-    [IconEnum.Link24]: <IconLink24 />,
-    [IconEnum.Link32]: <IconLink32 />,
-    [IconEnum.LinkBox]: <IconLinkBox />,
-    [IconEnum.LinkBox12]: <IconLinkBox12 />,
-    [IconEnum.LinkBox16]: <IconLinkBox16 />,
-    [IconEnum.LinkBox20]: <IconLinkBox20 />,
-    [IconEnum.LinkBox24]: <IconLinkBox24 />,
-    [IconEnum.LinkBox32]: <IconLinkBox32 />,
-    [IconEnum.LinkOpen]: <IconLinkOpen />,
-    [IconEnum.LinkOpen12]: <IconLinkOpen12 />,
-    [IconEnum.LinkOpen16]: <IconLinkOpen16 />,
-    [IconEnum.LinkOpen20]: <IconLinkOpen20 />,
-    [IconEnum.LinkOpen24]: <IconLinkOpen24 />,
-    [IconEnum.LinkOpen32]: <IconLinkOpen32 />,
-    [IconEnum.ListBullet]: <IconListBullet />,
-    [IconEnum.ListBullet12]: <IconListBullet12 />,
-    [IconEnum.ListBullet16]: <IconListBullet16 />,
-    [IconEnum.ListBullet20]: <IconListBullet20 />,
-    [IconEnum.ListBullet24]: <IconListBullet24 />,
-    [IconEnum.ListBullet32]: <IconListBullet32 />,
-    [IconEnum.ListCheck]: <IconListCheck />,
-    [IconEnum.ListCheck12]: <IconListCheck12 />,
-    [IconEnum.ListCheck16]: <IconListCheck16 />,
-    [IconEnum.ListCheck20]: <IconListCheck20 />,
-    [IconEnum.ListCheck24]: <IconListCheck24 />,
-    [IconEnum.ListCheck32]: <IconListCheck32 />,
-    [IconEnum.ListIndented]: <IconListIndented />,
-    [IconEnum.ListIndented12]: <IconListIndented12 />,
-    [IconEnum.ListIndented16]: <IconListIndented16 />,
-    [IconEnum.ListIndented20]: <IconListIndented20 />,
-    [IconEnum.ListIndented24]: <IconListIndented24 />,
-    [IconEnum.ListIndented32]: <IconListIndented32 />,
-    [IconEnum.ListNumbers]: <IconListNumbers />,
-    [IconEnum.ListNumbers12]: <IconListNumbers12 />,
-    [IconEnum.ListNumbers16]: <IconListNumbers16 />,
-    [IconEnum.ListNumbers20]: <IconListNumbers20 />,
-    [IconEnum.ListNumbers24]: <IconListNumbers24 />,
-    [IconEnum.ListNumbers32]: <IconListNumbers32 />,
-    [IconEnum.LockClosed]: <IconLockClosed />,
-    [IconEnum.LockClosed12]: <IconLockClosed12 />,
-    [IconEnum.LockClosed16]: <IconLockClosed16 />,
-    [IconEnum.LockClosed20]: <IconLockClosed20 />,
-    [IconEnum.LockClosed24]: <IconLockClosed24 />,
-    [IconEnum.LockClosed32]: <IconLockClosed32 />,
-    [IconEnum.LockOpen]: <IconLockOpen />,
-    [IconEnum.LockOpen12]: <IconLockOpen12 />,
-    [IconEnum.LockOpen12Filled]: <IconLockOpen12Filled />,
-    [IconEnum.LockOpen16]: <IconLockOpen16 />,
-    [IconEnum.LockOpen16Filled]: <IconLockOpen16Filled />,
-    [IconEnum.LockOpen20]: <IconLockOpen20 />,
-    [IconEnum.LockOpen20Filled]: <IconLockOpen20Filled />,
-    [IconEnum.LockOpen24]: <IconLockOpen24 />,
-    [IconEnum.LockOpen24Filled]: <IconLockOpen24Filled />,
-    [IconEnum.LockOpen32]: <IconLockOpen32 />,
-    [IconEnum.LockOpen32Filled]: <IconLockOpen32Filled />,
-    [IconEnum.Lollipop]: <IconLollipop />,
-    [IconEnum.Lollipop12]: <IconLollipop12 />,
-    [IconEnum.Lollipop16]: <IconLollipop16 />,
-    [IconEnum.Lollipop20]: <IconLollipop20 />,
-    [IconEnum.Lollipop24]: <IconLollipop24 />,
-    [IconEnum.Lollipop32]: <IconLollipop32 />,
-    [IconEnum.Magnifier]: <IconMagnifier />,
-    [IconEnum.Magnifier12]: <IconMagnifier12 />,
-    [IconEnum.Magnifier16]: <IconMagnifier16 />,
-    [IconEnum.Magnifier20]: <IconMagnifier20 />,
-    [IconEnum.Magnifier24]: <IconMagnifier24 />,
-    [IconEnum.Magnifier32]: <IconMagnifier32 />,
-    [IconEnum.MagnifierPlus]: <IconMagnifierPlus />,
-    [IconEnum.MagnifierPlus12]: <IconMagnifierPlus12 />,
-    [IconEnum.MagnifierPlus16]: <IconMagnifierPlus16 />,
-    [IconEnum.MagnifierPlus20]: <IconMagnifierPlus20 />,
-    [IconEnum.MagnifierPlus24]: <IconMagnifierPlus24 />,
-    [IconEnum.MagnifierPlus32]: <IconMagnifierPlus32 />,
-    [IconEnum.MapPointer]: <IconMapPointer />,
-    [IconEnum.MapPointer12]: <IconMapPointer12 />,
-    [IconEnum.MapPointer16]: <IconMapPointer16 />,
-    [IconEnum.MapPointer20]: <IconMapPointer20 />,
-    [IconEnum.MapPointer24]: <IconMapPointer24 />,
-    [IconEnum.MapPointer32]: <IconMapPointer32 />,
-    [IconEnum.MarkArea]: <IconMarkArea />,
-    [IconEnum.MarkArea12]: <IconMarkArea12 />,
-    [IconEnum.MarkArea16]: <IconMarkArea16 />,
-    [IconEnum.MarkArea20]: <IconMarkArea20 />,
-    [IconEnum.MarkArea24]: <IconMarkArea24 />,
-    [IconEnum.MarkArea32]: <IconMarkArea32 />,
-    [IconEnum.Masonry]: <IconMasonry />,
-    [IconEnum.Masonry12]: <IconMasonry12 />,
-    [IconEnum.Masonry16]: <IconMasonry16 />,
-    [IconEnum.Masonry20]: <IconMasonry20 />,
-    [IconEnum.Masonry24]: <IconMasonry24 />,
-    [IconEnum.Masonry32]: <IconMasonry32 />,
-    [IconEnum.MediaObject]: <IconMediaObject />,
-    [IconEnum.MediaObject12]: <IconMediaObject12 />,
-    [IconEnum.MediaObject16]: <IconMediaObject16 />,
-    [IconEnum.MediaObject20]: <IconMediaObject20 />,
-    [IconEnum.MediaObject24]: <IconMediaObject24 />,
-    [IconEnum.MediaObject32]: <IconMediaObject32 />,
-    [IconEnum.MediaObjectIcon]: <IconMediaObjectIcon />,
-    [IconEnum.MediaObjectIcon12]: <IconMediaObjectIcon12 />,
-    [IconEnum.MediaObjectIcon16]: <IconMediaObjectIcon16 />,
-    [IconEnum.MediaObjectIcon20]: <IconMediaObjectIcon20 />,
-    [IconEnum.MediaObjectIcon24]: <IconMediaObjectIcon24 />,
-    [IconEnum.MediaObjectIcon32]: <IconMediaObjectIcon32 />,
-    [IconEnum.MediaObjectImage]: <IconMediaObjectImage />,
-    [IconEnum.MediaObjectImage12]: <IconMediaObjectImage12 />,
-    [IconEnum.MediaObjectImage16]: <IconMediaObjectImage16 />,
-    [IconEnum.MediaObjectImage20]: <IconMediaObjectImage20 />,
-    [IconEnum.MediaObjectImage24]: <IconMediaObjectImage24 />,
-    [IconEnum.MediaObjectImage32]: <IconMediaObjectImage32 />,
-    [IconEnum.MediaObjectRatio1To1]: <IconMediaObjectRatio1To1 />,
-    [IconEnum.MediaObjectRatio1To112]: <IconMediaObjectRatio1To112 />,
-    [IconEnum.MediaObjectRatio1To116]: <IconMediaObjectRatio1To116 />,
-    [IconEnum.MediaObjectRatio1To120]: <IconMediaObjectRatio1To120 />,
-    [IconEnum.MediaObjectRatio1To124]: <IconMediaObjectRatio1To124 />,
-    [IconEnum.MediaObjectRatio1To132]: <IconMediaObjectRatio1To132 />,
-    [IconEnum.MediaObjectRatio1To2]: <IconMediaObjectRatio1To2 />,
-    [IconEnum.MediaObjectRatio1To212]: <IconMediaObjectRatio1To212 />,
-    [IconEnum.MediaObjectRatio1To216]: <IconMediaObjectRatio1To216 />,
-    [IconEnum.MediaObjectRatio1To220]: <IconMediaObjectRatio1To220 />,
-    [IconEnum.MediaObjectRatio1To224]: <IconMediaObjectRatio1To224 />,
-    [IconEnum.MediaObjectRatio1To232]: <IconMediaObjectRatio1To232 />,
-    [IconEnum.MediaObjectRatio1To3]: <IconMediaObjectRatio1To3 />,
-    [IconEnum.MediaObjectRatio1To312]: <IconMediaObjectRatio1To312 />,
-    [IconEnum.MediaObjectRatio1To316]: <IconMediaObjectRatio1To316 />,
-    [IconEnum.MediaObjectRatio1To320]: <IconMediaObjectRatio1To320 />,
-    [IconEnum.MediaObjectRatio1To324]: <IconMediaObjectRatio1To324 />,
-    [IconEnum.MediaObjectRatio1To332]: <IconMediaObjectRatio1To332 />,
-    [IconEnum.MediaObjectRatio2To1]: <IconMediaObjectRatio2To1 />,
-    [IconEnum.MediaObjectRatio2To112]: <IconMediaObjectRatio2To112 />,
-    [IconEnum.MediaObjectRatio2To116]: <IconMediaObjectRatio2To116 />,
-    [IconEnum.MediaObjectRatio2To120]: <IconMediaObjectRatio2To120 />,
-    [IconEnum.MediaObjectRatio2To124]: <IconMediaObjectRatio2To124 />,
-    [IconEnum.MediaObjectRatio2To132]: <IconMediaObjectRatio2To132 />,
-    [IconEnum.MediaObjectRatio3To1]: <IconMediaObjectRatio3To1 />,
-    [IconEnum.MediaObjectRatio3To112]: <IconMediaObjectRatio3To112 />,
-    [IconEnum.MediaObjectRatio3To116]: <IconMediaObjectRatio3To116 />,
-    [IconEnum.MediaObjectRatio3To120]: <IconMediaObjectRatio3To120 />,
-    [IconEnum.MediaObjectRatio3To124]: <IconMediaObjectRatio3To124 />,
-    [IconEnum.MediaObjectRatio3To132]: <IconMediaObjectRatio3To132 />,
-    [IconEnum.MediaObjectTextBottom]: <IconMediaObjectTextBottom />,
-    [IconEnum.MediaObjectTextBottom12]: <IconMediaObjectTextBottom12 />,
-    [IconEnum.MediaObjectTextBottom16]: <IconMediaObjectTextBottom16 />,
-    [IconEnum.MediaObjectTextBottom20]: <IconMediaObjectTextBottom20 />,
-    [IconEnum.MediaObjectTextBottom24]: <IconMediaObjectTextBottom24 />,
-    [IconEnum.MediaObjectTextBottom32]: <IconMediaObjectTextBottom32 />,
-    [IconEnum.MediaObjectTextLeft]: <IconMediaObjectTextLeft />,
-    [IconEnum.MediaObjectTextLeft12]: <IconMediaObjectTextLeft12 />,
-    [IconEnum.MediaObjectTextLeft16]: <IconMediaObjectTextLeft16 />,
-    [IconEnum.MediaObjectTextLeft20]: <IconMediaObjectTextLeft20 />,
-    [IconEnum.MediaObjectTextLeft24]: <IconMediaObjectTextLeft24 />,
-    [IconEnum.MediaObjectTextLeft32]: <IconMediaObjectTextLeft32 />,
-    [IconEnum.MediaObjectTextRight]: <IconMediaObjectTextRight />,
-    [IconEnum.MediaObjectTextRight12]: <IconMediaObjectTextRight12 />,
-    [IconEnum.MediaObjectTextRight16]: <IconMediaObjectTextRight16 />,
-    [IconEnum.MediaObjectTextRight20]: <IconMediaObjectTextRight20 />,
-    [IconEnum.MediaObjectTextRight24]: <IconMediaObjectTextRight24 />,
-    [IconEnum.MediaObjectTextRight32]: <IconMediaObjectTextRight32 />,
-    [IconEnum.MediaObjectTextTop]: <IconMediaObjectTextTop />,
-    [IconEnum.MediaObjectTextTop12]: <IconMediaObjectTextTop12 />,
-    [IconEnum.MediaObjectTextTop16]: <IconMediaObjectTextTop16 />,
-    [IconEnum.MediaObjectTextTop20]: <IconMediaObjectTextTop20 />,
-    [IconEnum.MediaObjectTextTop24]: <IconMediaObjectTextTop24 />,
-    [IconEnum.MediaObjectTextTop32]: <IconMediaObjectTextTop32 />,
-    [IconEnum.MediaObjectVideo]: <IconMediaObjectVideo />,
-    [IconEnum.MediaObjectVideo12]: <IconMediaObjectVideo12 />,
-    [IconEnum.MediaObjectVideo16]: <IconMediaObjectVideo16 />,
-    [IconEnum.MediaObjectVideo20]: <IconMediaObjectVideo20 />,
-    [IconEnum.MediaObjectVideo24]: <IconMediaObjectVideo24 />,
-    [IconEnum.MediaObjectVideo32]: <IconMediaObjectVideo32 />,
-    [IconEnum.MegaMenu]: <IconMegaMenu />,
-    [IconEnum.MegaMenu12]: <IconMegaMenu12 />,
-    [IconEnum.MegaMenu16]: <IconMegaMenu16 />,
-    [IconEnum.MegaMenu20]: <IconMegaMenu20 />,
-    [IconEnum.MegaMenu24]: <IconMegaMenu24 />,
-    [IconEnum.MegaMenu32]: <IconMegaMenu32 />,
-    [IconEnum.Megaphone]: <IconMegaphone />,
-    [IconEnum.Megaphone12]: <IconMegaphone12 />,
-    [IconEnum.Megaphone16]: <IconMegaphone16 />,
-    [IconEnum.Megaphone20]: <IconMegaphone20 />,
-    [IconEnum.Megaphone24]: <IconMegaphone24 />,
-    [IconEnum.Megaphone32]: <IconMegaphone32 />,
-    [IconEnum.Minus]: <IconMinus />,
-    [IconEnum.Minus12]: <IconMinus12 />,
-    [IconEnum.Minus16]: <IconMinus16 />,
-    [IconEnum.Minus20]: <IconMinus20 />,
-    [IconEnum.Minus24]: <IconMinus24 />,
-    [IconEnum.Minus32]: <IconMinus32 />,
-    [IconEnum.MinusCircle]: <IconMinusCircle />,
-    [IconEnum.MinusCircle12]: <IconMinusCircle12 />,
-    [IconEnum.MinusCircle16]: <IconMinusCircle16 />,
-    [IconEnum.MinusCircle20]: <IconMinusCircle20 />,
-    [IconEnum.MinusCircle24]: <IconMinusCircle24 />,
-    [IconEnum.MinusCircle32]: <IconMinusCircle32 />,
-    [IconEnum.Money]: <IconMoney />,
-    [IconEnum.Money12]: <IconMoney12 />,
-    [IconEnum.Money16]: <IconMoney16 />,
-    [IconEnum.Money20]: <IconMoney20 />,
-    [IconEnum.Money24]: <IconMoney24 />,
-    [IconEnum.Money32]: <IconMoney32 />,
-    [IconEnum.MoneyCircle]: <IconMoneyCircle />,
-    [IconEnum.MoneyCircle12Filled]: <IconMoneyCircle12Filled />,
-    [IconEnum.MoneyCircle16Filled]: <IconMoneyCircle16Filled />,
-    [IconEnum.MoneyCircle20Filled]: <IconMoneyCircle20Filled />,
-    [IconEnum.MoneyCircle24Filled]: <IconMoneyCircle24Filled />,
-    [IconEnum.MoneyCircle32Filled]: <IconMoneyCircle32Filled />,
-    [IconEnum.MusicNote]: <IconMusicNote />,
-    [IconEnum.MusicNote12]: <IconMusicNote12 />,
-    [IconEnum.MusicNote16]: <IconMusicNote16 />,
-    [IconEnum.MusicNote20]: <IconMusicNote20 />,
-    [IconEnum.MusicNote24]: <IconMusicNote24 />,
-    [IconEnum.MusicNote32]: <IconMusicNote32 />,
-    [IconEnum.Nook]: <IconNook />,
-    [IconEnum.Nook12]: <IconNook12 />,
-    [IconEnum.Nook16]: <IconNook16 />,
-    [IconEnum.Nook20]: <IconNook20 />,
-    [IconEnum.Nook24]: <IconNook24 />,
-    [IconEnum.Nook32]: <IconNook32 />,
-    [IconEnum.OrientationBoxes]: <IconOrientationBoxes />,
-    [IconEnum.OrientationBoxes12]: <IconOrientationBoxes12 />,
-    [IconEnum.OrientationBoxes16]: <IconOrientationBoxes16 />,
-    [IconEnum.OrientationBoxes20]: <IconOrientationBoxes20 />,
-    [IconEnum.OrientationBoxes24]: <IconOrientationBoxes24 />,
-    [IconEnum.OrientationBoxes32]: <IconOrientationBoxes32 />,
-    [IconEnum.Paintbrush]: <IconPaintbrush />,
-    [IconEnum.Paintbrush12]: <IconPaintbrush12 />,
-    [IconEnum.Paintbrush16]: <IconPaintbrush16 />,
-    [IconEnum.Paintbrush20]: <IconPaintbrush20 />,
-    [IconEnum.Paintbrush24]: <IconPaintbrush24 />,
-    [IconEnum.Paintbrush32]: <IconPaintbrush32 />,
-    [IconEnum.Paperclip]: <IconPaperclip />,
-    [IconEnum.Paperclip12]: <IconPaperclip12 />,
-    [IconEnum.Paperclip16]: <IconPaperclip16 />,
-    [IconEnum.Paperclip20]: <IconPaperclip20 />,
-    [IconEnum.Paperclip24]: <IconPaperclip24 />,
-    [IconEnum.Paperclip32]: <IconPaperclip32 />,
-    [IconEnum.Paperplane]: <IconPaperplane />,
-    [IconEnum.Paperplane12]: <IconPaperplane12 />,
-    [IconEnum.Paperplane16]: <IconPaperplane16 />,
-    [IconEnum.Paperplane20]: <IconPaperplane20 />,
-    [IconEnum.Paperplane24]: <IconPaperplane24 />,
-    [IconEnum.Paperplane32]: <IconPaperplane32 />,
-    [IconEnum.Pause]: <IconPause />,
-    [IconEnum.Pause12]: <IconPause12 />,
-    [IconEnum.Pause12Filled]: <IconPause12Filled />,
-    [IconEnum.Pause16]: <IconPause16 />,
-    [IconEnum.Pause16Filled]: <IconPause16Filled />,
-    [IconEnum.Pause20]: <IconPause20 />,
-    [IconEnum.Pause20Filled]: <IconPause20Filled />,
-    [IconEnum.Pause24]: <IconPause24 />,
-    [IconEnum.Pause24Filled]: <IconPause24Filled />,
-    [IconEnum.Pause32]: <IconPause32 />,
-    [IconEnum.Pause32Filled]: <IconPause32Filled />,
-    [IconEnum.Pen]: <IconPen />,
-    [IconEnum.Pen12]: <IconPen12 />,
-    [IconEnum.Pen16]: <IconPen16 />,
-    [IconEnum.Pen20]: <IconPen20 />,
-    [IconEnum.Pen24]: <IconPen24 />,
-    [IconEnum.Pen32]: <IconPen32 />,
-    [IconEnum.People]: <IconPeople />,
-    [IconEnum.People12]: <IconPeople12 />,
-    [IconEnum.People16]: <IconPeople16 />,
-    [IconEnum.People20]: <IconPeople20 />,
-    [IconEnum.People24]: <IconPeople24 />,
-    [IconEnum.People32]: <IconPeople32 />,
-    [IconEnum.Person]: <IconPerson />,
-    [IconEnum.Person12]: <IconPerson12 />,
-    [IconEnum.Person16]: <IconPerson16 />,
-    [IconEnum.Person20]: <IconPerson20 />,
-    [IconEnum.Person24]: <IconPerson24 />,
-    [IconEnum.Person32]: <IconPerson32 />,
-    [IconEnum.PersonAnonymous]: <IconPersonAnonymous />,
-    [IconEnum.PersonAnonymous12]: <IconPersonAnonymous12 />,
-    [IconEnum.PersonAnonymous16]: <IconPersonAnonymous16 />,
-    [IconEnum.PersonAnonymous20]: <IconPersonAnonymous20 />,
-    [IconEnum.PersonAnonymous24]: <IconPersonAnonymous24 />,
-    [IconEnum.PersonAnonymous32]: <IconPersonAnonymous32 />,
-    [IconEnum.PersonSlash]: <IconPersonSlash />,
-    [IconEnum.PersonSlash12]: <IconPersonSlash12 />,
-    [IconEnum.PersonSlash16]: <IconPersonSlash16 />,
-    [IconEnum.PersonSlash20]: <IconPersonSlash20 />,
-    [IconEnum.PersonSlash24]: <IconPersonSlash24 />,
-    [IconEnum.PersonSlash32]: <IconPersonSlash32 />,
-    [IconEnum.Pills]: <IconPills />,
-    [IconEnum.Pills12]: <IconPills12 />,
-    [IconEnum.Pills16]: <IconPills16 />,
-    [IconEnum.Pills20]: <IconPills20 />,
-    [IconEnum.Pills24]: <IconPills24 />,
-    [IconEnum.Pills32]: <IconPills32 />,
-    [IconEnum.Pin]: <IconPin />,
-    [IconEnum.Pin12]: <IconPin12 />,
-    [IconEnum.Pin12Filled]: <IconPin12Filled />,
-    [IconEnum.Pin16]: <IconPin16 />,
-    [IconEnum.Pin16Filled]: <IconPin16Filled />,
-    [IconEnum.Pin20]: <IconPin20 />,
-    [IconEnum.Pin20Filled]: <IconPin20Filled />,
-    [IconEnum.Pin24]: <IconPin24 />,
-    [IconEnum.Pin24Filled]: <IconPin24Filled />,
-    [IconEnum.Pin32]: <IconPin32 />,
-    [IconEnum.Pin32Filled]: <IconPin32Filled />,
-    [IconEnum.Play]: <IconPlay />,
-    [IconEnum.Play12]: <IconPlay12 />,
-    [IconEnum.Play12Filled]: <IconPlay12Filled />,
-    [IconEnum.Play16]: <IconPlay16 />,
-    [IconEnum.Play16Filled]: <IconPlay16Filled />,
-    [IconEnum.Play20]: <IconPlay20 />,
-    [IconEnum.Play20Filled]: <IconPlay20Filled />,
-    [IconEnum.Play24]: <IconPlay24 />,
-    [IconEnum.Play24Filled]: <IconPlay24Filled />,
-    [IconEnum.Play32]: <IconPlay32 />,
-    [IconEnum.Play32Filled]: <IconPlay32Filled />,
-    [IconEnum.PlayBox]: <IconPlayBox />,
-    [IconEnum.PlayBox12]: <IconPlayBox12 />,
-    [IconEnum.PlayBox16]: <IconPlayBox16 />,
-    [IconEnum.PlayBox20]: <IconPlayBox20 />,
-    [IconEnum.PlayBox24]: <IconPlayBox24 />,
-    [IconEnum.PlayBox32]: <IconPlayBox32 />,
-    [IconEnum.PlayCircle]: <IconPlayCircle />,
-    [IconEnum.PlayCircle12]: <IconPlayCircle12 />,
-    [IconEnum.PlayCircle16]: <IconPlayCircle16 />,
-    [IconEnum.PlayCircle20]: <IconPlayCircle20 />,
-    [IconEnum.PlayCircle24]: <IconPlayCircle24 />,
-    [IconEnum.PlayCircle32]: <IconPlayCircle32 />,
-    [IconEnum.PlayFrame]: <IconPlayFrame />,
-    [IconEnum.PlayFrame12]: <IconPlayFrame12 />,
-    [IconEnum.PlayFrame16]: <IconPlayFrame16 />,
-    [IconEnum.PlayFrame20]: <IconPlayFrame20 />,
-    [IconEnum.PlayFrame24]: <IconPlayFrame24 />,
-    [IconEnum.PlayFrame32]: <IconPlayFrame32 />,
-    [IconEnum.Plus]: <IconPlus />,
-    [IconEnum.Plus12]: <IconPlus12 />,
-    [IconEnum.Plus16]: <IconPlus16 />,
-    [IconEnum.Plus20]: <IconPlus20 />,
-    [IconEnum.Plus24]: <IconPlus24 />,
-    [IconEnum.Plus32]: <IconPlus32 />,
-    [IconEnum.PlusBoxStack]: <IconPlusBoxStack />,
-    [IconEnum.PlusBoxStack12]: <IconPlusBoxStack12 />,
-    [IconEnum.PlusBoxStack16]: <IconPlusBoxStack16 />,
-    [IconEnum.PlusBoxStack20]: <IconPlusBoxStack20 />,
-    [IconEnum.PlusBoxStack24]: <IconPlusBoxStack24 />,
-    [IconEnum.PlusBoxStack32]: <IconPlusBoxStack32 />,
-    [IconEnum.PlusCircle]: <IconPlusCircle />,
-    [IconEnum.PlusCircle12]: <IconPlusCircle12 />,
-    [IconEnum.PlusCircle16]: <IconPlusCircle16 />,
-    [IconEnum.PlusCircle20]: <IconPlusCircle20 />,
-    [IconEnum.PlusCircle24]: <IconPlusCircle24 />,
-    [IconEnum.PlusCircle32]: <IconPlusCircle32 />,
-    [IconEnum.PlusFrame]: <IconPlusFrame />,
-    [IconEnum.PlusFrame12]: <IconPlusFrame12 />,
-    [IconEnum.PlusFrame16]: <IconPlusFrame16 />,
-    [IconEnum.PlusFrame20]: <IconPlusFrame20 />,
-    [IconEnum.PlusFrame24]: <IconPlusFrame24 />,
-    [IconEnum.PlusFrame32]: <IconPlusFrame32 />,
-    [IconEnum.Pointout]: <IconPointout />,
-    [IconEnum.Pointout12]: <IconPointout12 />,
-    [IconEnum.Pointout16]: <IconPointout16 />,
-    [IconEnum.Pointout20]: <IconPointout20 />,
-    [IconEnum.Pointout24]: <IconPointout24 />,
-    [IconEnum.Pointout32]: <IconPointout32 />,
-    [IconEnum.Polygon]: <IconPolygon />,
-    [IconEnum.Polygon12]: <IconPolygon12 />,
-    [IconEnum.Polygon16]: <IconPolygon16 />,
-    [IconEnum.Polygon20]: <IconPolygon20 />,
-    [IconEnum.Polygon24]: <IconPolygon24 />,
-    [IconEnum.Polygon32]: <IconPolygon32 />,
-    [IconEnum.Puzzle]: <IconPuzzle />,
-    [IconEnum.Puzzle12]: <IconPuzzle12 />,
-    [IconEnum.Puzzle16]: <IconPuzzle16 />,
-    [IconEnum.Puzzle20]: <IconPuzzle20 />,
-    [IconEnum.Puzzle24]: <IconPuzzle24 />,
-    [IconEnum.Puzzle32]: <IconPuzzle32 />,
-    [IconEnum.QuestionMark]: <IconQuestionMark />,
-    [IconEnum.QuestionMark12]: <IconQuestionMark12 />,
-    [IconEnum.QuestionMark16]: <IconQuestionMark16 />,
-    [IconEnum.QuestionMark20]: <IconQuestionMark20 />,
-    [IconEnum.QuestionMark24]: <IconQuestionMark24 />,
-    [IconEnum.QuestionMark32]: <IconQuestionMark32 />,
-    [IconEnum.QuestionMarkCircle]: <IconQuestionMarkCircle />,
-    [IconEnum.QuestionMarkCircle12]: <IconQuestionMarkCircle12 />,
-    [IconEnum.QuestionMarkCircle16]: <IconQuestionMarkCircle16 />,
-    [IconEnum.QuestionMarkCircle20]: <IconQuestionMarkCircle20 />,
-    [IconEnum.QuestionMarkCircle24]: <IconQuestionMarkCircle24 />,
-    [IconEnum.QuestionMarkCircle32]: <IconQuestionMarkCircle32 />,
-    [IconEnum.RectangleLandscape]: <IconRectangleLandscape />,
-    [IconEnum.RectangleLandscape12]: <IconRectangleLandscape12 />,
-    [IconEnum.RectangleLandscape16]: <IconRectangleLandscape16 />,
-    [IconEnum.RectangleLandscape20]: <IconRectangleLandscape20 />,
-    [IconEnum.RectangleLandscape24]: <IconRectangleLandscape24 />,
-    [IconEnum.RectangleLandscape32]: <IconRectangleLandscape32 />,
-    [IconEnum.RectanglePortraitSquare]: <IconRectanglePortraitSquare />,
-    [IconEnum.RectanglePortraitSquare12]: <IconRectanglePortraitSquare12 />,
-    [IconEnum.RectanglePortraitSquare16]: <IconRectanglePortraitSquare16 />,
-    [IconEnum.RectanglePortraitSquare20]: <IconRectanglePortraitSquare20 />,
-    [IconEnum.RectanglePortraitSquare24]: <IconRectanglePortraitSquare24 />,
-    [IconEnum.RectanglePortraitSquare32]: <IconRectanglePortraitSquare32 />,
-    [IconEnum.Report]: <IconReport />,
-    [IconEnum.Report12]: <IconReport12 />,
-    [IconEnum.Report16]: <IconReport16 />,
-    [IconEnum.Report20]: <IconReport20 />,
-    [IconEnum.Report24]: <IconReport24 />,
-    [IconEnum.Report32]: <IconReport32 />,
-    [IconEnum.Rocket]: <IconRocket />,
-    [IconEnum.Rocket12]: <IconRocket12 />,
-    [IconEnum.Rocket16]: <IconRocket16 />,
-    [IconEnum.Rocket20]: <IconRocket20 />,
-    [IconEnum.Rocket24]: <IconRocket24 />,
-    [IconEnum.Rocket32]: <IconRocket32 />,
-    [IconEnum.Ruler]: <IconRuler />,
-    [IconEnum.Ruler12]: <IconRuler12 />,
-    [IconEnum.Ruler16]: <IconRuler16 />,
-    [IconEnum.Ruler20]: <IconRuler20 />,
-    [IconEnum.Ruler24]: <IconRuler24 />,
-    [IconEnum.Ruler32]: <IconRuler32 />,
-    [IconEnum.ScaleLarge]: <IconScaleLarge />,
-    [IconEnum.ScaleLarge12]: <IconScaleLarge12 />,
-    [IconEnum.ScaleLarge16]: <IconScaleLarge16 />,
-    [IconEnum.ScaleLarge20]: <IconScaleLarge20 />,
-    [IconEnum.ScaleLarge24]: <IconScaleLarge24 />,
-    [IconEnum.ScaleLarge32]: <IconScaleLarge32 />,
-    [IconEnum.ScaleSmall]: <IconScaleSmall />,
-    [IconEnum.ScaleSmall12]: <IconScaleSmall12 />,
-    [IconEnum.ScaleSmall16]: <IconScaleSmall16 />,
-    [IconEnum.ScaleSmall20]: <IconScaleSmall20 />,
-    [IconEnum.ScaleSmall24]: <IconScaleSmall24 />,
-    [IconEnum.ScaleSmall32]: <IconScaleSmall32 />,
-    [IconEnum.Scissors]: <IconScissors />,
-    [IconEnum.Scissors12]: <IconScissors12 />,
-    [IconEnum.Scissors16]: <IconScissors16 />,
-    [IconEnum.Scissors20]: <IconScissors20 />,
-    [IconEnum.Scissors24]: <IconScissors24 />,
-    [IconEnum.Scissors32]: <IconScissors32 />,
-    [IconEnum.Side]: <IconSide />,
-    [IconEnum.Side12]: <IconSide12 />,
-    [IconEnum.Side16]: <IconSide16 />,
-    [IconEnum.Side20]: <IconSide20 />,
-    [IconEnum.Side24]: <IconSide24 />,
-    [IconEnum.Side32]: <IconSide32 />,
-    [IconEnum.SidebarLeft]: <IconSidebarLeft />,
-    [IconEnum.SidebarLeft12]: <IconSidebarLeft12 />,
-    [IconEnum.SidebarLeft16]: <IconSidebarLeft16 />,
-    [IconEnum.SidebarLeft20]: <IconSidebarLeft20 />,
-    [IconEnum.SidebarLeft24]: <IconSidebarLeft24 />,
-    [IconEnum.SidebarLeft32]: <IconSidebarLeft32 />,
-    [IconEnum.SidebarLeftHide]: <IconSidebarLeftHide />,
-    [IconEnum.SidebarLeftHide12]: <IconSidebarLeftHide12 />,
-    [IconEnum.SidebarLeftHide12Filled]: <IconSidebarLeftHide12Filled />,
-    [IconEnum.SidebarLeftHide16]: <IconSidebarLeftHide16 />,
-    [IconEnum.SidebarLeftHide16Filled]: <IconSidebarLeftHide16Filled />,
-    [IconEnum.SidebarLeftHide20]: <IconSidebarLeftHide20 />,
-    [IconEnum.SidebarLeftHide20Filled]: <IconSidebarLeftHide20Filled />,
-    [IconEnum.SidebarLeftHide24]: <IconSidebarLeftHide24 />,
-    [IconEnum.SidebarLeftHide24Filled]: <IconSidebarLeftHide24Filled />,
-    [IconEnum.SidebarLeftHide32]: <IconSidebarLeftHide32 />,
-    [IconEnum.SidebarLeftHide32Filled]: <IconSidebarLeftHide32Filled />,
-    [IconEnum.SidebarLeftShow]: <IconSidebarLeftShow />,
-    [IconEnum.SidebarLeftShow12]: <IconSidebarLeftShow12 />,
-    [IconEnum.SidebarLeftShow12Filled]: <IconSidebarLeftShow12Filled />,
-    [IconEnum.SidebarLeftShow16]: <IconSidebarLeftShow16 />,
-    [IconEnum.SidebarLeftShow16Filled]: <IconSidebarLeftShow16Filled />,
-    [IconEnum.SidebarLeftShow20]: <IconSidebarLeftShow20 />,
-    [IconEnum.SidebarLeftShow20Filled]: <IconSidebarLeftShow20Filled />,
-    [IconEnum.SidebarLeftShow24]: <IconSidebarLeftShow24 />,
-    [IconEnum.SidebarLeftShow24Filled]: <IconSidebarLeftShow24Filled />,
-    [IconEnum.SidebarLeftShow32]: <IconSidebarLeftShow32 />,
-    [IconEnum.SidebarLeftShow32Filled]: <IconSidebarLeftShow32Filled />,
-    [IconEnum.SidebarRight]: <IconSidebarRight />,
-    [IconEnum.SidebarRight12]: <IconSidebarRight12 />,
-    [IconEnum.SidebarRight16]: <IconSidebarRight16 />,
-    [IconEnum.SidebarRight20]: <IconSidebarRight20 />,
-    [IconEnum.SidebarRight24]: <IconSidebarRight24 />,
-    [IconEnum.SidebarRight32]: <IconSidebarRight32 />,
-    [IconEnum.SidebarRightHide]: <IconSidebarRightHide />,
-    [IconEnum.SidebarRightHide12]: <IconSidebarRightHide12 />,
-    [IconEnum.SidebarRightHide12Filled]: <IconSidebarRightHide12Filled />,
-    [IconEnum.SidebarRightHide16]: <IconSidebarRightHide16 />,
-    [IconEnum.SidebarRightHide16Filled]: <IconSidebarRightHide16Filled />,
-    [IconEnum.SidebarRightHide20]: <IconSidebarRightHide20 />,
-    [IconEnum.SidebarRightHide20Filled]: <IconSidebarRightHide20Filled />,
-    [IconEnum.SidebarRightHide24]: <IconSidebarRightHide24 />,
-    [IconEnum.SidebarRightHide24Filled]: <IconSidebarRightHide24Filled />,
-    [IconEnum.SidebarRightHide32]: <IconSidebarRightHide32 />,
-    [IconEnum.SidebarRightHide32Filled]: <IconSidebarRightHide32Filled />,
-    [IconEnum.SidebarRightShow]: <IconSidebarRightShow />,
-    [IconEnum.SidebarRightShow12]: <IconSidebarRightShow12 />,
-    [IconEnum.SidebarRightShow12Filled]: <IconSidebarRightShow12Filled />,
-    [IconEnum.SidebarRightShow16]: <IconSidebarRightShow16 />,
-    [IconEnum.SidebarRightShow16Filled]: <IconSidebarRightShow16Filled />,
-    [IconEnum.SidebarRightShow20]: <IconSidebarRightShow20 />,
-    [IconEnum.SidebarRightShow20Filled]: <IconSidebarRightShow20Filled />,
-    [IconEnum.SidebarRightShow24]: <IconSidebarRightShow24 />,
-    [IconEnum.SidebarRightShow24Filled]: <IconSidebarRightShow24Filled />,
-    [IconEnum.SidebarRightShow32]: <IconSidebarRightShow32 />,
-    [IconEnum.SidebarRightShow32Filled]: <IconSidebarRightShow32Filled />,
-    [IconEnum.Sketch]: <IconSketch />,
-    [IconEnum.Sketch12]: <IconSketch12 />,
-    [IconEnum.Sketch16]: <IconSketch16 />,
-    [IconEnum.Sketch20]: <IconSketch20 />,
-    [IconEnum.Sketch24]: <IconSketch24 />,
-    [IconEnum.Sketch32]: <IconSketch32 />,
-    [IconEnum.Skip5SecondsBackward]: <IconSkip5SecondsBackward />,
-    [IconEnum.Skip5SecondsBackward12]: <IconSkip5SecondsBackward12 />,
-    [IconEnum.Skip5SecondsBackward16]: <IconSkip5SecondsBackward16 />,
-    [IconEnum.Skip5SecondsBackward20]: <IconSkip5SecondsBackward20 />,
-    [IconEnum.Skip5SecondsBackward24]: <IconSkip5SecondsBackward24 />,
-    [IconEnum.Skip5SecondsBackward32]: <IconSkip5SecondsBackward32 />,
-    [IconEnum.Skip5SecondsForward]: <IconSkip5SecondsForward />,
-    [IconEnum.Skip5SecondsForward12]: <IconSkip5SecondsForward12 />,
-    [IconEnum.Skip5SecondsForward16]: <IconSkip5SecondsForward16 />,
-    [IconEnum.Skip5SecondsForward20]: <IconSkip5SecondsForward20 />,
-    [IconEnum.Skip5SecondsForward24]: <IconSkip5SecondsForward24 />,
-    [IconEnum.Skip5SecondsForward32]: <IconSkip5SecondsForward32 />,
-    [IconEnum.SoundWave]: <IconSoundWave />,
-    [IconEnum.SoundWave12]: <IconSoundWave12 />,
-    [IconEnum.SoundWave16]: <IconSoundWave16 />,
-    [IconEnum.SoundWave20]: <IconSoundWave20 />,
-    [IconEnum.SoundWave24]: <IconSoundWave24 />,
-    [IconEnum.SoundWave32]: <IconSoundWave32 />,
-    [IconEnum.Speaker]: <IconSpeaker />,
-    [IconEnum.Speaker12]: <IconSpeaker12 />,
-    [IconEnum.Speaker16]: <IconSpeaker16 />,
-    [IconEnum.Speaker20]: <IconSpeaker20 />,
-    [IconEnum.Speaker24]: <IconSpeaker24 />,
-    [IconEnum.Speaker32]: <IconSpeaker32 />,
-    [IconEnum.SpeakerOff]: <IconSpeakerOff />,
-    [IconEnum.SpeakerOff12]: <IconSpeakerOff12 />,
-    [IconEnum.SpeakerOff16]: <IconSpeakerOff16 />,
-    [IconEnum.SpeakerOff20]: <IconSpeakerOff20 />,
-    [IconEnum.SpeakerOff24]: <IconSpeakerOff24 />,
-    [IconEnum.SpeakerOff32]: <IconSpeakerOff32 />,
-    [IconEnum.SpeechBubbleDots]: <IconSpeechBubbleDots />,
-    [IconEnum.SpeechBubbleDots12]: <IconSpeechBubbleDots12 />,
-    [IconEnum.SpeechBubbleDots16]: <IconSpeechBubbleDots16 />,
-    [IconEnum.SpeechBubbleDots20]: <IconSpeechBubbleDots20 />,
-    [IconEnum.SpeechBubbleDots24]: <IconSpeechBubbleDots24 />,
-    [IconEnum.SpeechBubbleDots32]: <IconSpeechBubbleDots32 />,
-    [IconEnum.SpeechBubbleEmpty]: <IconSpeechBubbleEmpty />,
-    [IconEnum.SpeechBubbleEmpty12]: <IconSpeechBubbleEmpty12 />,
-    [IconEnum.SpeechBubbleEmpty16]: <IconSpeechBubbleEmpty16 />,
-    [IconEnum.SpeechBubbleEmpty20]: <IconSpeechBubbleEmpty20 />,
-    [IconEnum.SpeechBubbleEmpty24]: <IconSpeechBubbleEmpty24 />,
-    [IconEnum.SpeechBubbleEmpty32]: <IconSpeechBubbleEmpty32 />,
-    [IconEnum.SpeechBubbleQuote]: <IconSpeechBubbleQuote />,
-    [IconEnum.SpeechBubbleQuote12]: <IconSpeechBubbleQuote12 />,
-    [IconEnum.SpeechBubbleQuote16]: <IconSpeechBubbleQuote16 />,
-    [IconEnum.SpeechBubbleQuote20]: <IconSpeechBubbleQuote20 />,
-    [IconEnum.SpeechBubbleQuote24]: <IconSpeechBubbleQuote24 />,
-    [IconEnum.SpeechBubbleQuote32]: <IconSpeechBubbleQuote32 />,
-    [IconEnum.SpeechBubbleStack]: <IconSpeechBubbleStack />,
-    [IconEnum.SpeechBubbleStack12]: <IconSpeechBubbleStack12 />,
-    [IconEnum.SpeechBubbleStack16]: <IconSpeechBubbleStack16 />,
-    [IconEnum.SpeechBubbleStack20]: <IconSpeechBubbleStack20 />,
-    [IconEnum.SpeechBubbleStack24]: <IconSpeechBubbleStack24 />,
-    [IconEnum.SpeechBubbleStack32]: <IconSpeechBubbleStack32 />,
-    [IconEnum.StackHorizontal]: <IconStackHorizontal />,
-    [IconEnum.StackHorizontal12]: <IconStackHorizontal12 />,
-    [IconEnum.StackHorizontal16]: <IconStackHorizontal16 />,
-    [IconEnum.StackHorizontal20]: <IconStackHorizontal20 />,
-    [IconEnum.StackHorizontal24]: <IconStackHorizontal24 />,
-    [IconEnum.StackHorizontal32]: <IconStackHorizontal32 />,
-    [IconEnum.StackHorizontalAltFour]: <IconStackHorizontalAltFour />,
-    [IconEnum.StackHorizontalAltFour12]: <IconStackHorizontalAltFour12 />,
-    [IconEnum.StackHorizontalAltFour16]: <IconStackHorizontalAltFour16 />,
-    [IconEnum.StackHorizontalAltFour20]: <IconStackHorizontalAltFour20 />,
-    [IconEnum.StackHorizontalAltFour24]: <IconStackHorizontalAltFour24 />,
-    [IconEnum.StackHorizontalAltFour32]: <IconStackHorizontalAltFour32 />,
-    [IconEnum.StackHorizontalAltThree]: <IconStackHorizontalAltThree />,
-    [IconEnum.StackHorizontalAltThree12]: <IconStackHorizontalAltThree12 />,
-    [IconEnum.StackHorizontalAltThree16]: <IconStackHorizontalAltThree16 />,
-    [IconEnum.StackHorizontalAltThree20]: <IconStackHorizontalAltThree20 />,
-    [IconEnum.StackHorizontalAltThree24]: <IconStackHorizontalAltThree24 />,
-    [IconEnum.StackHorizontalAltThree32]: <IconStackHorizontalAltThree32 />,
-    [IconEnum.StackHorizontalAltTwo]: <IconStackHorizontalAltTwo />,
-    [IconEnum.StackHorizontalAltTwo12]: <IconStackHorizontalAltTwo12 />,
-    [IconEnum.StackHorizontalAltTwo16]: <IconStackHorizontalAltTwo16 />,
-    [IconEnum.StackHorizontalAltTwo20]: <IconStackHorizontalAltTwo20 />,
-    [IconEnum.StackHorizontalAltTwo24]: <IconStackHorizontalAltTwo24 />,
-    [IconEnum.StackHorizontalAltTwo32]: <IconStackHorizontalAltTwo32 />,
-    [IconEnum.StackVertical]: <IconStackVertical />,
-    [IconEnum.StackVertical12]: <IconStackVertical12 />,
-    [IconEnum.StackVertical16]: <IconStackVertical16 />,
-    [IconEnum.StackVertical20]: <IconStackVertical20 />,
-    [IconEnum.StackVertical24]: <IconStackVertical24 />,
-    [IconEnum.StackVertical32]: <IconStackVertical32 />,
-    [IconEnum.Star]: <IconStar />,
-    [IconEnum.Star12]: <IconStar12 />,
-    [IconEnum.Star12Filled]: <IconStar12Filled />,
-    [IconEnum.Star16]: <IconStar16 />,
-    [IconEnum.Star16Filled]: <IconStar16Filled />,
-    [IconEnum.Star20]: <IconStar20 />,
-    [IconEnum.Star20Filled]: <IconStar20Filled />,
-    [IconEnum.Star24]: <IconStar24 />,
-    [IconEnum.Star24Filled]: <IconStar24Filled />,
-    [IconEnum.Star32]: <IconStar32 />,
-    [IconEnum.Star32Filled]: <IconStar32Filled />,
-    [IconEnum.Storybook]: <IconStorybook />,
-    [IconEnum.Storybook12]: <IconStorybook12 />,
-    [IconEnum.Storybook16]: <IconStorybook16 />,
-    [IconEnum.Storybook20]: <IconStorybook20 />,
-    [IconEnum.Storybook24]: <IconStorybook24 />,
-    [IconEnum.Storybook32]: <IconStorybook32 />,
-    [IconEnum.StrikethroughBox]: <IconStrikethroughBox />,
-    [IconEnum.StrikethroughBox12]: <IconStrikethroughBox12 />,
-    [IconEnum.StrikethroughBox16]: <IconStrikethroughBox16 />,
-    [IconEnum.StrikethroughBox20]: <IconStrikethroughBox20 />,
-    [IconEnum.StrikethroughBox24]: <IconStrikethroughBox24 />,
-    [IconEnum.StrikethroughBox32]: <IconStrikethroughBox32 />,
-    [IconEnum.Subscript]: <IconSubscript />,
-    [IconEnum.Subscript12]: <IconSubscript12 />,
-    [IconEnum.Subscript16]: <IconSubscript16 />,
-    [IconEnum.Subscript20]: <IconSubscript20 />,
-    [IconEnum.Subscript24]: <IconSubscript24 />,
-    [IconEnum.Subscript32]: <IconSubscript32 />,
-    [IconEnum.Suitcase]: <IconSuitcase />,
-    [IconEnum.Suitcase12]: <IconSuitcase12 />,
-    [IconEnum.Suitcase16]: <IconSuitcase16 />,
-    [IconEnum.Suitcase20]: <IconSuitcase20 />,
-    [IconEnum.Suitcase24]: <IconSuitcase24 />,
-    [IconEnum.Suitcase32]: <IconSuitcase32 />,
-    [IconEnum.SuitcaseTag]: <IconSuitcaseTag />,
-    [IconEnum.SuitcaseTag12]: <IconSuitcaseTag12 />,
-    [IconEnum.SuitcaseTag16]: <IconSuitcaseTag16 />,
-    [IconEnum.SuitcaseTag20]: <IconSuitcaseTag20 />,
-    [IconEnum.SuitcaseTag24]: <IconSuitcaseTag24 />,
-    [IconEnum.SuitcaseTag32]: <IconSuitcaseTag32 />,
-    [IconEnum.Superscript]: <IconSuperscript />,
-    [IconEnum.Superscript12]: <IconSuperscript12 />,
-    [IconEnum.Superscript16]: <IconSuperscript16 />,
-    [IconEnum.Superscript20]: <IconSuperscript20 />,
-    [IconEnum.Superscript24]: <IconSuperscript24 />,
-    [IconEnum.Superscript32]: <IconSuperscript32 />,
-    [IconEnum.Table]: <IconTable />,
-    [IconEnum.Table12]: <IconTable12 />,
-    [IconEnum.Table16]: <IconTable16 />,
-    [IconEnum.Table20]: <IconTable20 />,
-    [IconEnum.Table24]: <IconTable24 />,
-    [IconEnum.Table32]: <IconTable32 />,
-    [IconEnum.Target]: <IconTarget />,
-    [IconEnum.Target12]: <IconTarget12 />,
-    [IconEnum.Target16]: <IconTarget16 />,
-    [IconEnum.Target20]: <IconTarget20 />,
-    [IconEnum.Target24]: <IconTarget24 />,
-    [IconEnum.Target32]: <IconTarget32 />,
-    [IconEnum.Tester]: <IconTester />,
-    [IconEnum.Tester12]: <IconTester12 />,
-    [IconEnum.Tester16]: <IconTester16 />,
-    [IconEnum.Tester20]: <IconTester20 />,
-    [IconEnum.Tester24]: <IconTester24 />,
-    [IconEnum.Tester32]: <IconTester32 />,
-    [IconEnum.TextAlignmentCentre]: <IconTextAlignmentCentre />,
-    [IconEnum.TextAlignmentCentre12]: <IconTextAlignmentCentre12 />,
-    [IconEnum.TextAlignmentCentre16]: <IconTextAlignmentCentre16 />,
-    [IconEnum.TextAlignmentCentre20]: <IconTextAlignmentCentre20 />,
-    [IconEnum.TextAlignmentCentre24]: <IconTextAlignmentCentre24 />,
-    [IconEnum.TextAlignmentCentre32]: <IconTextAlignmentCentre32 />,
-    [IconEnum.TextAlignmentJustify]: <IconTextAlignmentJustify />,
-    [IconEnum.TextAlignmentJustify12]: <IconTextAlignmentJustify12 />,
-    [IconEnum.TextAlignmentJustify16]: <IconTextAlignmentJustify16 />,
-    [IconEnum.TextAlignmentJustify20]: <IconTextAlignmentJustify20 />,
-    [IconEnum.TextAlignmentJustify24]: <IconTextAlignmentJustify24 />,
-    [IconEnum.TextAlignmentJustify32]: <IconTextAlignmentJustify32 />,
-    [IconEnum.TextAlignmentLeft]: <IconTextAlignmentLeft />,
-    [IconEnum.TextAlignmentLeft12]: <IconTextAlignmentLeft12 />,
-    [IconEnum.TextAlignmentLeft16]: <IconTextAlignmentLeft16 />,
-    [IconEnum.TextAlignmentLeft20]: <IconTextAlignmentLeft20 />,
-    [IconEnum.TextAlignmentLeft24]: <IconTextAlignmentLeft24 />,
-    [IconEnum.TextAlignmentLeft32]: <IconTextAlignmentLeft32 />,
-    [IconEnum.TextAlignmentRight]: <IconTextAlignmentRight />,
-    [IconEnum.TextAlignmentRight12]: <IconTextAlignmentRight12 />,
-    [IconEnum.TextAlignmentRight16]: <IconTextAlignmentRight16 />,
-    [IconEnum.TextAlignmentRight20]: <IconTextAlignmentRight20 />,
-    [IconEnum.TextAlignmentRight24]: <IconTextAlignmentRight24 />,
-    [IconEnum.TextAlignmentRight32]: <IconTextAlignmentRight32 />,
-    [IconEnum.TextBoxStack]: <IconTextBoxStack />,
-    [IconEnum.TextBoxStack12]: <IconTextBoxStack12 />,
-    [IconEnum.TextBoxStack16]: <IconTextBoxStack16 />,
-    [IconEnum.TextBoxStack20]: <IconTextBoxStack20 />,
-    [IconEnum.TextBoxStack24]: <IconTextBoxStack24 />,
-    [IconEnum.TextBoxStack32]: <IconTextBoxStack32 />,
-    [IconEnum.TextBrackets]: <IconTextBrackets />,
-    [IconEnum.TextBrackets12]: <IconTextBrackets12 />,
-    [IconEnum.TextBrackets16]: <IconTextBrackets16 />,
-    [IconEnum.TextBrackets20]: <IconTextBrackets20 />,
-    [IconEnum.TextBrackets24]: <IconTextBrackets24 />,
-    [IconEnum.TextBrackets32]: <IconTextBrackets32 />,
-    [IconEnum.TextColumnBreak]: <IconTextColumnBreak />,
-    [IconEnum.TextColumnBreak12]: <IconTextColumnBreak12 />,
-    [IconEnum.TextColumnBreak16]: <IconTextColumnBreak16 />,
-    [IconEnum.TextColumnBreak20]: <IconTextColumnBreak20 />,
-    [IconEnum.TextColumnBreak24]: <IconTextColumnBreak24 />,
-    [IconEnum.TextColumnBreak32]: <IconTextColumnBreak32 />,
-    [IconEnum.TextColumns]: <IconTextColumns />,
-    [IconEnum.TextColumns12]: <IconTextColumns12 />,
-    [IconEnum.TextColumns16]: <IconTextColumns16 />,
-    [IconEnum.TextColumns20]: <IconTextColumns20 />,
-    [IconEnum.TextColumns24]: <IconTextColumns24 />,
-    [IconEnum.TextColumns32]: <IconTextColumns32 />,
-    [IconEnum.TextFormatBold]: <IconTextFormatBold />,
-    [IconEnum.TextFormatBold12]: <IconTextFormatBold12 />,
-    [IconEnum.TextFormatBold16]: <IconTextFormatBold16 />,
-    [IconEnum.TextFormatBold20]: <IconTextFormatBold20 />,
-    [IconEnum.TextFormatBold24]: <IconTextFormatBold24 />,
-    [IconEnum.TextFormatBold32]: <IconTextFormatBold32 />,
-    [IconEnum.TextFormatItalic]: <IconTextFormatItalic />,
-    [IconEnum.TextFormatItalic12]: <IconTextFormatItalic12 />,
-    [IconEnum.TextFormatItalic16]: <IconTextFormatItalic16 />,
-    [IconEnum.TextFormatItalic20]: <IconTextFormatItalic20 />,
-    [IconEnum.TextFormatItalic24]: <IconTextFormatItalic24 />,
-    [IconEnum.TextFormatItalic32]: <IconTextFormatItalic32 />,
-    [IconEnum.TextFormatStrikethrough]: <IconTextFormatStrikethrough />,
-    [IconEnum.TextFormatStrikethrough12]: <IconTextFormatStrikethrough12 />,
-    [IconEnum.TextFormatStrikethrough16]: <IconTextFormatStrikethrough16 />,
-    [IconEnum.TextFormatStrikethrough20]: <IconTextFormatStrikethrough20 />,
-    [IconEnum.TextFormatStrikethrough24]: <IconTextFormatStrikethrough24 />,
-    [IconEnum.TextFormatStrikethrough32]: <IconTextFormatStrikethrough32 />,
-    [IconEnum.TextFormatUnderline]: <IconTextFormatUnderline />,
-    [IconEnum.TextFormatUnderline12]: <IconTextFormatUnderline12 />,
-    [IconEnum.TextFormatUnderline16]: <IconTextFormatUnderline16 />,
-    [IconEnum.TextFormatUnderline20]: <IconTextFormatUnderline20 />,
-    [IconEnum.TextFormatUnderline24]: <IconTextFormatUnderline24 />,
-    [IconEnum.TextFormatUnderline32]: <IconTextFormatUnderline32 />,
-    [IconEnum.TextIndent]: <IconTextIndent />,
-    [IconEnum.TextIndent12]: <IconTextIndent12 />,
-    [IconEnum.TextIndent16]: <IconTextIndent16 />,
-    [IconEnum.TextIndent20]: <IconTextIndent20 />,
-    [IconEnum.TextIndent24]: <IconTextIndent24 />,
-    [IconEnum.TextIndent32]: <IconTextIndent32 />,
-    [IconEnum.TextOutdent]: <IconTextOutdent />,
-    [IconEnum.TextOutdent12]: <IconTextOutdent12 />,
-    [IconEnum.TextOutdent16]: <IconTextOutdent16 />,
-    [IconEnum.TextOutdent20]: <IconTextOutdent20 />,
-    [IconEnum.TextOutdent24]: <IconTextOutdent24 />,
-    [IconEnum.TextOutdent32]: <IconTextOutdent32 />,
-    [IconEnum.TrashBin]: <IconTrashBin />,
-    [IconEnum.TrashBin12]: <IconTrashBin12 />,
-    [IconEnum.TrashBin16]: <IconTrashBin16 />,
-    [IconEnum.TrashBin20]: <IconTrashBin20 />,
-    [IconEnum.TrashBin24]: <IconTrashBin24 />,
-    [IconEnum.TrashBin32]: <IconTrashBin32 />,
-    [IconEnum.TrashBinOpen]: <IconTrashBinOpen />,
-    [IconEnum.TrashBinOpen12]: <IconTrashBinOpen12 />,
-    [IconEnum.TrashBinOpen16]: <IconTrashBinOpen16 />,
-    [IconEnum.TrashBinOpen20]: <IconTrashBinOpen20 />,
-    [IconEnum.TrashBinOpen24]: <IconTrashBinOpen24 />,
-    [IconEnum.TrashBinOpen32]: <IconTrashBinOpen32 />,
-    [IconEnum.Triangle]: <IconTriangle />,
-    [IconEnum.Triangle12]: <IconTriangle12 />,
-    [IconEnum.Triangle16]: <IconTriangle16 />,
-    [IconEnum.Triangle20]: <IconTriangle20 />,
-    [IconEnum.Triangle24]: <IconTriangle24 />,
-    [IconEnum.Triangle32]: <IconTriangle32 />,
-    [IconEnum.Twitter]: <IconTwitter />,
-    [IconEnum.Twitter12]: <IconTwitter12 />,
-    [IconEnum.Twitter16]: <IconTwitter16 />,
-    [IconEnum.Twitter20]: <IconTwitter20 />,
-    [IconEnum.Twitter24]: <IconTwitter24 />,
-    [IconEnum.Twitter32]: <IconTwitter32 />,
-    [IconEnum.Typography]: <IconTypography />,
-    [IconEnum.Typography12]: <IconTypography12 />,
-    [IconEnum.Typography16]: <IconTypography16 />,
-    [IconEnum.Typography20]: <IconTypography20 />,
-    [IconEnum.Typography24]: <IconTypography24 />,
-    [IconEnum.Typography32]: <IconTypography32 />,
-    [IconEnum.TypographyBox]: <IconTypographyBox />,
-    [IconEnum.TypographyBox12]: <IconTypographyBox12 />,
-    [IconEnum.TypographyBox16]: <IconTypographyBox16 />,
-    [IconEnum.TypographyBox20]: <IconTypographyBox20 />,
-    [IconEnum.TypographyBox24]: <IconTypographyBox24 />,
-    [IconEnum.TypographyBox32]: <IconTypographyBox32 />,
-    [IconEnum.TypographyCase]: <IconTypographyCase />,
-    [IconEnum.TypographyCase12]: <IconTypographyCase12 />,
-    [IconEnum.TypographyCase16]: <IconTypographyCase16 />,
-    [IconEnum.TypographyCase20]: <IconTypographyCase20 />,
-    [IconEnum.TypographyCase24]: <IconTypographyCase24 />,
-    [IconEnum.TypographyCase32]: <IconTypographyCase32 />,
-    [IconEnum.TypographyMultiple]: <IconTypographyMultiple />,
-    [IconEnum.TypographyMultiple12]: <IconTypographyMultiple12 />,
-    [IconEnum.TypographyMultiple16]: <IconTypographyMultiple16 />,
-    [IconEnum.TypographyMultiple20]: <IconTypographyMultiple20 />,
-    [IconEnum.TypographyMultiple24]: <IconTypographyMultiple24 />,
-    [IconEnum.TypographyMultiple32]: <IconTypographyMultiple32 />,
-    [IconEnum.Vimeo]: <IconVimeo />,
-    [IconEnum.Vimeo12]: <IconVimeo12 />,
-    [IconEnum.Vimeo16]: <IconVimeo16 />,
-    [IconEnum.Vimeo20]: <IconVimeo20 />,
-    [IconEnum.Vimeo24]: <IconVimeo24 />,
-    [IconEnum.Vimeo32]: <IconVimeo32 />,
-    [IconEnum.WorkflowLanes]: <IconWorkflowLanes />,
-    [IconEnum.WorkflowLanes12]: <IconWorkflowLanes12 />,
-    [IconEnum.WorkflowLanes16]: <IconWorkflowLanes16 />,
-    [IconEnum.WorkflowLanes20]: <IconWorkflowLanes20 />,
-    [IconEnum.WorkflowLanes24]: <IconWorkflowLanes24 />,
-    [IconEnum.WorkflowLanes32]: <IconWorkflowLanes32 />,
-    [IconEnum.WorkflowShapes]: <IconWorkflowShapes />,
-    [IconEnum.WorkflowShapes12]: <IconWorkflowShapes12 />,
-    [IconEnum.WorkflowShapes16]: <IconWorkflowShapes16 />,
-    [IconEnum.WorkflowShapes20]: <IconWorkflowShapes20 />,
-    [IconEnum.WorkflowShapes24]: <IconWorkflowShapes24 />,
-    [IconEnum.WorkflowShapes32]: <IconWorkflowShapes32 />,
-};
