+++ src/foundation/Icon/index.ts
@@ -1,11 +1,6 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 export * from './Generated';
-export * from './Icon';
 export * from './IconEnum';
 export * from './IconProps';
 export * from './IconSize';
-export * from './IconsMap';
-
-import * as AllIcons from './Generated';
-export { AllIcons };
